--- src/auto-imports.d.ts
@@ -1,3970 +1,0 @@
-// Generated by 'unplugin-auto-import'
-export {}
-declare global {
-  const API_URL: typeof import('./shared/constants/url-config')['API_URL']
-  const API_WS_URL: typeof import('./shared/constants/url-config')['API_WS_URL']
-  const AUTH_URL: typeof import('./shared/constants/url-config')['AUTH_URL']
-  const BASE_URL: typeof import('./shared/constants/url-config')['BASE_URL']
-  const BIcon123: typeof import('bootstrap-icons-vue')['BIcon123']
-  const BIconActivity: typeof import('bootstrap-icons-vue')['BIconActivity']
-  const BIconAlarm: typeof import('bootstrap-icons-vue')['BIconAlarm']
-  const BIconAlarmFill: typeof import('bootstrap-icons-vue')['BIconAlarmFill']
-  const BIconAlignBottom: typeof import('bootstrap-icons-vue')['BIconAlignBottom']
-  const BIconAlignCenter: typeof import('bootstrap-icons-vue')['BIconAlignCenter']
-  const BIconAlignEnd: typeof import('bootstrap-icons-vue')['BIconAlignEnd']
-  const BIconAlignMiddle: typeof import('bootstrap-icons-vue')['BIconAlignMiddle']
-  const BIconAlignStart: typeof import('bootstrap-icons-vue')['BIconAlignStart']
-  const BIconAlignTop: typeof import('bootstrap-icons-vue')['BIconAlignTop']
-  const BIconAlt: typeof import('bootstrap-icons-vue')['BIconAlt']
-  const BIconApp: typeof import('bootstrap-icons-vue')['BIconApp']
-  const BIconAppIndicator: typeof import('bootstrap-icons-vue')['BIconAppIndicator']
-  const BIconApple: typeof import('bootstrap-icons-vue')['BIconApple']
-  const BIconArchive: typeof import('bootstrap-icons-vue')['BIconArchive']
-  const BIconArchiveFill: typeof import('bootstrap-icons-vue')['BIconArchiveFill']
-  const BIconArrow90degDown: typeof import('bootstrap-icons-vue')['BIconArrow90degDown']
-  const BIconArrow90degLeft: typeof import('bootstrap-icons-vue')['BIconArrow90degLeft']
-  const BIconArrow90degRight: typeof import('bootstrap-icons-vue')['BIconArrow90degRight']
-  const BIconArrow90degUp: typeof import('bootstrap-icons-vue')['BIconArrow90degUp']
-  const BIconArrowBarDown: typeof import('bootstrap-icons-vue')['BIconArrowBarDown']
-  const BIconArrowBarLeft: typeof import('bootstrap-icons-vue')['BIconArrowBarLeft']
-  const BIconArrowBarRight: typeof import('bootstrap-icons-vue')['BIconArrowBarRight']
-  const BIconArrowBarUp: typeof import('bootstrap-icons-vue')['BIconArrowBarUp']
-  const BIconArrowClockwise: typeof import('bootstrap-icons-vue')['BIconArrowClockwise']
-  const BIconArrowCounterclockwise: typeof import('bootstrap-icons-vue')['BIconArrowCounterclockwise']
-  const BIconArrowDown: typeof import('bootstrap-icons-vue')['BIconArrowDown']
-  const BIconArrowDownCircle: typeof import('bootstrap-icons-vue')['BIconArrowDownCircle']
-  const BIconArrowDownCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowDownCircleFill']
-  const BIconArrowDownLeft: typeof import('bootstrap-icons-vue')['BIconArrowDownLeft']
-  const BIconArrowDownLeftCircle: typeof import('bootstrap-icons-vue')['BIconArrowDownLeftCircle']
-  const BIconArrowDownLeftCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowDownLeftCircleFill']
-  const BIconArrowDownLeftSquare: typeof import('bootstrap-icons-vue')['BIconArrowDownLeftSquare']
-  const BIconArrowDownLeftSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowDownLeftSquareFill']
-  const BIconArrowDownRight: typeof import('bootstrap-icons-vue')['BIconArrowDownRight']
-  const BIconArrowDownRightCircle: typeof import('bootstrap-icons-vue')['BIconArrowDownRightCircle']
-  const BIconArrowDownRightCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowDownRightCircleFill']
-  const BIconArrowDownRightSquare: typeof import('bootstrap-icons-vue')['BIconArrowDownRightSquare']
-  const BIconArrowDownRightSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowDownRightSquareFill']
-  const BIconArrowDownShort: typeof import('bootstrap-icons-vue')['BIconArrowDownShort']
-  const BIconArrowDownSquare: typeof import('bootstrap-icons-vue')['BIconArrowDownSquare']
-  const BIconArrowDownSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowDownSquareFill']
-  const BIconArrowDownUp: typeof import('bootstrap-icons-vue')['BIconArrowDownUp']
-  const BIconArrowLeft: typeof import('bootstrap-icons-vue')['BIconArrowLeft']
-  const BIconArrowLeftCircle: typeof import('bootstrap-icons-vue')['BIconArrowLeftCircle']
-  const BIconArrowLeftCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowLeftCircleFill']
-  const BIconArrowLeftRight: typeof import('bootstrap-icons-vue')['BIconArrowLeftRight']
-  const BIconArrowLeftShort: typeof import('bootstrap-icons-vue')['BIconArrowLeftShort']
-  const BIconArrowLeftSquare: typeof import('bootstrap-icons-vue')['BIconArrowLeftSquare']
-  const BIconArrowLeftSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowLeftSquareFill']
-  const BIconArrowRepeat: typeof import('bootstrap-icons-vue')['BIconArrowRepeat']
-  const BIconArrowReturnLeft: typeof import('bootstrap-icons-vue')['BIconArrowReturnLeft']
-  const BIconArrowReturnRight: typeof import('bootstrap-icons-vue')['BIconArrowReturnRight']
-  const BIconArrowRight: typeof import('bootstrap-icons-vue')['BIconArrowRight']
-  const BIconArrowRightCircle: typeof import('bootstrap-icons-vue')['BIconArrowRightCircle']
-  const BIconArrowRightCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowRightCircleFill']
-  const BIconArrowRightShort: typeof import('bootstrap-icons-vue')['BIconArrowRightShort']
-  const BIconArrowRightSquare: typeof import('bootstrap-icons-vue')['BIconArrowRightSquare']
-  const BIconArrowRightSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowRightSquareFill']
-  const BIconArrowThroughHeart: typeof import('bootstrap-icons-vue')['BIconArrowThroughHeart']
-  const BIconArrowThroughHeartFill: typeof import('bootstrap-icons-vue')['BIconArrowThroughHeartFill']
-  const BIconArrowUp: typeof import('bootstrap-icons-vue')['BIconArrowUp']
-  const BIconArrowUpCircle: typeof import('bootstrap-icons-vue')['BIconArrowUpCircle']
-  const BIconArrowUpCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowUpCircleFill']
-  const BIconArrowUpLeft: typeof import('bootstrap-icons-vue')['BIconArrowUpLeft']
-  const BIconArrowUpLeftCircle: typeof import('bootstrap-icons-vue')['BIconArrowUpLeftCircle']
-  const BIconArrowUpLeftCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowUpLeftCircleFill']
-  const BIconArrowUpLeftSquare: typeof import('bootstrap-icons-vue')['BIconArrowUpLeftSquare']
-  const BIconArrowUpLeftSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowUpLeftSquareFill']
-  const BIconArrowUpRight: typeof import('bootstrap-icons-vue')['BIconArrowUpRight']
-  const BIconArrowUpRightCircle: typeof import('bootstrap-icons-vue')['BIconArrowUpRightCircle']
-  const BIconArrowUpRightCircleFill: typeof import('bootstrap-icons-vue')['BIconArrowUpRightCircleFill']
-  const BIconArrowUpRightSquare: typeof import('bootstrap-icons-vue')['BIconArrowUpRightSquare']
-  const BIconArrowUpRightSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowUpRightSquareFill']
-  const BIconArrowUpShort: typeof import('bootstrap-icons-vue')['BIconArrowUpShort']
-  const BIconArrowUpSquare: typeof import('bootstrap-icons-vue')['BIconArrowUpSquare']
-  const BIconArrowUpSquareFill: typeof import('bootstrap-icons-vue')['BIconArrowUpSquareFill']
-  const BIconArrowsAngleContract: typeof import('bootstrap-icons-vue')['BIconArrowsAngleContract']
-  const BIconArrowsAngleExpand: typeof import('bootstrap-icons-vue')['BIconArrowsAngleExpand']
-  const BIconArrowsCollapse: typeof import('bootstrap-icons-vue')['BIconArrowsCollapse']
-  const BIconArrowsExpand: typeof import('bootstrap-icons-vue')['BIconArrowsExpand']
-  const BIconArrowsFullscreen: typeof import('bootstrap-icons-vue')['BIconArrowsFullscreen']
-  const BIconArrowsMove: typeof import('bootstrap-icons-vue')['BIconArrowsMove']
-  const BIconAspectRatio: typeof import('bootstrap-icons-vue')['BIconAspectRatio']
-  const BIconAspectRatioFill: typeof import('bootstrap-icons-vue')['BIconAspectRatioFill']
-  const BIconAsterisk: typeof import('bootstrap-icons-vue')['BIconAsterisk']
-  const BIconAt: typeof import('bootstrap-icons-vue')['BIconAt']
-  const BIconAward: typeof import('bootstrap-icons-vue')['BIconAward']
-  const BIconAwardFill: typeof import('bootstrap-icons-vue')['BIconAwardFill']
-  const BIconBack: typeof import('bootstrap-icons-vue')['BIconBack']
-  const BIconBackspace: typeof import('bootstrap-icons-vue')['BIconBackspace']
-  const BIconBackspaceFill: typeof import('bootstrap-icons-vue')['BIconBackspaceFill']
-  const BIconBackspaceReverse: typeof import('bootstrap-icons-vue')['BIconBackspaceReverse']
-  const BIconBackspaceReverseFill: typeof import('bootstrap-icons-vue')['BIconBackspaceReverseFill']
-  const BIconBadge3d: typeof import('bootstrap-icons-vue')['BIconBadge3d']
-  const BIconBadge3dFill: typeof import('bootstrap-icons-vue')['BIconBadge3dFill']
-  const BIconBadge4k: typeof import('bootstrap-icons-vue')['BIconBadge4k']
-  const BIconBadge4kFill: typeof import('bootstrap-icons-vue')['BIconBadge4kFill']
-  const BIconBadge8k: typeof import('bootstrap-icons-vue')['BIconBadge8k']
-  const BIconBadge8kFill: typeof import('bootstrap-icons-vue')['BIconBadge8kFill']
-  const BIconBadgeAd: typeof import('bootstrap-icons-vue')['BIconBadgeAd']
-  const BIconBadgeAdFill: typeof import('bootstrap-icons-vue')['BIconBadgeAdFill']
-  const BIconBadgeAr: typeof import('bootstrap-icons-vue')['BIconBadgeAr']
-  const BIconBadgeArFill: typeof import('bootstrap-icons-vue')['BIconBadgeArFill']
-  const BIconBadgeCc: typeof import('bootstrap-icons-vue')['BIconBadgeCc']
-  const BIconBadgeCcFill: typeof import('bootstrap-icons-vue')['BIconBadgeCcFill']
-  const BIconBadgeHd: typeof import('bootstrap-icons-vue')['BIconBadgeHd']
-  const BIconBadgeHdFill: typeof import('bootstrap-icons-vue')['BIconBadgeHdFill']
-  const BIconBadgeSd: typeof import('bootstrap-icons-vue')['BIconBadgeSd']
-  const BIconBadgeSdFill: typeof import('bootstrap-icons-vue')['BIconBadgeSdFill']
-  const BIconBadgeTm: typeof import('bootstrap-icons-vue')['BIconBadgeTm']
-  const BIconBadgeTmFill: typeof import('bootstrap-icons-vue')['BIconBadgeTmFill']
-  const BIconBadgeVo: typeof import('bootstrap-icons-vue')['BIconBadgeVo']
-  const BIconBadgeVoFill: typeof import('bootstrap-icons-vue')['BIconBadgeVoFill']
-  const BIconBadgeVr: typeof import('bootstrap-icons-vue')['BIconBadgeVr']
-  const BIconBadgeVrFill: typeof import('bootstrap-icons-vue')['BIconBadgeVrFill']
-  const BIconBadgeWc: typeof import('bootstrap-icons-vue')['BIconBadgeWc']
-  const BIconBadgeWcFill: typeof import('bootstrap-icons-vue')['BIconBadgeWcFill']
-  const BIconBag: typeof import('bootstrap-icons-vue')['BIconBag']
-  const BIconBagCheck: typeof import('bootstrap-icons-vue')['BIconBagCheck']
-  const BIconBagCheckFill: typeof import('bootstrap-icons-vue')['BIconBagCheckFill']
-  const BIconBagDash: typeof import('bootstrap-icons-vue')['BIconBagDash']
-  const BIconBagDashFill: typeof import('bootstrap-icons-vue')['BIconBagDashFill']
-  const BIconBagFill: typeof import('bootstrap-icons-vue')['BIconBagFill']
-  const BIconBagHeart: typeof import('bootstrap-icons-vue')['BIconBagHeart']
-  const BIconBagHeartFill: typeof import('bootstrap-icons-vue')['BIconBagHeartFill']
-  const BIconBagPlus: typeof import('bootstrap-icons-vue')['BIconBagPlus']
-  const BIconBagPlusFill: typeof import('bootstrap-icons-vue')['BIconBagPlusFill']
-  const BIconBagX: typeof import('bootstrap-icons-vue')['BIconBagX']
-  const BIconBagXFill: typeof import('bootstrap-icons-vue')['BIconBagXFill']
-  const BIconBalloon: typeof import('bootstrap-icons-vue')['BIconBalloon']
-  const BIconBalloonFill: typeof import('bootstrap-icons-vue')['BIconBalloonFill']
-  const BIconBalloonHeart: typeof import('bootstrap-icons-vue')['BIconBalloonHeart']
-  const BIconBalloonHeartFill: typeof import('bootstrap-icons-vue')['BIconBalloonHeartFill']
-  const BIconBandaid: typeof import('bootstrap-icons-vue')['BIconBandaid']
-  const BIconBandaidFill: typeof import('bootstrap-icons-vue')['BIconBandaidFill']
-  const BIconBank2: typeof import('bootstrap-icons-vue')['BIconBank2']
-  const BIconBank: typeof import('bootstrap-icons-vue')['BIconBank']
-  const BIconBarChart: typeof import('bootstrap-icons-vue')['BIconBarChart']
-  const BIconBarChartFill: typeof import('bootstrap-icons-vue')['BIconBarChartFill']
-  const BIconBarChartLine: typeof import('bootstrap-icons-vue')['BIconBarChartLine']
-  const BIconBarChartLineFill: typeof import('bootstrap-icons-vue')['BIconBarChartLineFill']
-  const BIconBarChartSteps: typeof import('bootstrap-icons-vue')['BIconBarChartSteps']
-  const BIconBasket2: typeof import('bootstrap-icons-vue')['BIconBasket2']
-  const BIconBasket2Fill: typeof import('bootstrap-icons-vue')['BIconBasket2Fill']
-  const BIconBasket3: typeof import('bootstrap-icons-vue')['BIconBasket3']
-  const BIconBasket3Fill: typeof import('bootstrap-icons-vue')['BIconBasket3Fill']
-  const BIconBasket: typeof import('bootstrap-icons-vue')['BIconBasket']
-  const BIconBasketFill: typeof import('bootstrap-icons-vue')['BIconBasketFill']
-  const BIconBattery: typeof import('bootstrap-icons-vue')['BIconBattery']
-  const BIconBatteryCharging: typeof import('bootstrap-icons-vue')['BIconBatteryCharging']
-  const BIconBatteryFull: typeof import('bootstrap-icons-vue')['BIconBatteryFull']
-  const BIconBatteryHalf: typeof import('bootstrap-icons-vue')['BIconBatteryHalf']
-  const BIconBehance: typeof import('bootstrap-icons-vue')['BIconBehance']
-  const BIconBell: typeof import('bootstrap-icons-vue')['BIconBell']
-  const BIconBellFill: typeof import('bootstrap-icons-vue')['BIconBellFill']
-  const BIconBellSlash: typeof import('bootstrap-icons-vue')['BIconBellSlash']
-  const BIconBellSlashFill: typeof import('bootstrap-icons-vue')['BIconBellSlashFill']
-  const BIconBezier2: typeof import('bootstrap-icons-vue')['BIconBezier2']
-  const BIconBezier: typeof import('bootstrap-icons-vue')['BIconBezier']
-  const BIconBicycle: typeof import('bootstrap-icons-vue')['BIconBicycle']
-  const BIconBinoculars: typeof import('bootstrap-icons-vue')['BIconBinoculars']
-  const BIconBinocularsFill: typeof import('bootstrap-icons-vue')['BIconBinocularsFill']
-  const BIconBlockquoteLeft: typeof import('bootstrap-icons-vue')['BIconBlockquoteLeft']
-  const BIconBlockquoteRight: typeof import('bootstrap-icons-vue')['BIconBlockquoteRight']
-  const BIconBluetooth: typeof import('bootstrap-icons-vue')['BIconBluetooth']
-  const BIconBodyText: typeof import('bootstrap-icons-vue')['BIconBodyText']
-  const BIconBook: typeof import('bootstrap-icons-vue')['BIconBook']
-  const BIconBookFill: typeof import('bootstrap-icons-vue')['BIconBookFill']
-  const BIconBookHalf: typeof import('bootstrap-icons-vue')['BIconBookHalf']
-  const BIconBookmark: typeof import('bootstrap-icons-vue')['BIconBookmark']
-  const BIconBookmarkCheck: typeof import('bootstrap-icons-vue')['BIconBookmarkCheck']
-  const BIconBookmarkCheckFill: typeof import('bootstrap-icons-vue')['BIconBookmarkCheckFill']
-  const BIconBookmarkDash: typeof import('bootstrap-icons-vue')['BIconBookmarkDash']
-  const BIconBookmarkDashFill: typeof import('bootstrap-icons-vue')['BIconBookmarkDashFill']
-  const BIconBookmarkFill: typeof import('bootstrap-icons-vue')['BIconBookmarkFill']
-  const BIconBookmarkHeart: typeof import('bootstrap-icons-vue')['BIconBookmarkHeart']
-  const BIconBookmarkHeartFill: typeof import('bootstrap-icons-vue')['BIconBookmarkHeartFill']
-  const BIconBookmarkPlus: typeof import('bootstrap-icons-vue')['BIconBookmarkPlus']
-  const BIconBookmarkPlusFill: typeof import('bootstrap-icons-vue')['BIconBookmarkPlusFill']
-  const BIconBookmarkStar: typeof import('bootstrap-icons-vue')['BIconBookmarkStar']
-  const BIconBookmarkStarFill: typeof import('bootstrap-icons-vue')['BIconBookmarkStarFill']
-  const BIconBookmarkX: typeof import('bootstrap-icons-vue')['BIconBookmarkX']
-  const BIconBookmarkXFill: typeof import('bootstrap-icons-vue')['BIconBookmarkXFill']
-  const BIconBookmarks: typeof import('bootstrap-icons-vue')['BIconBookmarks']
-  const BIconBookmarksFill: typeof import('bootstrap-icons-vue')['BIconBookmarksFill']
-  const BIconBookshelf: typeof import('bootstrap-icons-vue')['BIconBookshelf']
-  const BIconBoombox: typeof import('bootstrap-icons-vue')['BIconBoombox']
-  const BIconBoomboxFill: typeof import('bootstrap-icons-vue')['BIconBoomboxFill']
-  const BIconBootstrap: typeof import('bootstrap-icons-vue')['BIconBootstrap']
-  const BIconBootstrapFill: typeof import('bootstrap-icons-vue')['BIconBootstrapFill']
-  const BIconBootstrapReboot: typeof import('bootstrap-icons-vue')['BIconBootstrapReboot']
-  const BIconBorder: typeof import('bootstrap-icons-vue')['BIconBorder']
-  const BIconBorderAll: typeof import('bootstrap-icons-vue')['BIconBorderAll']
-  const BIconBorderBottom: typeof import('bootstrap-icons-vue')['BIconBorderBottom']
-  const BIconBorderCenter: typeof import('bootstrap-icons-vue')['BIconBorderCenter']
-  const BIconBorderInner: typeof import('bootstrap-icons-vue')['BIconBorderInner']
-  const BIconBorderLeft: typeof import('bootstrap-icons-vue')['BIconBorderLeft']
-  const BIconBorderMiddle: typeof import('bootstrap-icons-vue')['BIconBorderMiddle']
-  const BIconBorderOuter: typeof import('bootstrap-icons-vue')['BIconBorderOuter']
-  const BIconBorderRight: typeof import('bootstrap-icons-vue')['BIconBorderRight']
-  const BIconBorderStyle: typeof import('bootstrap-icons-vue')['BIconBorderStyle']
-  const BIconBorderTop: typeof import('bootstrap-icons-vue')['BIconBorderTop']
-  const BIconBorderWidth: typeof import('bootstrap-icons-vue')['BIconBorderWidth']
-  const BIconBoundingBox: typeof import('bootstrap-icons-vue')['BIconBoundingBox']
-  const BIconBoundingBoxCircles: typeof import('bootstrap-icons-vue')['BIconBoundingBoxCircles']
-  const BIconBox2: typeof import('bootstrap-icons-vue')['BIconBox2']
-  const BIconBox2Fill: typeof import('bootstrap-icons-vue')['BIconBox2Fill']
-  const BIconBox2Heart: typeof import('bootstrap-icons-vue')['BIconBox2Heart']
-  const BIconBox2HeartFill: typeof import('bootstrap-icons-vue')['BIconBox2HeartFill']
-  const BIconBox: typeof import('bootstrap-icons-vue')['BIconBox']
-  const BIconBoxArrowDown: typeof import('bootstrap-icons-vue')['BIconBoxArrowDown']
-  const BIconBoxArrowDownLeft: typeof import('bootstrap-icons-vue')['BIconBoxArrowDownLeft']
-  const BIconBoxArrowDownRight: typeof import('bootstrap-icons-vue')['BIconBoxArrowDownRight']
-  const BIconBoxArrowInDown: typeof import('bootstrap-icons-vue')['BIconBoxArrowInDown']
-  const BIconBoxArrowInDownLeft: typeof import('bootstrap-icons-vue')['BIconBoxArrowInDownLeft']
-  const BIconBoxArrowInDownRight: typeof import('bootstrap-icons-vue')['BIconBoxArrowInDownRight']
-  const BIconBoxArrowInLeft: typeof import('bootstrap-icons-vue')['BIconBoxArrowInLeft']
-  const BIconBoxArrowInRight: typeof import('bootstrap-icons-vue')['BIconBoxArrowInRight']
-  const BIconBoxArrowInUp: typeof import('bootstrap-icons-vue')['BIconBoxArrowInUp']
-  const BIconBoxArrowInUpLeft: typeof import('bootstrap-icons-vue')['BIconBoxArrowInUpLeft']
-  const BIconBoxArrowInUpRight: typeof import('bootstrap-icons-vue')['BIconBoxArrowInUpRight']
-  const BIconBoxArrowLeft: typeof import('bootstrap-icons-vue')['BIconBoxArrowLeft']
-  const BIconBoxArrowRight: typeof import('bootstrap-icons-vue')['BIconBoxArrowRight']
-  const BIconBoxArrowUp: typeof import('bootstrap-icons-vue')['BIconBoxArrowUp']
-  const BIconBoxArrowUpLeft: typeof import('bootstrap-icons-vue')['BIconBoxArrowUpLeft']
-  const BIconBoxArrowUpRight: typeof import('bootstrap-icons-vue')['BIconBoxArrowUpRight']
-  const BIconBoxSeam: typeof import('bootstrap-icons-vue')['BIconBoxSeam']
-  const BIconBoxes: typeof import('bootstrap-icons-vue')['BIconBoxes']
-  const BIconBraces: typeof import('bootstrap-icons-vue')['BIconBraces']
-  const BIconBracesAsterisk: typeof import('bootstrap-icons-vue')['BIconBracesAsterisk']
-  const BIconBricks: typeof import('bootstrap-icons-vue')['BIconBricks']
-  const BIconBriefcase: typeof import('bootstrap-icons-vue')['BIconBriefcase']
-  const BIconBriefcaseFill: typeof import('bootstrap-icons-vue')['BIconBriefcaseFill']
-  const BIconBrightnessAltHigh: typeof import('bootstrap-icons-vue')['BIconBrightnessAltHigh']
-  const BIconBrightnessAltHighFill: typeof import('bootstrap-icons-vue')['BIconBrightnessAltHighFill']
-  const BIconBrightnessAltLow: typeof import('bootstrap-icons-vue')['BIconBrightnessAltLow']
-  const BIconBrightnessAltLowFill: typeof import('bootstrap-icons-vue')['BIconBrightnessAltLowFill']
-  const BIconBrightnessHigh: typeof import('bootstrap-icons-vue')['BIconBrightnessHigh']
-  const BIconBrightnessHighFill: typeof import('bootstrap-icons-vue')['BIconBrightnessHighFill']
-  const BIconBrightnessLow: typeof import('bootstrap-icons-vue')['BIconBrightnessLow']
-  const BIconBrightnessLowFill: typeof import('bootstrap-icons-vue')['BIconBrightnessLowFill']
-  const BIconBroadcast: typeof import('bootstrap-icons-vue')['BIconBroadcast']
-  const BIconBroadcastPin: typeof import('bootstrap-icons-vue')['BIconBroadcastPin']
-  const BIconBrush: typeof import('bootstrap-icons-vue')['BIconBrush']
-  const BIconBrushFill: typeof import('bootstrap-icons-vue')['BIconBrushFill']
-  const BIconBucket: typeof import('bootstrap-icons-vue')['BIconBucket']
-  const BIconBucketFill: typeof import('bootstrap-icons-vue')['BIconBucketFill']
-  const BIconBug: typeof import('bootstrap-icons-vue')['BIconBug']
-  const BIconBugFill: typeof import('bootstrap-icons-vue')['BIconBugFill']
-  const BIconBuilding: typeof import('bootstrap-icons-vue')['BIconBuilding']
-  const BIconBullseye: typeof import('bootstrap-icons-vue')['BIconBullseye']
-  const BIconCalculator: typeof import('bootstrap-icons-vue')['BIconCalculator']
-  const BIconCalculatorFill: typeof import('bootstrap-icons-vue')['BIconCalculatorFill']
-  const BIconCalendar2: typeof import('bootstrap-icons-vue')['BIconCalendar2']
-  const BIconCalendar2Check: typeof import('bootstrap-icons-vue')['BIconCalendar2Check']
-  const BIconCalendar2CheckFill: typeof import('bootstrap-icons-vue')['BIconCalendar2CheckFill']
-  const BIconCalendar2Date: typeof import('bootstrap-icons-vue')['BIconCalendar2Date']
-  const BIconCalendar2DateFill: typeof import('bootstrap-icons-vue')['BIconCalendar2DateFill']
-  const BIconCalendar2Day: typeof import('bootstrap-icons-vue')['BIconCalendar2Day']
-  const BIconCalendar2DayFill: typeof import('bootstrap-icons-vue')['BIconCalendar2DayFill']
-  const BIconCalendar2Event: typeof import('bootstrap-icons-vue')['BIconCalendar2Event']
-  const BIconCalendar2EventFill: typeof import('bootstrap-icons-vue')['BIconCalendar2EventFill']
-  const BIconCalendar2Fill: typeof import('bootstrap-icons-vue')['BIconCalendar2Fill']
-  const BIconCalendar2Heart: typeof import('bootstrap-icons-vue')['BIconCalendar2Heart']
-  const BIconCalendar2HeartFill: typeof import('bootstrap-icons-vue')['BIconCalendar2HeartFill']
-  const BIconCalendar2Minus: typeof import('bootstrap-icons-vue')['BIconCalendar2Minus']
-  const BIconCalendar2MinusFill: typeof import('bootstrap-icons-vue')['BIconCalendar2MinusFill']
-  const BIconCalendar2Month: typeof import('bootstrap-icons-vue')['BIconCalendar2Month']
-  const BIconCalendar2MonthFill: typeof import('bootstrap-icons-vue')['BIconCalendar2MonthFill']
-  const BIconCalendar2Plus: typeof import('bootstrap-icons-vue')['BIconCalendar2Plus']
-  const BIconCalendar2PlusFill: typeof import('bootstrap-icons-vue')['BIconCalendar2PlusFill']
-  const BIconCalendar2Range: typeof import('bootstrap-icons-vue')['BIconCalendar2Range']
-  const BIconCalendar2RangeFill: typeof import('bootstrap-icons-vue')['BIconCalendar2RangeFill']
-  const BIconCalendar2Week: typeof import('bootstrap-icons-vue')['BIconCalendar2Week']
-  const BIconCalendar2WeekFill: typeof import('bootstrap-icons-vue')['BIconCalendar2WeekFill']
-  const BIconCalendar2X: typeof import('bootstrap-icons-vue')['BIconCalendar2X']
-  const BIconCalendar2XFill: typeof import('bootstrap-icons-vue')['BIconCalendar2XFill']
-  const BIconCalendar3: typeof import('bootstrap-icons-vue')['BIconCalendar3']
-  const BIconCalendar3Event: typeof import('bootstrap-icons-vue')['BIconCalendar3Event']
-  const BIconCalendar3EventFill: typeof import('bootstrap-icons-vue')['BIconCalendar3EventFill']
-  const BIconCalendar3Fill: typeof import('bootstrap-icons-vue')['BIconCalendar3Fill']
-  const BIconCalendar3Range: typeof import('bootstrap-icons-vue')['BIconCalendar3Range']
-  const BIconCalendar3RangeFill: typeof import('bootstrap-icons-vue')['BIconCalendar3RangeFill']
-  const BIconCalendar3Week: typeof import('bootstrap-icons-vue')['BIconCalendar3Week']
-  const BIconCalendar3WeekFill: typeof import('bootstrap-icons-vue')['BIconCalendar3WeekFill']
-  const BIconCalendar4: typeof import('bootstrap-icons-vue')['BIconCalendar4']
-  const BIconCalendar4Event: typeof import('bootstrap-icons-vue')['BIconCalendar4Event']
-  const BIconCalendar4Range: typeof import('bootstrap-icons-vue')['BIconCalendar4Range']
-  const BIconCalendar4Week: typeof import('bootstrap-icons-vue')['BIconCalendar4Week']
-  const BIconCalendar: typeof import('bootstrap-icons-vue')['BIconCalendar']
-  const BIconCalendarCheck: typeof import('bootstrap-icons-vue')['BIconCalendarCheck']
-  const BIconCalendarCheckFill: typeof import('bootstrap-icons-vue')['BIconCalendarCheckFill']
-  const BIconCalendarDate: typeof import('bootstrap-icons-vue')['BIconCalendarDate']
-  const BIconCalendarDateFill: typeof import('bootstrap-icons-vue')['BIconCalendarDateFill']
-  const BIconCalendarDay: typeof import('bootstrap-icons-vue')['BIconCalendarDay']
-  const BIconCalendarDayFill: typeof import('bootstrap-icons-vue')['BIconCalendarDayFill']
-  const BIconCalendarEvent: typeof import('bootstrap-icons-vue')['BIconCalendarEvent']
-  const BIconCalendarEventFill: typeof import('bootstrap-icons-vue')['BIconCalendarEventFill']
-  const BIconCalendarFill: typeof import('bootstrap-icons-vue')['BIconCalendarFill']
-  const BIconCalendarHeart: typeof import('bootstrap-icons-vue')['BIconCalendarHeart']
-  const BIconCalendarHeartFill: typeof import('bootstrap-icons-vue')['BIconCalendarHeartFill']
-  const BIconCalendarMinus: typeof import('bootstrap-icons-vue')['BIconCalendarMinus']
-  const BIconCalendarMinusFill: typeof import('bootstrap-icons-vue')['BIconCalendarMinusFill']
-  const BIconCalendarMonth: typeof import('bootstrap-icons-vue')['BIconCalendarMonth']
-  const BIconCalendarMonthFill: typeof import('bootstrap-icons-vue')['BIconCalendarMonthFill']
-  const BIconCalendarPlus: typeof import('bootstrap-icons-vue')['BIconCalendarPlus']
-  const BIconCalendarPlusFill: typeof import('bootstrap-icons-vue')['BIconCalendarPlusFill']
-  const BIconCalendarRange: typeof import('bootstrap-icons-vue')['BIconCalendarRange']
-  const BIconCalendarRangeFill: typeof import('bootstrap-icons-vue')['BIconCalendarRangeFill']
-  const BIconCalendarWeek: typeof import('bootstrap-icons-vue')['BIconCalendarWeek']
-  const BIconCalendarWeekFill: typeof import('bootstrap-icons-vue')['BIconCalendarWeekFill']
-  const BIconCalendarX: typeof import('bootstrap-icons-vue')['BIconCalendarX']
-  const BIconCalendarXFill: typeof import('bootstrap-icons-vue')['BIconCalendarXFill']
-  const BIconCamera2: typeof import('bootstrap-icons-vue')['BIconCamera2']
-  const BIconCamera: typeof import('bootstrap-icons-vue')['BIconCamera']
-  const BIconCameraFill: typeof import('bootstrap-icons-vue')['BIconCameraFill']
-  const BIconCameraReels: typeof import('bootstrap-icons-vue')['BIconCameraReels']
-  const BIconCameraReelsFill: typeof import('bootstrap-icons-vue')['BIconCameraReelsFill']
-  const BIconCameraVideo: typeof import('bootstrap-icons-vue')['BIconCameraVideo']
-  const BIconCameraVideoFill: typeof import('bootstrap-icons-vue')['BIconCameraVideoFill']
-  const BIconCameraVideoOff: typeof import('bootstrap-icons-vue')['BIconCameraVideoOff']
-  const BIconCameraVideoOffFill: typeof import('bootstrap-icons-vue')['BIconCameraVideoOffFill']
-  const BIconCapslock: typeof import('bootstrap-icons-vue')['BIconCapslock']
-  const BIconCapslockFill: typeof import('bootstrap-icons-vue')['BIconCapslockFill']
-  const BIconCardChecklist: typeof import('bootstrap-icons-vue')['BIconCardChecklist']
-  const BIconCardHeading: typeof import('bootstrap-icons-vue')['BIconCardHeading']
-  const BIconCardImage: typeof import('bootstrap-icons-vue')['BIconCardImage']
-  const BIconCardList: typeof import('bootstrap-icons-vue')['BIconCardList']
-  const BIconCardText: typeof import('bootstrap-icons-vue')['BIconCardText']
-  const BIconCaretDown: typeof import('bootstrap-icons-vue')['BIconCaretDown']
-  const BIconCaretDownFill: typeof import('bootstrap-icons-vue')['BIconCaretDownFill']
-  const BIconCaretDownSquare: typeof import('bootstrap-icons-vue')['BIconCaretDownSquare']
-  const BIconCaretDownSquareFill: typeof import('bootstrap-icons-vue')['BIconCaretDownSquareFill']
-  const BIconCaretLeft: typeof import('bootstrap-icons-vue')['BIconCaretLeft']
-  const BIconCaretLeftFill: typeof import('bootstrap-icons-vue')['BIconCaretLeftFill']
-  const BIconCaretLeftSquare: typeof import('bootstrap-icons-vue')['BIconCaretLeftSquare']
-  const BIconCaretLeftSquareFill: typeof import('bootstrap-icons-vue')['BIconCaretLeftSquareFill']
-  const BIconCaretRight: typeof import('bootstrap-icons-vue')['BIconCaretRight']
-  const BIconCaretRightFill: typeof import('bootstrap-icons-vue')['BIconCaretRightFill']
-  const BIconCaretRightSquare: typeof import('bootstrap-icons-vue')['BIconCaretRightSquare']
-  const BIconCaretRightSquareFill: typeof import('bootstrap-icons-vue')['BIconCaretRightSquareFill']
-  const BIconCaretUp: typeof import('bootstrap-icons-vue')['BIconCaretUp']
-  const BIconCaretUpFill: typeof import('bootstrap-icons-vue')['BIconCaretUpFill']
-  const BIconCaretUpSquare: typeof import('bootstrap-icons-vue')['BIconCaretUpSquare']
-  const BIconCaretUpSquareFill: typeof import('bootstrap-icons-vue')['BIconCaretUpSquareFill']
-  const BIconCart2: typeof import('bootstrap-icons-vue')['BIconCart2']
-  const BIconCart3: typeof import('bootstrap-icons-vue')['BIconCart3']
-  const BIconCart4: typeof import('bootstrap-icons-vue')['BIconCart4']
-  const BIconCart: typeof import('bootstrap-icons-vue')['BIconCart']
-  const BIconCartCheck: typeof import('bootstrap-icons-vue')['BIconCartCheck']
-  const BIconCartCheckFill: typeof import('bootstrap-icons-vue')['BIconCartCheckFill']
-  const BIconCartDash: typeof import('bootstrap-icons-vue')['BIconCartDash']
-  const BIconCartDashFill: typeof import('bootstrap-icons-vue')['BIconCartDashFill']
-  const BIconCartFill: typeof import('bootstrap-icons-vue')['BIconCartFill']
-  const BIconCartPlus: typeof import('bootstrap-icons-vue')['BIconCartPlus']
-  const BIconCartPlusFill: typeof import('bootstrap-icons-vue')['BIconCartPlusFill']
-  const BIconCartX: typeof import('bootstrap-icons-vue')['BIconCartX']
-  const BIconCartXFill: typeof import('bootstrap-icons-vue')['BIconCartXFill']
-  const BIconCash: typeof import('bootstrap-icons-vue')['BIconCash']
-  const BIconCashCoin: typeof import('bootstrap-icons-vue')['BIconCashCoin']
-  const BIconCashStack: typeof import('bootstrap-icons-vue')['BIconCashStack']
-  const BIconCast: typeof import('bootstrap-icons-vue')['BIconCast']
-  const BIconChat: typeof import('bootstrap-icons-vue')['BIconChat']
-  const BIconChatDots: typeof import('bootstrap-icons-vue')['BIconChatDots']
-  const BIconChatDotsFill: typeof import('bootstrap-icons-vue')['BIconChatDotsFill']
-  const BIconChatFill: typeof import('bootstrap-icons-vue')['BIconChatFill']
-  const BIconChatHeart: typeof import('bootstrap-icons-vue')['BIconChatHeart']
-  const BIconChatHeartFill: typeof import('bootstrap-icons-vue')['BIconChatHeartFill']
-  const BIconChatLeft: typeof import('bootstrap-icons-vue')['BIconChatLeft']
-  const BIconChatLeftDots: typeof import('bootstrap-icons-vue')['BIconChatLeftDots']
-  const BIconChatLeftDotsFill: typeof import('bootstrap-icons-vue')['BIconChatLeftDotsFill']
-  const BIconChatLeftFill: typeof import('bootstrap-icons-vue')['BIconChatLeftFill']
-  const BIconChatLeftHeart: typeof import('bootstrap-icons-vue')['BIconChatLeftHeart']
-  const BIconChatLeftHeartFill: typeof import('bootstrap-icons-vue')['BIconChatLeftHeartFill']
-  const BIconChatLeftQuote: typeof import('bootstrap-icons-vue')['BIconChatLeftQuote']
-  const BIconChatLeftQuoteFill: typeof import('bootstrap-icons-vue')['BIconChatLeftQuoteFill']
-  const BIconChatLeftText: typeof import('bootstrap-icons-vue')['BIconChatLeftText']
-  const BIconChatLeftTextFill: typeof import('bootstrap-icons-vue')['BIconChatLeftTextFill']
-  const BIconChatQuote: typeof import('bootstrap-icons-vue')['BIconChatQuote']
-  const BIconChatQuoteFill: typeof import('bootstrap-icons-vue')['BIconChatQuoteFill']
-  const BIconChatRight: typeof import('bootstrap-icons-vue')['BIconChatRight']
-  const BIconChatRightDots: typeof import('bootstrap-icons-vue')['BIconChatRightDots']
-  const BIconChatRightDotsFill: typeof import('bootstrap-icons-vue')['BIconChatRightDotsFill']
-  const BIconChatRightFill: typeof import('bootstrap-icons-vue')['BIconChatRightFill']
-  const BIconChatRightHeart: typeof import('bootstrap-icons-vue')['BIconChatRightHeart']
-  const BIconChatRightHeartFill: typeof import('bootstrap-icons-vue')['BIconChatRightHeartFill']
-  const BIconChatRightQuote: typeof import('bootstrap-icons-vue')['BIconChatRightQuote']
-  const BIconChatRightQuoteFill: typeof import('bootstrap-icons-vue')['BIconChatRightQuoteFill']
-  const BIconChatRightText: typeof import('bootstrap-icons-vue')['BIconChatRightText']
-  const BIconChatRightTextFill: typeof import('bootstrap-icons-vue')['BIconChatRightTextFill']
-  const BIconChatSquare: typeof import('bootstrap-icons-vue')['BIconChatSquare']
-  const BIconChatSquareDots: typeof import('bootstrap-icons-vue')['BIconChatSquareDots']
-  const BIconChatSquareDotsFill: typeof import('bootstrap-icons-vue')['BIconChatSquareDotsFill']
-  const BIconChatSquareFill: typeof import('bootstrap-icons-vue')['BIconChatSquareFill']
-  const BIconChatSquareHeart: typeof import('bootstrap-icons-vue')['BIconChatSquareHeart']
-  const BIconChatSquareHeartFill: typeof import('bootstrap-icons-vue')['BIconChatSquareHeartFill']
-  const BIconChatSquareQuote: typeof import('bootstrap-icons-vue')['BIconChatSquareQuote']
-  const BIconChatSquareQuoteFill: typeof import('bootstrap-icons-vue')['BIconChatSquareQuoteFill']
-  const BIconChatSquareText: typeof import('bootstrap-icons-vue')['BIconChatSquareText']
-  const BIconChatSquareTextFill: typeof import('bootstrap-icons-vue')['BIconChatSquareTextFill']
-  const BIconChatText: typeof import('bootstrap-icons-vue')['BIconChatText']
-  const BIconChatTextFill: typeof import('bootstrap-icons-vue')['BIconChatTextFill']
-  const BIconCheck2: typeof import('bootstrap-icons-vue')['BIconCheck2']
-  const BIconCheck2All: typeof import('bootstrap-icons-vue')['BIconCheck2All']
-  const BIconCheck2Circle: typeof import('bootstrap-icons-vue')['BIconCheck2Circle']
-  const BIconCheck2Square: typeof import('bootstrap-icons-vue')['BIconCheck2Square']
-  const BIconCheck: typeof import('bootstrap-icons-vue')['BIconCheck']
-  const BIconCheckAll: typeof import('bootstrap-icons-vue')['BIconCheckAll']
-  const BIconCheckCircle: typeof import('bootstrap-icons-vue')['BIconCheckCircle']
-  const BIconCheckCircleFill: typeof import('bootstrap-icons-vue')['BIconCheckCircleFill']
-  const BIconCheckLg: typeof import('bootstrap-icons-vue')['BIconCheckLg']
-  const BIconCheckSquare: typeof import('bootstrap-icons-vue')['BIconCheckSquare']
-  const BIconCheckSquareFill: typeof import('bootstrap-icons-vue')['BIconCheckSquareFill']
-  const BIconChevronBarContract: typeof import('bootstrap-icons-vue')['BIconChevronBarContract']
-  const BIconChevronBarDown: typeof import('bootstrap-icons-vue')['BIconChevronBarDown']
-  const BIconChevronBarExpand: typeof import('bootstrap-icons-vue')['BIconChevronBarExpand']
-  const BIconChevronBarLeft: typeof import('bootstrap-icons-vue')['BIconChevronBarLeft']
-  const BIconChevronBarRight: typeof import('bootstrap-icons-vue')['BIconChevronBarRight']
-  const BIconChevronBarUp: typeof import('bootstrap-icons-vue')['BIconChevronBarUp']
-  const BIconChevronCompactDown: typeof import('bootstrap-icons-vue')['BIconChevronCompactDown']
-  const BIconChevronCompactLeft: typeof import('bootstrap-icons-vue')['BIconChevronCompactLeft']
-  const BIconChevronCompactRight: typeof import('bootstrap-icons-vue')['BIconChevronCompactRight']
-  const BIconChevronCompactUp: typeof import('bootstrap-icons-vue')['BIconChevronCompactUp']
-  const BIconChevronContract: typeof import('bootstrap-icons-vue')['BIconChevronContract']
-  const BIconChevronDoubleDown: typeof import('bootstrap-icons-vue')['BIconChevronDoubleDown']
-  const BIconChevronDoubleLeft: typeof import('bootstrap-icons-vue')['BIconChevronDoubleLeft']
-  const BIconChevronDoubleRight: typeof import('bootstrap-icons-vue')['BIconChevronDoubleRight']
-  const BIconChevronDoubleUp: typeof import('bootstrap-icons-vue')['BIconChevronDoubleUp']
-  const BIconChevronDown: typeof import('bootstrap-icons-vue')['BIconChevronDown']
-  const BIconChevronExpand: typeof import('bootstrap-icons-vue')['BIconChevronExpand']
-  const BIconChevronLeft: typeof import('bootstrap-icons-vue')['BIconChevronLeft']
-  const BIconChevronRight: typeof import('bootstrap-icons-vue')['BIconChevronRight']
-  const BIconChevronUp: typeof import('bootstrap-icons-vue')['BIconChevronUp']
-  const BIconCircle: typeof import('bootstrap-icons-vue')['BIconCircle']
-  const BIconCircleFill: typeof import('bootstrap-icons-vue')['BIconCircleFill']
-  const BIconCircleHalf: typeof import('bootstrap-icons-vue')['BIconCircleHalf']
-  const BIconCircleSquare: typeof import('bootstrap-icons-vue')['BIconCircleSquare']
-  const BIconClipboard2: typeof import('bootstrap-icons-vue')['BIconClipboard2']
-  const BIconClipboard2Check: typeof import('bootstrap-icons-vue')['BIconClipboard2Check']
-  const BIconClipboard2CheckFill: typeof import('bootstrap-icons-vue')['BIconClipboard2CheckFill']
-  const BIconClipboard2Data: typeof import('bootstrap-icons-vue')['BIconClipboard2Data']
-  const BIconClipboard2DataFill: typeof import('bootstrap-icons-vue')['BIconClipboard2DataFill']
-  const BIconClipboard2Fill: typeof import('bootstrap-icons-vue')['BIconClipboard2Fill']
-  const BIconClipboard2Heart: typeof import('bootstrap-icons-vue')['BIconClipboard2Heart']
-  const BIconClipboard2HeartFill: typeof import('bootstrap-icons-vue')['BIconClipboard2HeartFill']
-  const BIconClipboard2Minus: typeof import('bootstrap-icons-vue')['BIconClipboard2Minus']
-  const BIconClipboard2MinusFill: typeof import('bootstrap-icons-vue')['BIconClipboard2MinusFill']
-  const BIconClipboard2Plus: typeof import('bootstrap-icons-vue')['BIconClipboard2Plus']
-  const BIconClipboard2PlusFill: typeof import('bootstrap-icons-vue')['BIconClipboard2PlusFill']
-  const BIconClipboard2Pulse: typeof import('bootstrap-icons-vue')['BIconClipboard2Pulse']
-  const BIconClipboard2PulseFill: typeof import('bootstrap-icons-vue')['BIconClipboard2PulseFill']
-  const BIconClipboard2X: typeof import('bootstrap-icons-vue')['BIconClipboard2X']
-  const BIconClipboard2XFill: typeof import('bootstrap-icons-vue')['BIconClipboard2XFill']
-  const BIconClipboard: typeof import('bootstrap-icons-vue')['BIconClipboard']
-  const BIconClipboardCheck: typeof import('bootstrap-icons-vue')['BIconClipboardCheck']
-  const BIconClipboardCheckFill: typeof import('bootstrap-icons-vue')['BIconClipboardCheckFill']
-  const BIconClipboardData: typeof import('bootstrap-icons-vue')['BIconClipboardData']
-  const BIconClipboardDataFill: typeof import('bootstrap-icons-vue')['BIconClipboardDataFill']
-  const BIconClipboardFill: typeof import('bootstrap-icons-vue')['BIconClipboardFill']
-  const BIconClipboardHeart: typeof import('bootstrap-icons-vue')['BIconClipboardHeart']
-  const BIconClipboardHeartFill: typeof import('bootstrap-icons-vue')['BIconClipboardHeartFill']
-  const BIconClipboardMinus: typeof import('bootstrap-icons-vue')['BIconClipboardMinus']
-  const BIconClipboardMinusFill: typeof import('bootstrap-icons-vue')['BIconClipboardMinusFill']
-  const BIconClipboardPlus: typeof import('bootstrap-icons-vue')['BIconClipboardPlus']
-  const BIconClipboardPlusFill: typeof import('bootstrap-icons-vue')['BIconClipboardPlusFill']
-  const BIconClipboardPulse: typeof import('bootstrap-icons-vue')['BIconClipboardPulse']
-  const BIconClipboardX: typeof import('bootstrap-icons-vue')['BIconClipboardX']
-  const BIconClipboardXFill: typeof import('bootstrap-icons-vue')['BIconClipboardXFill']
-  const BIconClock: typeof import('bootstrap-icons-vue')['BIconClock']
-  const BIconClockFill: typeof import('bootstrap-icons-vue')['BIconClockFill']
-  const BIconClockHistory: typeof import('bootstrap-icons-vue')['BIconClockHistory']
-  const BIconCloud: typeof import('bootstrap-icons-vue')['BIconCloud']
-  const BIconCloudArrowDown: typeof import('bootstrap-icons-vue')['BIconCloudArrowDown']
-  const BIconCloudArrowDownFill: typeof import('bootstrap-icons-vue')['BIconCloudArrowDownFill']
-  const BIconCloudArrowUp: typeof import('bootstrap-icons-vue')['BIconCloudArrowUp']
-  const BIconCloudArrowUpFill: typeof import('bootstrap-icons-vue')['BIconCloudArrowUpFill']
-  const BIconCloudCheck: typeof import('bootstrap-icons-vue')['BIconCloudCheck']
-  const BIconCloudCheckFill: typeof import('bootstrap-icons-vue')['BIconCloudCheckFill']
-  const BIconCloudDownload: typeof import('bootstrap-icons-vue')['BIconCloudDownload']
-  const BIconCloudDownloadFill: typeof import('bootstrap-icons-vue')['BIconCloudDownloadFill']
-  const BIconCloudDrizzle: typeof import('bootstrap-icons-vue')['BIconCloudDrizzle']
-  const BIconCloudDrizzleFill: typeof import('bootstrap-icons-vue')['BIconCloudDrizzleFill']
-  const BIconCloudFill: typeof import('bootstrap-icons-vue')['BIconCloudFill']
-  const BIconCloudFog2: typeof import('bootstrap-icons-vue')['BIconCloudFog2']
-  const BIconCloudFog2Fill: typeof import('bootstrap-icons-vue')['BIconCloudFog2Fill']
-  const BIconCloudFog: typeof import('bootstrap-icons-vue')['BIconCloudFog']
-  const BIconCloudFogFill: typeof import('bootstrap-icons-vue')['BIconCloudFogFill']
-  const BIconCloudHail: typeof import('bootstrap-icons-vue')['BIconCloudHail']
-  const BIconCloudHailFill: typeof import('bootstrap-icons-vue')['BIconCloudHailFill']
-  const BIconCloudHaze2: typeof import('bootstrap-icons-vue')['BIconCloudHaze2']
-  const BIconCloudHaze2Fill: typeof import('bootstrap-icons-vue')['BIconCloudHaze2Fill']
-  const BIconCloudHaze: typeof import('bootstrap-icons-vue')['BIconCloudHaze']
-  const BIconCloudHazeFill: typeof import('bootstrap-icons-vue')['BIconCloudHazeFill']
-  const BIconCloudLightning: typeof import('bootstrap-icons-vue')['BIconCloudLightning']
-  const BIconCloudLightningFill: typeof import('bootstrap-icons-vue')['BIconCloudLightningFill']
-  const BIconCloudLightningRain: typeof import('bootstrap-icons-vue')['BIconCloudLightningRain']
-  const BIconCloudLightningRainFill: typeof import('bootstrap-icons-vue')['BIconCloudLightningRainFill']
-  const BIconCloudMinus: typeof import('bootstrap-icons-vue')['BIconCloudMinus']
-  const BIconCloudMinusFill: typeof import('bootstrap-icons-vue')['BIconCloudMinusFill']
-  const BIconCloudMoon: typeof import('bootstrap-icons-vue')['BIconCloudMoon']
-  const BIconCloudMoonFill: typeof import('bootstrap-icons-vue')['BIconCloudMoonFill']
-  const BIconCloudPlus: typeof import('bootstrap-icons-vue')['BIconCloudPlus']
-  const BIconCloudPlusFill: typeof import('bootstrap-icons-vue')['BIconCloudPlusFill']
-  const BIconCloudRain: typeof import('bootstrap-icons-vue')['BIconCloudRain']
-  const BIconCloudRainFill: typeof import('bootstrap-icons-vue')['BIconCloudRainFill']
-  const BIconCloudRainHeavy: typeof import('bootstrap-icons-vue')['BIconCloudRainHeavy']
-  const BIconCloudRainHeavyFill: typeof import('bootstrap-icons-vue')['BIconCloudRainHeavyFill']
-  const BIconCloudSlash: typeof import('bootstrap-icons-vue')['BIconCloudSlash']
-  const BIconCloudSlashFill: typeof import('bootstrap-icons-vue')['BIconCloudSlashFill']
-  const BIconCloudSleet: typeof import('bootstrap-icons-vue')['BIconCloudSleet']
-  const BIconCloudSleetFill: typeof import('bootstrap-icons-vue')['BIconCloudSleetFill']
-  const BIconCloudSnow: typeof import('bootstrap-icons-vue')['BIconCloudSnow']
-  const BIconCloudSnowFill: typeof import('bootstrap-icons-vue')['BIconCloudSnowFill']
-  const BIconCloudSun: typeof import('bootstrap-icons-vue')['BIconCloudSun']
-  const BIconCloudSunFill: typeof import('bootstrap-icons-vue')['BIconCloudSunFill']
-  const BIconCloudUpload: typeof import('bootstrap-icons-vue')['BIconCloudUpload']
-  const BIconCloudUploadFill: typeof import('bootstrap-icons-vue')['BIconCloudUploadFill']
-  const BIconClouds: typeof import('bootstrap-icons-vue')['BIconClouds']
-  const BIconCloudsFill: typeof import('bootstrap-icons-vue')['BIconCloudsFill']
-  const BIconCloudy: typeof import('bootstrap-icons-vue')['BIconCloudy']
-  const BIconCloudyFill: typeof import('bootstrap-icons-vue')['BIconCloudyFill']
-  const BIconCode: typeof import('bootstrap-icons-vue')['BIconCode']
-  const BIconCodeSlash: typeof import('bootstrap-icons-vue')['BIconCodeSlash']
-  const BIconCodeSquare: typeof import('bootstrap-icons-vue')['BIconCodeSquare']
-  const BIconCoin: typeof import('bootstrap-icons-vue')['BIconCoin']
-  const BIconCollection: typeof import('bootstrap-icons-vue')['BIconCollection']
-  const BIconCollectionFill: typeof import('bootstrap-icons-vue')['BIconCollectionFill']
-  const BIconCollectionPlay: typeof import('bootstrap-icons-vue')['BIconCollectionPlay']
-  const BIconCollectionPlayFill: typeof import('bootstrap-icons-vue')['BIconCollectionPlayFill']
-  const BIconColumns: typeof import('bootstrap-icons-vue')['BIconColumns']
-  const BIconColumnsGap: typeof import('bootstrap-icons-vue')['BIconColumnsGap']
-  const BIconCommand: typeof import('bootstrap-icons-vue')['BIconCommand']
-  const BIconCompass: typeof import('bootstrap-icons-vue')['BIconCompass']
-  const BIconCompassFill: typeof import('bootstrap-icons-vue')['BIconCompassFill']
-  const BIconCone: typeof import('bootstrap-icons-vue')['BIconCone']
-  const BIconConeStriped: typeof import('bootstrap-icons-vue')['BIconConeStriped']
-  const BIconController: typeof import('bootstrap-icons-vue')['BIconController']
-  const BIconCpu: typeof import('bootstrap-icons-vue')['BIconCpu']
-  const BIconCpuFill: typeof import('bootstrap-icons-vue')['BIconCpuFill']
-  const BIconCreditCard2Back: typeof import('bootstrap-icons-vue')['BIconCreditCard2Back']
-  const BIconCreditCard2BackFill: typeof import('bootstrap-icons-vue')['BIconCreditCard2BackFill']
-  const BIconCreditCard2Front: typeof import('bootstrap-icons-vue')['BIconCreditCard2Front']
-  const BIconCreditCard2FrontFill: typeof import('bootstrap-icons-vue')['BIconCreditCard2FrontFill']
-  const BIconCreditCard: typeof import('bootstrap-icons-vue')['BIconCreditCard']
-  const BIconCreditCardFill: typeof import('bootstrap-icons-vue')['BIconCreditCardFill']
-  const BIconCrop: typeof import('bootstrap-icons-vue')['BIconCrop']
-  const BIconCup: typeof import('bootstrap-icons-vue')['BIconCup']
-  const BIconCupFill: typeof import('bootstrap-icons-vue')['BIconCupFill']
-  const BIconCupStraw: typeof import('bootstrap-icons-vue')['BIconCupStraw']
-  const BIconCurrencyBitcoin: typeof import('bootstrap-icons-vue')['BIconCurrencyBitcoin']
-  const BIconCurrencyDollar: typeof import('bootstrap-icons-vue')['BIconCurrencyDollar']
-  const BIconCurrencyEuro: typeof import('bootstrap-icons-vue')['BIconCurrencyEuro']
-  const BIconCurrencyExchange: typeof import('bootstrap-icons-vue')['BIconCurrencyExchange']
-  const BIconCurrencyPound: typeof import('bootstrap-icons-vue')['BIconCurrencyPound']
-  const BIconCurrencyYen: typeof import('bootstrap-icons-vue')['BIconCurrencyYen']
-  const BIconCursor: typeof import('bootstrap-icons-vue')['BIconCursor']
-  const BIconCursorFill: typeof import('bootstrap-icons-vue')['BIconCursorFill']
-  const BIconCursorText: typeof import('bootstrap-icons-vue')['BIconCursorText']
-  const BIconDash: typeof import('bootstrap-icons-vue')['BIconDash']
-  const BIconDashCircle: typeof import('bootstrap-icons-vue')['BIconDashCircle']
-  const BIconDashCircleDotted: typeof import('bootstrap-icons-vue')['BIconDashCircleDotted']
-  const BIconDashCircleFill: typeof import('bootstrap-icons-vue')['BIconDashCircleFill']
-  const BIconDashLg: typeof import('bootstrap-icons-vue')['BIconDashLg']
-  const BIconDashSquare: typeof import('bootstrap-icons-vue')['BIconDashSquare']
-  const BIconDashSquareDotted: typeof import('bootstrap-icons-vue')['BIconDashSquareDotted']
-  const BIconDashSquareFill: typeof import('bootstrap-icons-vue')['BIconDashSquareFill']
-  const BIconDeviceHdd: typeof import('bootstrap-icons-vue')['BIconDeviceHdd']
-  const BIconDeviceHddFill: typeof import('bootstrap-icons-vue')['BIconDeviceHddFill']
-  const BIconDeviceSsd: typeof import('bootstrap-icons-vue')['BIconDeviceSsd']
-  const BIconDeviceSsdFill: typeof import('bootstrap-icons-vue')['BIconDeviceSsdFill']
-  const BIconDiagram2: typeof import('bootstrap-icons-vue')['BIconDiagram2']
-  const BIconDiagram2Fill: typeof import('bootstrap-icons-vue')['BIconDiagram2Fill']
-  const BIconDiagram3: typeof import('bootstrap-icons-vue')['BIconDiagram3']
-  const BIconDiagram3Fill: typeof import('bootstrap-icons-vue')['BIconDiagram3Fill']
-  const BIconDiamond: typeof import('bootstrap-icons-vue')['BIconDiamond']
-  const BIconDiamondFill: typeof import('bootstrap-icons-vue')['BIconDiamondFill']
-  const BIconDiamondHalf: typeof import('bootstrap-icons-vue')['BIconDiamondHalf']
-  const BIconDice1: typeof import('bootstrap-icons-vue')['BIconDice1']
-  const BIconDice1Fill: typeof import('bootstrap-icons-vue')['BIconDice1Fill']
-  const BIconDice2: typeof import('bootstrap-icons-vue')['BIconDice2']
-  const BIconDice2Fill: typeof import('bootstrap-icons-vue')['BIconDice2Fill']
-  const BIconDice3: typeof import('bootstrap-icons-vue')['BIconDice3']
-  const BIconDice3Fill: typeof import('bootstrap-icons-vue')['BIconDice3Fill']
-  const BIconDice4: typeof import('bootstrap-icons-vue')['BIconDice4']
-  const BIconDice4Fill: typeof import('bootstrap-icons-vue')['BIconDice4Fill']
-  const BIconDice5: typeof import('bootstrap-icons-vue')['BIconDice5']
-  const BIconDice5Fill: typeof import('bootstrap-icons-vue')['BIconDice5Fill']
-  const BIconDice6: typeof import('bootstrap-icons-vue')['BIconDice6']
-  const BIconDice6Fill: typeof import('bootstrap-icons-vue')['BIconDice6Fill']
-  const BIconDisc: typeof import('bootstrap-icons-vue')['BIconDisc']
-  const BIconDiscFill: typeof import('bootstrap-icons-vue')['BIconDiscFill']
-  const BIconDiscord: typeof import('bootstrap-icons-vue')['BIconDiscord']
-  const BIconDisplay: typeof import('bootstrap-icons-vue')['BIconDisplay']
-  const BIconDisplayFill: typeof import('bootstrap-icons-vue')['BIconDisplayFill']
-  const BIconDisplayport: typeof import('bootstrap-icons-vue')['BIconDisplayport']
-  const BIconDisplayportFill: typeof import('bootstrap-icons-vue')['BIconDisplayportFill']
-  const BIconDistributeHorizontal: typeof import('bootstrap-icons-vue')['BIconDistributeHorizontal']
-  const BIconDistributeVertical: typeof import('bootstrap-icons-vue')['BIconDistributeVertical']
-  const BIconDoorClosed: typeof import('bootstrap-icons-vue')['BIconDoorClosed']
-  const BIconDoorClosedFill: typeof import('bootstrap-icons-vue')['BIconDoorClosedFill']
-  const BIconDoorOpen: typeof import('bootstrap-icons-vue')['BIconDoorOpen']
-  const BIconDoorOpenFill: typeof import('bootstrap-icons-vue')['BIconDoorOpenFill']
-  const BIconDot: typeof import('bootstrap-icons-vue')['BIconDot']
-  const BIconDownload: typeof import('bootstrap-icons-vue')['BIconDownload']
-  const BIconDpad: typeof import('bootstrap-icons-vue')['BIconDpad']
-  const BIconDpadFill: typeof import('bootstrap-icons-vue')['BIconDpadFill']
-  const BIconDribbble: typeof import('bootstrap-icons-vue')['BIconDribbble']
-  const BIconDroplet: typeof import('bootstrap-icons-vue')['BIconDroplet']
-  const BIconDropletFill: typeof import('bootstrap-icons-vue')['BIconDropletFill']
-  const BIconDropletHalf: typeof import('bootstrap-icons-vue')['BIconDropletHalf']
-  const BIconEar: typeof import('bootstrap-icons-vue')['BIconEar']
-  const BIconEarFill: typeof import('bootstrap-icons-vue')['BIconEarFill']
-  const BIconEarbuds: typeof import('bootstrap-icons-vue')['BIconEarbuds']
-  const BIconEasel2: typeof import('bootstrap-icons-vue')['BIconEasel2']
-  const BIconEasel2Fill: typeof import('bootstrap-icons-vue')['BIconEasel2Fill']
-  const BIconEasel3: typeof import('bootstrap-icons-vue')['BIconEasel3']
-  const BIconEasel3Fill: typeof import('bootstrap-icons-vue')['BIconEasel3Fill']
-  const BIconEasel: typeof import('bootstrap-icons-vue')['BIconEasel']
-  const BIconEaselFill: typeof import('bootstrap-icons-vue')['BIconEaselFill']
-  const BIconEgg: typeof import('bootstrap-icons-vue')['BIconEgg']
-  const BIconEggFill: typeof import('bootstrap-icons-vue')['BIconEggFill']
-  const BIconEggFried: typeof import('bootstrap-icons-vue')['BIconEggFried']
-  const BIconEject: typeof import('bootstrap-icons-vue')['BIconEject']
-  const BIconEjectFill: typeof import('bootstrap-icons-vue')['BIconEjectFill']
-  const BIconEmojiAngry: typeof import('bootstrap-icons-vue')['BIconEmojiAngry']
-  const BIconEmojiAngryFill: typeof import('bootstrap-icons-vue')['BIconEmojiAngryFill']
-  const BIconEmojiDizzy: typeof import('bootstrap-icons-vue')['BIconEmojiDizzy']
-  const BIconEmojiDizzyFill: typeof import('bootstrap-icons-vue')['BIconEmojiDizzyFill']
-  const BIconEmojiExpressionless: typeof import('bootstrap-icons-vue')['BIconEmojiExpressionless']
-  const BIconEmojiExpressionlessFill: typeof import('bootstrap-icons-vue')['BIconEmojiExpressionlessFill']
-  const BIconEmojiFrown: typeof import('bootstrap-icons-vue')['BIconEmojiFrown']
-  const BIconEmojiFrownFill: typeof import('bootstrap-icons-vue')['BIconEmojiFrownFill']
-  const BIconEmojiHeartEyes: typeof import('bootstrap-icons-vue')['BIconEmojiHeartEyes']
-  const BIconEmojiHeartEyesFill: typeof import('bootstrap-icons-vue')['BIconEmojiHeartEyesFill']
-  const BIconEmojiKiss: typeof import('bootstrap-icons-vue')['BIconEmojiKiss']
-  const BIconEmojiKissFill: typeof import('bootstrap-icons-vue')['BIconEmojiKissFill']
-  const BIconEmojiLaughing: typeof import('bootstrap-icons-vue')['BIconEmojiLaughing']
-  const BIconEmojiLaughingFill: typeof import('bootstrap-icons-vue')['BIconEmojiLaughingFill']
-  const BIconEmojiNeutral: typeof import('bootstrap-icons-vue')['BIconEmojiNeutral']
-  const BIconEmojiNeutralFill: typeof import('bootstrap-icons-vue')['BIconEmojiNeutralFill']
-  const BIconEmojiSmile: typeof import('bootstrap-icons-vue')['BIconEmojiSmile']
-  const BIconEmojiSmileFill: typeof import('bootstrap-icons-vue')['BIconEmojiSmileFill']
-  const BIconEmojiSmileUpsideDown: typeof import('bootstrap-icons-vue')['BIconEmojiSmileUpsideDown']
-  const BIconEmojiSmileUpsideDownFill: typeof import('bootstrap-icons-vue')['BIconEmojiSmileUpsideDownFill']
-  const BIconEmojiSunglasses: typeof import('bootstrap-icons-vue')['BIconEmojiSunglasses']
-  const BIconEmojiSunglassesFill: typeof import('bootstrap-icons-vue')['BIconEmojiSunglassesFill']
-  const BIconEmojiWink: typeof import('bootstrap-icons-vue')['BIconEmojiWink']
-  const BIconEmojiWinkFill: typeof import('bootstrap-icons-vue')['BIconEmojiWinkFill']
-  const BIconEnvelope: typeof import('bootstrap-icons-vue')['BIconEnvelope']
-  const BIconEnvelopeCheck: typeof import('bootstrap-icons-vue')['BIconEnvelopeCheck']
-  const BIconEnvelopeCheckFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeCheckFill']
-  const BIconEnvelopeDash: typeof import('bootstrap-icons-vue')['BIconEnvelopeDash']
-  const BIconEnvelopeDashFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeDashFill']
-  const BIconEnvelopeExclamation: typeof import('bootstrap-icons-vue')['BIconEnvelopeExclamation']
-  const BIconEnvelopeExclamationFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeExclamationFill']
-  const BIconEnvelopeFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeFill']
-  const BIconEnvelopeHeart: typeof import('bootstrap-icons-vue')['BIconEnvelopeHeart']
-  const BIconEnvelopeHeartFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeHeartFill']
-  const BIconEnvelopeOpen: typeof import('bootstrap-icons-vue')['BIconEnvelopeOpen']
-  const BIconEnvelopeOpenFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeOpenFill']
-  const BIconEnvelopeOpenHeart: typeof import('bootstrap-icons-vue')['BIconEnvelopeOpenHeart']
-  const BIconEnvelopeOpenHeartFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeOpenHeartFill']
-  const BIconEnvelopePaper: typeof import('bootstrap-icons-vue')['BIconEnvelopePaper']
-  const BIconEnvelopePaperFill: typeof import('bootstrap-icons-vue')['BIconEnvelopePaperFill']
-  const BIconEnvelopePaperHeart: typeof import('bootstrap-icons-vue')['BIconEnvelopePaperHeart']
-  const BIconEnvelopePaperHeartFill: typeof import('bootstrap-icons-vue')['BIconEnvelopePaperHeartFill']
-  const BIconEnvelopePlus: typeof import('bootstrap-icons-vue')['BIconEnvelopePlus']
-  const BIconEnvelopePlusFill: typeof import('bootstrap-icons-vue')['BIconEnvelopePlusFill']
-  const BIconEnvelopeSlash: typeof import('bootstrap-icons-vue')['BIconEnvelopeSlash']
-  const BIconEnvelopeSlashFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeSlashFill']
-  const BIconEnvelopeX: typeof import('bootstrap-icons-vue')['BIconEnvelopeX']
-  const BIconEnvelopeXFill: typeof import('bootstrap-icons-vue')['BIconEnvelopeXFill']
-  const BIconEraser: typeof import('bootstrap-icons-vue')['BIconEraser']
-  const BIconEraserFill: typeof import('bootstrap-icons-vue')['BIconEraserFill']
-  const BIconEthernet: typeof import('bootstrap-icons-vue')['BIconEthernet']
-  const BIconExclamation: typeof import('bootstrap-icons-vue')['BIconExclamation']
-  const BIconExclamationCircle: typeof import('bootstrap-icons-vue')['BIconExclamationCircle']
-  const BIconExclamationCircleFill: typeof import('bootstrap-icons-vue')['BIconExclamationCircleFill']
-  const BIconExclamationDiamond: typeof import('bootstrap-icons-vue')['BIconExclamationDiamond']
-  const BIconExclamationDiamondFill: typeof import('bootstrap-icons-vue')['BIconExclamationDiamondFill']
-  const BIconExclamationLg: typeof import('bootstrap-icons-vue')['BIconExclamationLg']
-  const BIconExclamationOctagon: typeof import('bootstrap-icons-vue')['BIconExclamationOctagon']
-  const BIconExclamationOctagonFill: typeof import('bootstrap-icons-vue')['BIconExclamationOctagonFill']
-  const BIconExclamationSquare: typeof import('bootstrap-icons-vue')['BIconExclamationSquare']
-  const BIconExclamationSquareFill: typeof import('bootstrap-icons-vue')['BIconExclamationSquareFill']
-  const BIconExclamationTriangle: typeof import('bootstrap-icons-vue')['BIconExclamationTriangle']
-  const BIconExclamationTriangleFill: typeof import('bootstrap-icons-vue')['BIconExclamationTriangleFill']
-  const BIconExclude: typeof import('bootstrap-icons-vue')['BIconExclude']
-  const BIconExplicit: typeof import('bootstrap-icons-vue')['BIconExplicit']
-  const BIconExplicitFill: typeof import('bootstrap-icons-vue')['BIconExplicitFill']
-  const BIconEye: typeof import('bootstrap-icons-vue')['BIconEye']
-  const BIconEyeFill: typeof import('bootstrap-icons-vue')['BIconEyeFill']
-  const BIconEyeSlash: typeof import('bootstrap-icons-vue')['BIconEyeSlash']
-  const BIconEyeSlashFill: typeof import('bootstrap-icons-vue')['BIconEyeSlashFill']
-  const BIconEyedropper: typeof import('bootstrap-icons-vue')['BIconEyedropper']
-  const BIconEyeglasses: typeof import('bootstrap-icons-vue')['BIconEyeglasses']
-  const BIconFacebook: typeof import('bootstrap-icons-vue')['BIconFacebook']
-  const BIconFan: typeof import('bootstrap-icons-vue')['BIconFan']
-  const BIconFile: typeof import('bootstrap-icons-vue')['BIconFile']
-  const BIconFileArrowDown: typeof import('bootstrap-icons-vue')['BIconFileArrowDown']
-  const BIconFileArrowDownFill: typeof import('bootstrap-icons-vue')['BIconFileArrowDownFill']
-  const BIconFileArrowUp: typeof import('bootstrap-icons-vue')['BIconFileArrowUp']
-  const BIconFileArrowUpFill: typeof import('bootstrap-icons-vue')['BIconFileArrowUpFill']
-  const BIconFileBarGraph: typeof import('bootstrap-icons-vue')['BIconFileBarGraph']
-  const BIconFileBarGraphFill: typeof import('bootstrap-icons-vue')['BIconFileBarGraphFill']
-  const BIconFileBinary: typeof import('bootstrap-icons-vue')['BIconFileBinary']
-  const BIconFileBinaryFill: typeof import('bootstrap-icons-vue')['BIconFileBinaryFill']
-  const BIconFileBreak: typeof import('bootstrap-icons-vue')['BIconFileBreak']
-  const BIconFileBreakFill: typeof import('bootstrap-icons-vue')['BIconFileBreakFill']
-  const BIconFileCheck: typeof import('bootstrap-icons-vue')['BIconFileCheck']
-  const BIconFileCheckFill: typeof import('bootstrap-icons-vue')['BIconFileCheckFill']
-  const BIconFileCode: typeof import('bootstrap-icons-vue')['BIconFileCode']
-  const BIconFileCodeFill: typeof import('bootstrap-icons-vue')['BIconFileCodeFill']
-  const BIconFileDiff: typeof import('bootstrap-icons-vue')['BIconFileDiff']
-  const BIconFileDiffFill: typeof import('bootstrap-icons-vue')['BIconFileDiffFill']
-  const BIconFileEarmark: typeof import('bootstrap-icons-vue')['BIconFileEarmark']
-  const BIconFileEarmarkArrowDown: typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowDown']
-  const BIconFileEarmarkArrowDownFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowDownFill']
-  const BIconFileEarmarkArrowUp: typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowUp']
-  const BIconFileEarmarkArrowUpFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowUpFill']
-  const BIconFileEarmarkBarGraph: typeof import('bootstrap-icons-vue')['BIconFileEarmarkBarGraph']
-  const BIconFileEarmarkBarGraphFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkBarGraphFill']
-  const BIconFileEarmarkBinary: typeof import('bootstrap-icons-vue')['BIconFileEarmarkBinary']
-  const BIconFileEarmarkBinaryFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkBinaryFill']
-  const BIconFileEarmarkBreak: typeof import('bootstrap-icons-vue')['BIconFileEarmarkBreak']
-  const BIconFileEarmarkBreakFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkBreakFill']
-  const BIconFileEarmarkCheck: typeof import('bootstrap-icons-vue')['BIconFileEarmarkCheck']
-  const BIconFileEarmarkCheckFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkCheckFill']
-  const BIconFileEarmarkCode: typeof import('bootstrap-icons-vue')['BIconFileEarmarkCode']
-  const BIconFileEarmarkCodeFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkCodeFill']
-  const BIconFileEarmarkDiff: typeof import('bootstrap-icons-vue')['BIconFileEarmarkDiff']
-  const BIconFileEarmarkDiffFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkDiffFill']
-  const BIconFileEarmarkEasel: typeof import('bootstrap-icons-vue')['BIconFileEarmarkEasel']
-  const BIconFileEarmarkEaselFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkEaselFill']
-  const BIconFileEarmarkExcel: typeof import('bootstrap-icons-vue')['BIconFileEarmarkExcel']
-  const BIconFileEarmarkExcelFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkExcelFill']
-  const BIconFileEarmarkFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkFill']
-  const BIconFileEarmarkFont: typeof import('bootstrap-icons-vue')['BIconFileEarmarkFont']
-  const BIconFileEarmarkFontFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkFontFill']
-  const BIconFileEarmarkImage: typeof import('bootstrap-icons-vue')['BIconFileEarmarkImage']
-  const BIconFileEarmarkImageFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkImageFill']
-  const BIconFileEarmarkLock2: typeof import('bootstrap-icons-vue')['BIconFileEarmarkLock2']
-  const BIconFileEarmarkLock2Fill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkLock2Fill']
-  const BIconFileEarmarkLock: typeof import('bootstrap-icons-vue')['BIconFileEarmarkLock']
-  const BIconFileEarmarkLockFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkLockFill']
-  const BIconFileEarmarkMedical: typeof import('bootstrap-icons-vue')['BIconFileEarmarkMedical']
-  const BIconFileEarmarkMedicalFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkMedicalFill']
-  const BIconFileEarmarkMinus: typeof import('bootstrap-icons-vue')['BIconFileEarmarkMinus']
-  const BIconFileEarmarkMinusFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkMinusFill']
-  const BIconFileEarmarkMusic: typeof import('bootstrap-icons-vue')['BIconFileEarmarkMusic']
-  const BIconFileEarmarkMusicFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkMusicFill']
-  const BIconFileEarmarkPdf: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPdf']
-  const BIconFileEarmarkPdfFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPdfFill']
-  const BIconFileEarmarkPerson: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPerson']
-  const BIconFileEarmarkPersonFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPersonFill']
-  const BIconFileEarmarkPlay: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlay']
-  const BIconFileEarmarkPlayFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlayFill']
-  const BIconFileEarmarkPlus: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlus']
-  const BIconFileEarmarkPlusFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlusFill']
-  const BIconFileEarmarkPost: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPost']
-  const BIconFileEarmarkPostFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPostFill']
-  const BIconFileEarmarkPpt: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPpt']
-  const BIconFileEarmarkPptFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkPptFill']
-  const BIconFileEarmarkRichtext: typeof import('bootstrap-icons-vue')['BIconFileEarmarkRichtext']
-  const BIconFileEarmarkRichtextFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkRichtextFill']
-  const BIconFileEarmarkRuled: typeof import('bootstrap-icons-vue')['BIconFileEarmarkRuled']
-  const BIconFileEarmarkRuledFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkRuledFill']
-  const BIconFileEarmarkSlides: typeof import('bootstrap-icons-vue')['BIconFileEarmarkSlides']
-  const BIconFileEarmarkSlidesFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkSlidesFill']
-  const BIconFileEarmarkSpreadsheet: typeof import('bootstrap-icons-vue')['BIconFileEarmarkSpreadsheet']
-  const BIconFileEarmarkSpreadsheetFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkSpreadsheetFill']
-  const BIconFileEarmarkText: typeof import('bootstrap-icons-vue')['BIconFileEarmarkText']
-  const BIconFileEarmarkTextFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkTextFill']
-  const BIconFileEarmarkWord: typeof import('bootstrap-icons-vue')['BIconFileEarmarkWord']
-  const BIconFileEarmarkWordFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkWordFill']
-  const BIconFileEarmarkX: typeof import('bootstrap-icons-vue')['BIconFileEarmarkX']
-  const BIconFileEarmarkXFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkXFill']
-  const BIconFileEarmarkZip: typeof import('bootstrap-icons-vue')['BIconFileEarmarkZip']
-  const BIconFileEarmarkZipFill: typeof import('bootstrap-icons-vue')['BIconFileEarmarkZipFill']
-  const BIconFileEasel: typeof import('bootstrap-icons-vue')['BIconFileEasel']
-  const BIconFileEaselFill: typeof import('bootstrap-icons-vue')['BIconFileEaselFill']
-  const BIconFileExcel: typeof import('bootstrap-icons-vue')['BIconFileExcel']
-  const BIconFileExcelFill: typeof import('bootstrap-icons-vue')['BIconFileExcelFill']
-  const BIconFileFill: typeof import('bootstrap-icons-vue')['BIconFileFill']
-  const BIconFileFont: typeof import('bootstrap-icons-vue')['BIconFileFont']
-  const BIconFileFontFill: typeof import('bootstrap-icons-vue')['BIconFileFontFill']
-  const BIconFileImage: typeof import('bootstrap-icons-vue')['BIconFileImage']
-  const BIconFileImageFill: typeof import('bootstrap-icons-vue')['BIconFileImageFill']
-  const BIconFileLock2: typeof import('bootstrap-icons-vue')['BIconFileLock2']
-  const BIconFileLock2Fill: typeof import('bootstrap-icons-vue')['BIconFileLock2Fill']
-  const BIconFileLock: typeof import('bootstrap-icons-vue')['BIconFileLock']
-  const BIconFileLockFill: typeof import('bootstrap-icons-vue')['BIconFileLockFill']
-  const BIconFileMedical: typeof import('bootstrap-icons-vue')['BIconFileMedical']
-  const BIconFileMedicalFill: typeof import('bootstrap-icons-vue')['BIconFileMedicalFill']
-  const BIconFileMinus: typeof import('bootstrap-icons-vue')['BIconFileMinus']
-  const BIconFileMinusFill: typeof import('bootstrap-icons-vue')['BIconFileMinusFill']
-  const BIconFileMusic: typeof import('bootstrap-icons-vue')['BIconFileMusic']
-  const BIconFileMusicFill: typeof import('bootstrap-icons-vue')['BIconFileMusicFill']
-  const BIconFilePdf: typeof import('bootstrap-icons-vue')['BIconFilePdf']
-  const BIconFilePdfFill: typeof import('bootstrap-icons-vue')['BIconFilePdfFill']
-  const BIconFilePerson: typeof import('bootstrap-icons-vue')['BIconFilePerson']
-  const BIconFilePersonFill: typeof import('bootstrap-icons-vue')['BIconFilePersonFill']
-  const BIconFilePlay: typeof import('bootstrap-icons-vue')['BIconFilePlay']
-  const BIconFilePlayFill: typeof import('bootstrap-icons-vue')['BIconFilePlayFill']
-  const BIconFilePlus: typeof import('bootstrap-icons-vue')['BIconFilePlus']
-  const BIconFilePlusFill: typeof import('bootstrap-icons-vue')['BIconFilePlusFill']
-  const BIconFilePost: typeof import('bootstrap-icons-vue')['BIconFilePost']
-  const BIconFilePostFill: typeof import('bootstrap-icons-vue')['BIconFilePostFill']
-  const BIconFilePpt: typeof import('bootstrap-icons-vue')['BIconFilePpt']
-  const BIconFilePptFill: typeof import('bootstrap-icons-vue')['BIconFilePptFill']
-  const BIconFileRichtext: typeof import('bootstrap-icons-vue')['BIconFileRichtext']
-  const BIconFileRichtextFill: typeof import('bootstrap-icons-vue')['BIconFileRichtextFill']
-  const BIconFileRuled: typeof import('bootstrap-icons-vue')['BIconFileRuled']
-  const BIconFileRuledFill: typeof import('bootstrap-icons-vue')['BIconFileRuledFill']
-  const BIconFileSlides: typeof import('bootstrap-icons-vue')['BIconFileSlides']
-  const BIconFileSlidesFill: typeof import('bootstrap-icons-vue')['BIconFileSlidesFill']
-  const BIconFileSpreadsheet: typeof import('bootstrap-icons-vue')['BIconFileSpreadsheet']
-  const BIconFileSpreadsheetFill: typeof import('bootstrap-icons-vue')['BIconFileSpreadsheetFill']
-  const BIconFileText: typeof import('bootstrap-icons-vue')['BIconFileText']
-  const BIconFileTextFill: typeof import('bootstrap-icons-vue')['BIconFileTextFill']
-  const BIconFileWord: typeof import('bootstrap-icons-vue')['BIconFileWord']
-  const BIconFileWordFill: typeof import('bootstrap-icons-vue')['BIconFileWordFill']
-  const BIconFileX: typeof import('bootstrap-icons-vue')['BIconFileX']
-  const BIconFileXFill: typeof import('bootstrap-icons-vue')['BIconFileXFill']
-  const BIconFileZip: typeof import('bootstrap-icons-vue')['BIconFileZip']
-  const BIconFileZipFill: typeof import('bootstrap-icons-vue')['BIconFileZipFill']
-  const BIconFiles: typeof import('bootstrap-icons-vue')['BIconFiles']
-  const BIconFilesAlt: typeof import('bootstrap-icons-vue')['BIconFilesAlt']
-  const BIconFiletypeAac: typeof import('bootstrap-icons-vue')['BIconFiletypeAac']
-  const BIconFiletypeAi: typeof import('bootstrap-icons-vue')['BIconFiletypeAi']
-  const BIconFiletypeBmp: typeof import('bootstrap-icons-vue')['BIconFiletypeBmp']
-  const BIconFiletypeCs: typeof import('bootstrap-icons-vue')['BIconFiletypeCs']
-  const BIconFiletypeCss: typeof import('bootstrap-icons-vue')['BIconFiletypeCss']
-  const BIconFiletypeCsv: typeof import('bootstrap-icons-vue')['BIconFiletypeCsv']
-  const BIconFiletypeDoc: typeof import('bootstrap-icons-vue')['BIconFiletypeDoc']
-  const BIconFiletypeDocx: typeof import('bootstrap-icons-vue')['BIconFiletypeDocx']
-  const BIconFiletypeExe: typeof import('bootstrap-icons-vue')['BIconFiletypeExe']
-  const BIconFiletypeGif: typeof import('bootstrap-icons-vue')['BIconFiletypeGif']
-  const BIconFiletypeHeic: typeof import('bootstrap-icons-vue')['BIconFiletypeHeic']
-  const BIconFiletypeHtml: typeof import('bootstrap-icons-vue')['BIconFiletypeHtml']
-  const BIconFiletypeJava: typeof import('bootstrap-icons-vue')['BIconFiletypeJava']
-  const BIconFiletypeJpg: typeof import('bootstrap-icons-vue')['BIconFiletypeJpg']
-  const BIconFiletypeJs: typeof import('bootstrap-icons-vue')['BIconFiletypeJs']
-  const BIconFiletypeJson: typeof import('bootstrap-icons-vue')['BIconFiletypeJson']
-  const BIconFiletypeJsx: typeof import('bootstrap-icons-vue')['BIconFiletypeJsx']
-  const BIconFiletypeKey: typeof import('bootstrap-icons-vue')['BIconFiletypeKey']
-  const BIconFiletypeM4p: typeof import('bootstrap-icons-vue')['BIconFiletypeM4p']
-  const BIconFiletypeMd: typeof import('bootstrap-icons-vue')['BIconFiletypeMd']
-  const BIconFiletypeMdx: typeof import('bootstrap-icons-vue')['BIconFiletypeMdx']
-  const BIconFiletypeMov: typeof import('bootstrap-icons-vue')['BIconFiletypeMov']
-  const BIconFiletypeMp3: typeof import('bootstrap-icons-vue')['BIconFiletypeMp3']
-  const BIconFiletypeMp4: typeof import('bootstrap-icons-vue')['BIconFiletypeMp4']
-  const BIconFiletypeOtf: typeof import('bootstrap-icons-vue')['BIconFiletypeOtf']
-  const BIconFiletypePdf: typeof import('bootstrap-icons-vue')['BIconFiletypePdf']
-  const BIconFiletypePhp: typeof import('bootstrap-icons-vue')['BIconFiletypePhp']
-  const BIconFiletypePng: typeof import('bootstrap-icons-vue')['BIconFiletypePng']
-  const BIconFiletypePpt: typeof import('bootstrap-icons-vue')['BIconFiletypePpt']
-  const BIconFiletypePptx: typeof import('bootstrap-icons-vue')['BIconFiletypePptx']
-  const BIconFiletypePsd: typeof import('bootstrap-icons-vue')['BIconFiletypePsd']
-  const BIconFiletypePy: typeof import('bootstrap-icons-vue')['BIconFiletypePy']
-  const BIconFiletypeRaw: typeof import('bootstrap-icons-vue')['BIconFiletypeRaw']
-  const BIconFiletypeRb: typeof import('bootstrap-icons-vue')['BIconFiletypeRb']
-  const BIconFiletypeSass: typeof import('bootstrap-icons-vue')['BIconFiletypeSass']
-  const BIconFiletypeScss: typeof import('bootstrap-icons-vue')['BIconFiletypeScss']
-  const BIconFiletypeSh: typeof import('bootstrap-icons-vue')['BIconFiletypeSh']
-  const BIconFiletypeSvg: typeof import('bootstrap-icons-vue')['BIconFiletypeSvg']
-  const BIconFiletypeTiff: typeof import('bootstrap-icons-vue')['BIconFiletypeTiff']
-  const BIconFiletypeTsx: typeof import('bootstrap-icons-vue')['BIconFiletypeTsx']
-  const BIconFiletypeTtf: typeof import('bootstrap-icons-vue')['BIconFiletypeTtf']
-  const BIconFiletypeTxt: typeof import('bootstrap-icons-vue')['BIconFiletypeTxt']
-  const BIconFiletypeWav: typeof import('bootstrap-icons-vue')['BIconFiletypeWav']
-  const BIconFiletypeWoff: typeof import('bootstrap-icons-vue')['BIconFiletypeWoff']
-  const BIconFiletypeXls: typeof import('bootstrap-icons-vue')['BIconFiletypeXls']
-  const BIconFiletypeXlsx: typeof import('bootstrap-icons-vue')['BIconFiletypeXlsx']
-  const BIconFiletypeXml: typeof import('bootstrap-icons-vue')['BIconFiletypeXml']
-  const BIconFiletypeYml: typeof import('bootstrap-icons-vue')['BIconFiletypeYml']
-  const BIconFilm: typeof import('bootstrap-icons-vue')['BIconFilm']
-  const BIconFilter: typeof import('bootstrap-icons-vue')['BIconFilter']
-  const BIconFilterCircle: typeof import('bootstrap-icons-vue')['BIconFilterCircle']
-  const BIconFilterCircleFill: typeof import('bootstrap-icons-vue')['BIconFilterCircleFill']
-  const BIconFilterLeft: typeof import('bootstrap-icons-vue')['BIconFilterLeft']
-  const BIconFilterRight: typeof import('bootstrap-icons-vue')['BIconFilterRight']
-  const BIconFilterSquare: typeof import('bootstrap-icons-vue')['BIconFilterSquare']
-  const BIconFilterSquareFill: typeof import('bootstrap-icons-vue')['BIconFilterSquareFill']
-  const BIconFingerprint: typeof import('bootstrap-icons-vue')['BIconFingerprint']
-  const BIconFlag: typeof import('bootstrap-icons-vue')['BIconFlag']
-  const BIconFlagFill: typeof import('bootstrap-icons-vue')['BIconFlagFill']
-  const BIconFlower1: typeof import('bootstrap-icons-vue')['BIconFlower1']
-  const BIconFlower2: typeof import('bootstrap-icons-vue')['BIconFlower2']
-  const BIconFlower3: typeof import('bootstrap-icons-vue')['BIconFlower3']
-  const BIconFolder2: typeof import('bootstrap-icons-vue')['BIconFolder2']
-  const BIconFolder2Open: typeof import('bootstrap-icons-vue')['BIconFolder2Open']
-  const BIconFolder: typeof import('bootstrap-icons-vue')['BIconFolder']
-  const BIconFolderCheck: typeof import('bootstrap-icons-vue')['BIconFolderCheck']
-  const BIconFolderFill: typeof import('bootstrap-icons-vue')['BIconFolderFill']
-  const BIconFolderMinus: typeof import('bootstrap-icons-vue')['BIconFolderMinus']
-  const BIconFolderPlus: typeof import('bootstrap-icons-vue')['BIconFolderPlus']
-  const BIconFolderSymlink: typeof import('bootstrap-icons-vue')['BIconFolderSymlink']
-  const BIconFolderSymlinkFill: typeof import('bootstrap-icons-vue')['BIconFolderSymlinkFill']
-  const BIconFolderX: typeof import('bootstrap-icons-vue')['BIconFolderX']
-  const BIconFonts: typeof import('bootstrap-icons-vue')['BIconFonts']
-  const BIconForward: typeof import('bootstrap-icons-vue')['BIconForward']
-  const BIconForwardFill: typeof import('bootstrap-icons-vue')['BIconForwardFill']
-  const BIconFront: typeof import('bootstrap-icons-vue')['BIconFront']
-  const BIconFullscreen: typeof import('bootstrap-icons-vue')['BIconFullscreen']
-  const BIconFullscreenExit: typeof import('bootstrap-icons-vue')['BIconFullscreenExit']
-  const BIconFunnel: typeof import('bootstrap-icons-vue')['BIconFunnel']
-  const BIconFunnelFill: typeof import('bootstrap-icons-vue')['BIconFunnelFill']
-  const BIconGear: typeof import('bootstrap-icons-vue')['BIconGear']
-  const BIconGearFill: typeof import('bootstrap-icons-vue')['BIconGearFill']
-  const BIconGearWide: typeof import('bootstrap-icons-vue')['BIconGearWide']
-  const BIconGearWideConnected: typeof import('bootstrap-icons-vue')['BIconGearWideConnected']
-  const BIconGem: typeof import('bootstrap-icons-vue')['BIconGem']
-  const BIconGenderAmbiguous: typeof import('bootstrap-icons-vue')['BIconGenderAmbiguous']
-  const BIconGenderFemale: typeof import('bootstrap-icons-vue')['BIconGenderFemale']
-  const BIconGenderMale: typeof import('bootstrap-icons-vue')['BIconGenderMale']
-  const BIconGenderTrans: typeof import('bootstrap-icons-vue')['BIconGenderTrans']
-  const BIconGeo: typeof import('bootstrap-icons-vue')['BIconGeo']
-  const BIconGeoAlt: typeof import('bootstrap-icons-vue')['BIconGeoAlt']
-  const BIconGeoAltFill: typeof import('bootstrap-icons-vue')['BIconGeoAltFill']
-  const BIconGeoFill: typeof import('bootstrap-icons-vue')['BIconGeoFill']
-  const BIconGift: typeof import('bootstrap-icons-vue')['BIconGift']
-  const BIconGiftFill: typeof import('bootstrap-icons-vue')['BIconGiftFill']
-  const BIconGit: typeof import('bootstrap-icons-vue')['BIconGit']
-  const BIconGithub: typeof import('bootstrap-icons-vue')['BIconGithub']
-  const BIconGlobe2: typeof import('bootstrap-icons-vue')['BIconGlobe2']
-  const BIconGlobe: typeof import('bootstrap-icons-vue')['BIconGlobe']
-  const BIconGoogle: typeof import('bootstrap-icons-vue')['BIconGoogle']
-  const BIconGpuCard: typeof import('bootstrap-icons-vue')['BIconGpuCard']
-  const BIconGraphDown: typeof import('bootstrap-icons-vue')['BIconGraphDown']
-  const BIconGraphDownArrow: typeof import('bootstrap-icons-vue')['BIconGraphDownArrow']
-  const BIconGraphUp: typeof import('bootstrap-icons-vue')['BIconGraphUp']
-  const BIconGraphUpArrow: typeof import('bootstrap-icons-vue')['BIconGraphUpArrow']
-  const BIconGrid1x2: typeof import('bootstrap-icons-vue')['BIconGrid1x2']
-  const BIconGrid1x2Fill: typeof import('bootstrap-icons-vue')['BIconGrid1x2Fill']
-  const BIconGrid3x2: typeof import('bootstrap-icons-vue')['BIconGrid3x2']
-  const BIconGrid3x2Gap: typeof import('bootstrap-icons-vue')['BIconGrid3x2Gap']
-  const BIconGrid3x2GapFill: typeof import('bootstrap-icons-vue')['BIconGrid3x2GapFill']
-  const BIconGrid3x3: typeof import('bootstrap-icons-vue')['BIconGrid3x3']
-  const BIconGrid3x3Gap: typeof import('bootstrap-icons-vue')['BIconGrid3x3Gap']
-  const BIconGrid3x3GapFill: typeof import('bootstrap-icons-vue')['BIconGrid3x3GapFill']
-  const BIconGrid: typeof import('bootstrap-icons-vue')['BIconGrid']
-  const BIconGridFill: typeof import('bootstrap-icons-vue')['BIconGridFill']
-  const BIconGripHorizontal: typeof import('bootstrap-icons-vue')['BIconGripHorizontal']
-  const BIconGripVertical: typeof import('bootstrap-icons-vue')['BIconGripVertical']
-  const BIconHammer: typeof import('bootstrap-icons-vue')['BIconHammer']
-  const BIconHandIndex: typeof import('bootstrap-icons-vue')['BIconHandIndex']
-  const BIconHandIndexFill: typeof import('bootstrap-icons-vue')['BIconHandIndexFill']
-  const BIconHandIndexThumb: typeof import('bootstrap-icons-vue')['BIconHandIndexThumb']
-  const BIconHandIndexThumbFill: typeof import('bootstrap-icons-vue')['BIconHandIndexThumbFill']
-  const BIconHandThumbsDown: typeof import('bootstrap-icons-vue')['BIconHandThumbsDown']
-  const BIconHandThumbsDownFill: typeof import('bootstrap-icons-vue')['BIconHandThumbsDownFill']
-  const BIconHandThumbsUp: typeof import('bootstrap-icons-vue')['BIconHandThumbsUp']
-  const BIconHandThumbsUpFill: typeof import('bootstrap-icons-vue')['BIconHandThumbsUpFill']
-  const BIconHandbag: typeof import('bootstrap-icons-vue')['BIconHandbag']
-  const BIconHandbagFill: typeof import('bootstrap-icons-vue')['BIconHandbagFill']
-  const BIconHash: typeof import('bootstrap-icons-vue')['BIconHash']
-  const BIconHdd: typeof import('bootstrap-icons-vue')['BIconHdd']
-  const BIconHddFill: typeof import('bootstrap-icons-vue')['BIconHddFill']
-  const BIconHddNetwork: typeof import('bootstrap-icons-vue')['BIconHddNetwork']
-  const BIconHddNetworkFill: typeof import('bootstrap-icons-vue')['BIconHddNetworkFill']
-  const BIconHddRack: typeof import('bootstrap-icons-vue')['BIconHddRack']
-  const BIconHddRackFill: typeof import('bootstrap-icons-vue')['BIconHddRackFill']
-  const BIconHddStack: typeof import('bootstrap-icons-vue')['BIconHddStack']
-  const BIconHddStackFill: typeof import('bootstrap-icons-vue')['BIconHddStackFill']
-  const BIconHdmi: typeof import('bootstrap-icons-vue')['BIconHdmi']
-  const BIconHdmiFill: typeof import('bootstrap-icons-vue')['BIconHdmiFill']
-  const BIconHeadphones: typeof import('bootstrap-icons-vue')['BIconHeadphones']
-  const BIconHeadset: typeof import('bootstrap-icons-vue')['BIconHeadset']
-  const BIconHeadsetVr: typeof import('bootstrap-icons-vue')['BIconHeadsetVr']
-  const BIconHeart: typeof import('bootstrap-icons-vue')['BIconHeart']
-  const BIconHeartArrow: typeof import('bootstrap-icons-vue')['BIconHeartArrow']
-  const BIconHeartFill: typeof import('bootstrap-icons-vue')['BIconHeartFill']
-  const BIconHeartHalf: typeof import('bootstrap-icons-vue')['BIconHeartHalf']
-  const BIconHeartPulse: typeof import('bootstrap-icons-vue')['BIconHeartPulse']
-  const BIconHeartPulseFill: typeof import('bootstrap-icons-vue')['BIconHeartPulseFill']
-  const BIconHeartbreak: typeof import('bootstrap-icons-vue')['BIconHeartbreak']
-  const BIconHeartbreakFill: typeof import('bootstrap-icons-vue')['BIconHeartbreakFill']
-  const BIconHearts: typeof import('bootstrap-icons-vue')['BIconHearts']
-  const BIconHeptagon: typeof import('bootstrap-icons-vue')['BIconHeptagon']
-  const BIconHeptagonFill: typeof import('bootstrap-icons-vue')['BIconHeptagonFill']
-  const BIconHeptagonHalf: typeof import('bootstrap-icons-vue')['BIconHeptagonHalf']
-  const BIconHexagon: typeof import('bootstrap-icons-vue')['BIconHexagon']
-  const BIconHexagonFill: typeof import('bootstrap-icons-vue')['BIconHexagonFill']
-  const BIconHexagonHalf: typeof import('bootstrap-icons-vue')['BIconHexagonHalf']
-  const BIconHospital: typeof import('bootstrap-icons-vue')['BIconHospital']
-  const BIconHospitalFill: typeof import('bootstrap-icons-vue')['BIconHospitalFill']
-  const BIconHourglass: typeof import('bootstrap-icons-vue')['BIconHourglass']
-  const BIconHourglassBottom: typeof import('bootstrap-icons-vue')['BIconHourglassBottom']
-  const BIconHourglassSplit: typeof import('bootstrap-icons-vue')['BIconHourglassSplit']
-  const BIconHourglassTop: typeof import('bootstrap-icons-vue')['BIconHourglassTop']
-  const BIconHouse: typeof import('bootstrap-icons-vue')['BIconHouse']
-  const BIconHouseDoor: typeof import('bootstrap-icons-vue')['BIconHouseDoor']
-  const BIconHouseDoorFill: typeof import('bootstrap-icons-vue')['BIconHouseDoorFill']
-  const BIconHouseFill: typeof import('bootstrap-icons-vue')['BIconHouseFill']
-  const BIconHouseHeart: typeof import('bootstrap-icons-vue')['BIconHouseHeart']
-  const BIconHouseHeartFill: typeof import('bootstrap-icons-vue')['BIconHouseHeartFill']
-  const BIconHr: typeof import('bootstrap-icons-vue')['BIconHr']
-  const BIconHurricane: typeof import('bootstrap-icons-vue')['BIconHurricane']
-  const BIconHypnotize: typeof import('bootstrap-icons-vue')['BIconHypnotize']
-  const BIconImage: typeof import('bootstrap-icons-vue')['BIconImage']
-  const BIconImageAlt: typeof import('bootstrap-icons-vue')['BIconImageAlt']
-  const BIconImageFill: typeof import('bootstrap-icons-vue')['BIconImageFill']
-  const BIconImages: typeof import('bootstrap-icons-vue')['BIconImages']
-  const BIconInbox: typeof import('bootstrap-icons-vue')['BIconInbox']
-  const BIconInboxFill: typeof import('bootstrap-icons-vue')['BIconInboxFill']
-  const BIconInboxes: typeof import('bootstrap-icons-vue')['BIconInboxes']
-  const BIconInboxesFill: typeof import('bootstrap-icons-vue')['BIconInboxesFill']
-  const BIconIncognito: typeof import('bootstrap-icons-vue')['BIconIncognito']
-  const BIconInfinity: typeof import('bootstrap-icons-vue')['BIconInfinity']
-  const BIconInfo: typeof import('bootstrap-icons-vue')['BIconInfo']
-  const BIconInfoCircle: typeof import('bootstrap-icons-vue')['BIconInfoCircle']
-  const BIconInfoCircleFill: typeof import('bootstrap-icons-vue')['BIconInfoCircleFill']
-  const BIconInfoLg: typeof import('bootstrap-icons-vue')['BIconInfoLg']
-  const BIconInfoSquare: typeof import('bootstrap-icons-vue')['BIconInfoSquare']
-  const BIconInfoSquareFill: typeof import('bootstrap-icons-vue')['BIconInfoSquareFill']
-  const BIconInputCursor: typeof import('bootstrap-icons-vue')['BIconInputCursor']
-  const BIconInputCursorText: typeof import('bootstrap-icons-vue')['BIconInputCursorText']
-  const BIconInstagram: typeof import('bootstrap-icons-vue')['BIconInstagram']
-  const BIconIntersect: typeof import('bootstrap-icons-vue')['BIconIntersect']
-  const BIconJournal: typeof import('bootstrap-icons-vue')['BIconJournal']
-  const BIconJournalAlbum: typeof import('bootstrap-icons-vue')['BIconJournalAlbum']
-  const BIconJournalArrowDown: typeof import('bootstrap-icons-vue')['BIconJournalArrowDown']
-  const BIconJournalArrowUp: typeof import('bootstrap-icons-vue')['BIconJournalArrowUp']
-  const BIconJournalBookmark: typeof import('bootstrap-icons-vue')['BIconJournalBookmark']
-  const BIconJournalBookmarkFill: typeof import('bootstrap-icons-vue')['BIconJournalBookmarkFill']
-  const BIconJournalCheck: typeof import('bootstrap-icons-vue')['BIconJournalCheck']
-  const BIconJournalCode: typeof import('bootstrap-icons-vue')['BIconJournalCode']
-  const BIconJournalMedical: typeof import('bootstrap-icons-vue')['BIconJournalMedical']
-  const BIconJournalMinus: typeof import('bootstrap-icons-vue')['BIconJournalMinus']
-  const BIconJournalPlus: typeof import('bootstrap-icons-vue')['BIconJournalPlus']
-  const BIconJournalRichtext: typeof import('bootstrap-icons-vue')['BIconJournalRichtext']
-  const BIconJournalText: typeof import('bootstrap-icons-vue')['BIconJournalText']
-  const BIconJournalX: typeof import('bootstrap-icons-vue')['BIconJournalX']
-  const BIconJournals: typeof import('bootstrap-icons-vue')['BIconJournals']
-  const BIconJoystick: typeof import('bootstrap-icons-vue')['BIconJoystick']
-  const BIconJustify: typeof import('bootstrap-icons-vue')['BIconJustify']
-  const BIconJustifyLeft: typeof import('bootstrap-icons-vue')['BIconJustifyLeft']
-  const BIconJustifyRight: typeof import('bootstrap-icons-vue')['BIconJustifyRight']
-  const BIconKanban: typeof import('bootstrap-icons-vue')['BIconKanban']
-  const BIconKanbanFill: typeof import('bootstrap-icons-vue')['BIconKanbanFill']
-  const BIconKey: typeof import('bootstrap-icons-vue')['BIconKey']
-  const BIconKeyFill: typeof import('bootstrap-icons-vue')['BIconKeyFill']
-  const BIconKeyboard: typeof import('bootstrap-icons-vue')['BIconKeyboard']
-  const BIconKeyboardFill: typeof import('bootstrap-icons-vue')['BIconKeyboardFill']
-  const BIconLadder: typeof import('bootstrap-icons-vue')['BIconLadder']
-  const BIconLamp: typeof import('bootstrap-icons-vue')['BIconLamp']
-  const BIconLampFill: typeof import('bootstrap-icons-vue')['BIconLampFill']
-  const BIconLaptop: typeof import('bootstrap-icons-vue')['BIconLaptop']
-  const BIconLaptopFill: typeof import('bootstrap-icons-vue')['BIconLaptopFill']
-  const BIconLayerBackward: typeof import('bootstrap-icons-vue')['BIconLayerBackward']
-  const BIconLayerForward: typeof import('bootstrap-icons-vue')['BIconLayerForward']
-  const BIconLayers: typeof import('bootstrap-icons-vue')['BIconLayers']
-  const BIconLayersFill: typeof import('bootstrap-icons-vue')['BIconLayersFill']
-  const BIconLayersHalf: typeof import('bootstrap-icons-vue')['BIconLayersHalf']
-  const BIconLayoutSidebar: typeof import('bootstrap-icons-vue')['BIconLayoutSidebar']
-  const BIconLayoutSidebarInset: typeof import('bootstrap-icons-vue')['BIconLayoutSidebarInset']
-  const BIconLayoutSidebarInsetReverse: typeof import('bootstrap-icons-vue')['BIconLayoutSidebarInsetReverse']
-  const BIconLayoutSidebarReverse: typeof import('bootstrap-icons-vue')['BIconLayoutSidebarReverse']
-  const BIconLayoutSplit: typeof import('bootstrap-icons-vue')['BIconLayoutSplit']
-  const BIconLayoutTextSidebar: typeof import('bootstrap-icons-vue')['BIconLayoutTextSidebar']
-  const BIconLayoutTextSidebarReverse: typeof import('bootstrap-icons-vue')['BIconLayoutTextSidebarReverse']
-  const BIconLayoutTextWindow: typeof import('bootstrap-icons-vue')['BIconLayoutTextWindow']
-  const BIconLayoutTextWindowReverse: typeof import('bootstrap-icons-vue')['BIconLayoutTextWindowReverse']
-  const BIconLayoutThreeColumns: typeof import('bootstrap-icons-vue')['BIconLayoutThreeColumns']
-  const BIconLayoutWtf: typeof import('bootstrap-icons-vue')['BIconLayoutWtf']
-  const BIconLifePreserver: typeof import('bootstrap-icons-vue')['BIconLifePreserver']
-  const BIconLightbulb: typeof import('bootstrap-icons-vue')['BIconLightbulb']
-  const BIconLightbulbFill: typeof import('bootstrap-icons-vue')['BIconLightbulbFill']
-  const BIconLightbulbOff: typeof import('bootstrap-icons-vue')['BIconLightbulbOff']
-  const BIconLightbulbOffFill: typeof import('bootstrap-icons-vue')['BIconLightbulbOffFill']
-  const BIconLightning: typeof import('bootstrap-icons-vue')['BIconLightning']
-  const BIconLightningCharge: typeof import('bootstrap-icons-vue')['BIconLightningCharge']
-  const BIconLightningChargeFill: typeof import('bootstrap-icons-vue')['BIconLightningChargeFill']
-  const BIconLightningFill: typeof import('bootstrap-icons-vue')['BIconLightningFill']
-  const BIconLine: typeof import('bootstrap-icons-vue')['BIconLine']
-  const BIconLink45deg: typeof import('bootstrap-icons-vue')['BIconLink45deg']
-  const BIconLink: typeof import('bootstrap-icons-vue')['BIconLink']
-  const BIconLinkedin: typeof import('bootstrap-icons-vue')['BIconLinkedin']
-  const BIconList: typeof import('bootstrap-icons-vue')['BIconList']
-  const BIconListCheck: typeof import('bootstrap-icons-vue')['BIconListCheck']
-  const BIconListColumns: typeof import('bootstrap-icons-vue')['BIconListColumns']
-  const BIconListColumnsReverse: typeof import('bootstrap-icons-vue')['BIconListColumnsReverse']
-  const BIconListNested: typeof import('bootstrap-icons-vue')['BIconListNested']
-  const BIconListOl: typeof import('bootstrap-icons-vue')['BIconListOl']
-  const BIconListStars: typeof import('bootstrap-icons-vue')['BIconListStars']
-  const BIconListTask: typeof import('bootstrap-icons-vue')['BIconListTask']
-  const BIconListUl: typeof import('bootstrap-icons-vue')['BIconListUl']
-  const BIconLock: typeof import('bootstrap-icons-vue')['BIconLock']
-  const BIconLockFill: typeof import('bootstrap-icons-vue')['BIconLockFill']
-  const BIconMagic: typeof import('bootstrap-icons-vue')['BIconMagic']
-  const BIconMagnet: typeof import('bootstrap-icons-vue')['BIconMagnet']
-  const BIconMagnetFill: typeof import('bootstrap-icons-vue')['BIconMagnetFill']
-  const BIconMailbox2: typeof import('bootstrap-icons-vue')['BIconMailbox2']
-  const BIconMailbox: typeof import('bootstrap-icons-vue')['BIconMailbox']
-  const BIconMap: typeof import('bootstrap-icons-vue')['BIconMap']
-  const BIconMapFill: typeof import('bootstrap-icons-vue')['BIconMapFill']
-  const BIconMarkdown: typeof import('bootstrap-icons-vue')['BIconMarkdown']
-  const BIconMarkdownFill: typeof import('bootstrap-icons-vue')['BIconMarkdownFill']
-  const BIconMask: typeof import('bootstrap-icons-vue')['BIconMask']
-  const BIconMastodon: typeof import('bootstrap-icons-vue')['BIconMastodon']
-  const BIconMedium: typeof import('bootstrap-icons-vue')['BIconMedium']
-  const BIconMegaphone: typeof import('bootstrap-icons-vue')['BIconMegaphone']
-  const BIconMegaphoneFill: typeof import('bootstrap-icons-vue')['BIconMegaphoneFill']
-  const BIconMemory: typeof import('bootstrap-icons-vue')['BIconMemory']
-  const BIconMenuApp: typeof import('bootstrap-icons-vue')['BIconMenuApp']
-  const BIconMenuAppFill: typeof import('bootstrap-icons-vue')['BIconMenuAppFill']
-  const BIconMenuButton: typeof import('bootstrap-icons-vue')['BIconMenuButton']
-  const BIconMenuButtonFill: typeof import('bootstrap-icons-vue')['BIconMenuButtonFill']
-  const BIconMenuButtonWide: typeof import('bootstrap-icons-vue')['BIconMenuButtonWide']
-  const BIconMenuButtonWideFill: typeof import('bootstrap-icons-vue')['BIconMenuButtonWideFill']
-  const BIconMenuDown: typeof import('bootstrap-icons-vue')['BIconMenuDown']
-  const BIconMenuUp: typeof import('bootstrap-icons-vue')['BIconMenuUp']
-  const BIconMessenger: typeof import('bootstrap-icons-vue')['BIconMessenger']
-  const BIconMeta: typeof import('bootstrap-icons-vue')['BIconMeta']
-  const BIconMic: typeof import('bootstrap-icons-vue')['BIconMic']
-  const BIconMicFill: typeof import('bootstrap-icons-vue')['BIconMicFill']
-  const BIconMicMute: typeof import('bootstrap-icons-vue')['BIconMicMute']
-  const BIconMicMuteFill: typeof import('bootstrap-icons-vue')['BIconMicMuteFill']
-  const BIconMicrosoft: typeof import('bootstrap-icons-vue')['BIconMicrosoft']
-  const BIconMinecart: typeof import('bootstrap-icons-vue')['BIconMinecart']
-  const BIconMinecartLoaded: typeof import('bootstrap-icons-vue')['BIconMinecartLoaded']
-  const BIconModem: typeof import('bootstrap-icons-vue')['BIconModem']
-  const BIconModemFill: typeof import('bootstrap-icons-vue')['BIconModemFill']
-  const BIconMoisture: typeof import('bootstrap-icons-vue')['BIconMoisture']
-  const BIconMoon: typeof import('bootstrap-icons-vue')['BIconMoon']
-  const BIconMoonFill: typeof import('bootstrap-icons-vue')['BIconMoonFill']
-  const BIconMoonStars: typeof import('bootstrap-icons-vue')['BIconMoonStars']
-  const BIconMoonStarsFill: typeof import('bootstrap-icons-vue')['BIconMoonStarsFill']
-  const BIconMortarboard: typeof import('bootstrap-icons-vue')['BIconMortarboard']
-  const BIconMortarboardFill: typeof import('bootstrap-icons-vue')['BIconMortarboardFill']
-  const BIconMotherboard: typeof import('bootstrap-icons-vue')['BIconMotherboard']
-  const BIconMotherboardFill: typeof import('bootstrap-icons-vue')['BIconMotherboardFill']
-  const BIconMouse2: typeof import('bootstrap-icons-vue')['BIconMouse2']
-  const BIconMouse2Fill: typeof import('bootstrap-icons-vue')['BIconMouse2Fill']
-  const BIconMouse3: typeof import('bootstrap-icons-vue')['BIconMouse3']
-  const BIconMouse3Fill: typeof import('bootstrap-icons-vue')['BIconMouse3Fill']
-  const BIconMouse: typeof import('bootstrap-icons-vue')['BIconMouse']
-  const BIconMouseFill: typeof import('bootstrap-icons-vue')['BIconMouseFill']
-  const BIconMusicNote: typeof import('bootstrap-icons-vue')['BIconMusicNote']
-  const BIconMusicNoteBeamed: typeof import('bootstrap-icons-vue')['BIconMusicNoteBeamed']
-  const BIconMusicNoteList: typeof import('bootstrap-icons-vue')['BIconMusicNoteList']
-  const BIconMusicPlayer: typeof import('bootstrap-icons-vue')['BIconMusicPlayer']
-  const BIconMusicPlayerFill: typeof import('bootstrap-icons-vue')['BIconMusicPlayerFill']
-  const BIconNewspaper: typeof import('bootstrap-icons-vue')['BIconNewspaper']
-  const BIconNintendoSwitch: typeof import('bootstrap-icons-vue')['BIconNintendoSwitch']
-  const BIconNodeMinus: typeof import('bootstrap-icons-vue')['BIconNodeMinus']
-  const BIconNodeMinusFill: typeof import('bootstrap-icons-vue')['BIconNodeMinusFill']
-  const BIconNodePlus: typeof import('bootstrap-icons-vue')['BIconNodePlus']
-  const BIconNodePlusFill: typeof import('bootstrap-icons-vue')['BIconNodePlusFill']
-  const BIconNut: typeof import('bootstrap-icons-vue')['BIconNut']
-  const BIconNutFill: typeof import('bootstrap-icons-vue')['BIconNutFill']
-  const BIconOctagon: typeof import('bootstrap-icons-vue')['BIconOctagon']
-  const BIconOctagonFill: typeof import('bootstrap-icons-vue')['BIconOctagonFill']
-  const BIconOctagonHalf: typeof import('bootstrap-icons-vue')['BIconOctagonHalf']
-  const BIconOpticalAudio: typeof import('bootstrap-icons-vue')['BIconOpticalAudio']
-  const BIconOpticalAudioFill: typeof import('bootstrap-icons-vue')['BIconOpticalAudioFill']
-  const BIconOption: typeof import('bootstrap-icons-vue')['BIconOption']
-  const BIconOutlet: typeof import('bootstrap-icons-vue')['BIconOutlet']
-  const BIconPaintBucket: typeof import('bootstrap-icons-vue')['BIconPaintBucket']
-  const BIconPalette2: typeof import('bootstrap-icons-vue')['BIconPalette2']
-  const BIconPalette: typeof import('bootstrap-icons-vue')['BIconPalette']
-  const BIconPaletteFill: typeof import('bootstrap-icons-vue')['BIconPaletteFill']
-  const BIconPaperclip: typeof import('bootstrap-icons-vue')['BIconPaperclip']
-  const BIconParagraph: typeof import('bootstrap-icons-vue')['BIconParagraph']
-  const BIconPatchCheck: typeof import('bootstrap-icons-vue')['BIconPatchCheck']
-  const BIconPatchCheckFill: typeof import('bootstrap-icons-vue')['BIconPatchCheckFill']
-  const BIconPatchExclamation: typeof import('bootstrap-icons-vue')['BIconPatchExclamation']
-  const BIconPatchExclamationFill: typeof import('bootstrap-icons-vue')['BIconPatchExclamationFill']
-  const BIconPatchMinus: typeof import('bootstrap-icons-vue')['BIconPatchMinus']
-  const BIconPatchMinusFill: typeof import('bootstrap-icons-vue')['BIconPatchMinusFill']
-  const BIconPatchPlus: typeof import('bootstrap-icons-vue')['BIconPatchPlus']
-  const BIconPatchPlusFill: typeof import('bootstrap-icons-vue')['BIconPatchPlusFill']
-  const BIconPatchQuestion: typeof import('bootstrap-icons-vue')['BIconPatchQuestion']
-  const BIconPatchQuestionFill: typeof import('bootstrap-icons-vue')['BIconPatchQuestionFill']
-  const BIconPause: typeof import('bootstrap-icons-vue')['BIconPause']
-  const BIconPauseBtn: typeof import('bootstrap-icons-vue')['BIconPauseBtn']
-  const BIconPauseBtnFill: typeof import('bootstrap-icons-vue')['BIconPauseBtnFill']
-  const BIconPauseCircle: typeof import('bootstrap-icons-vue')['BIconPauseCircle']
-  const BIconPauseCircleFill: typeof import('bootstrap-icons-vue')['BIconPauseCircleFill']
-  const BIconPauseFill: typeof import('bootstrap-icons-vue')['BIconPauseFill']
-  const BIconPaypal: typeof import('bootstrap-icons-vue')['BIconPaypal']
-  const BIconPc: typeof import('bootstrap-icons-vue')['BIconPc']
-  const BIconPcDisplay: typeof import('bootstrap-icons-vue')['BIconPcDisplay']
-  const BIconPcDisplayHorizontal: typeof import('bootstrap-icons-vue')['BIconPcDisplayHorizontal']
-  const BIconPcHorizontal: typeof import('bootstrap-icons-vue')['BIconPcHorizontal']
-  const BIconPciCard: typeof import('bootstrap-icons-vue')['BIconPciCard']
-  const BIconPeace: typeof import('bootstrap-icons-vue')['BIconPeace']
-  const BIconPeaceFill: typeof import('bootstrap-icons-vue')['BIconPeaceFill']
-  const BIconPen: typeof import('bootstrap-icons-vue')['BIconPen']
-  const BIconPenFill: typeof import('bootstrap-icons-vue')['BIconPenFill']
-  const BIconPencil: typeof import('bootstrap-icons-vue')['BIconPencil']
-  const BIconPencilFill: typeof import('bootstrap-icons-vue')['BIconPencilFill']
-  const BIconPencilSquare: typeof import('bootstrap-icons-vue')['BIconPencilSquare']
-  const BIconPentagon: typeof import('bootstrap-icons-vue')['BIconPentagon']
-  const BIconPentagonFill: typeof import('bootstrap-icons-vue')['BIconPentagonFill']
-  const BIconPentagonHalf: typeof import('bootstrap-icons-vue')['BIconPentagonHalf']
-  const BIconPeople: typeof import('bootstrap-icons-vue')['BIconPeople']
-  const BIconPeopleFill: typeof import('bootstrap-icons-vue')['BIconPeopleFill']
-  const BIconPercent: typeof import('bootstrap-icons-vue')['BIconPercent']
-  const BIconPerson: typeof import('bootstrap-icons-vue')['BIconPerson']
-  const BIconPersonBadge: typeof import('bootstrap-icons-vue')['BIconPersonBadge']
-  const BIconPersonBadgeFill: typeof import('bootstrap-icons-vue')['BIconPersonBadgeFill']
-  const BIconPersonBoundingBox: typeof import('bootstrap-icons-vue')['BIconPersonBoundingBox']
-  const BIconPersonCheck: typeof import('bootstrap-icons-vue')['BIconPersonCheck']
-  const BIconPersonCheckFill: typeof import('bootstrap-icons-vue')['BIconPersonCheckFill']
-  const BIconPersonCircle: typeof import('bootstrap-icons-vue')['BIconPersonCircle']
-  const BIconPersonDash: typeof import('bootstrap-icons-vue')['BIconPersonDash']
-  const BIconPersonDashFill: typeof import('bootstrap-icons-vue')['BIconPersonDashFill']
-  const BIconPersonFill: typeof import('bootstrap-icons-vue')['BIconPersonFill']
-  const BIconPersonHeart: typeof import('bootstrap-icons-vue')['BIconPersonHeart']
-  const BIconPersonHearts: typeof import('bootstrap-icons-vue')['BIconPersonHearts']
-  const BIconPersonLinesFill: typeof import('bootstrap-icons-vue')['BIconPersonLinesFill']
-  const BIconPersonPlus: typeof import('bootstrap-icons-vue')['BIconPersonPlus']
-  const BIconPersonPlusFill: typeof import('bootstrap-icons-vue')['BIconPersonPlusFill']
-  const BIconPersonRolodex: typeof import('bootstrap-icons-vue')['BIconPersonRolodex']
-  const BIconPersonSquare: typeof import('bootstrap-icons-vue')['BIconPersonSquare']
-  const BIconPersonVideo2: typeof import('bootstrap-icons-vue')['BIconPersonVideo2']
-  const BIconPersonVideo3: typeof import('bootstrap-icons-vue')['BIconPersonVideo3']
-  const BIconPersonVideo: typeof import('bootstrap-icons-vue')['BIconPersonVideo']
-  const BIconPersonWorkspace: typeof import('bootstrap-icons-vue')['BIconPersonWorkspace']
-  const BIconPersonX: typeof import('bootstrap-icons-vue')['BIconPersonX']
-  const BIconPersonXFill: typeof import('bootstrap-icons-vue')['BIconPersonXFill']
-  const BIconPhone: typeof import('bootstrap-icons-vue')['BIconPhone']
-  const BIconPhoneFill: typeof import('bootstrap-icons-vue')['BIconPhoneFill']
-  const BIconPhoneFlip: typeof import('bootstrap-icons-vue')['BIconPhoneFlip']
-  const BIconPhoneLandscape: typeof import('bootstrap-icons-vue')['BIconPhoneLandscape']
-  const BIconPhoneLandscapeFill: typeof import('bootstrap-icons-vue')['BIconPhoneLandscapeFill']
-  const BIconPhoneVibrate: typeof import('bootstrap-icons-vue')['BIconPhoneVibrate']
-  const BIconPhoneVibrateFill: typeof import('bootstrap-icons-vue')['BIconPhoneVibrateFill']
-  const BIconPieChart: typeof import('bootstrap-icons-vue')['BIconPieChart']
-  const BIconPieChartFill: typeof import('bootstrap-icons-vue')['BIconPieChartFill']
-  const BIconPiggyBank: typeof import('bootstrap-icons-vue')['BIconPiggyBank']
-  const BIconPiggyBankFill: typeof import('bootstrap-icons-vue')['BIconPiggyBankFill']
-  const BIconPin: typeof import('bootstrap-icons-vue')['BIconPin']
-  const BIconPinAngle: typeof import('bootstrap-icons-vue')['BIconPinAngle']
-  const BIconPinAngleFill: typeof import('bootstrap-icons-vue')['BIconPinAngleFill']
-  const BIconPinFill: typeof import('bootstrap-icons-vue')['BIconPinFill']
-  const BIconPinMap: typeof import('bootstrap-icons-vue')['BIconPinMap']
-  const BIconPinMapFill: typeof import('bootstrap-icons-vue')['BIconPinMapFill']
-  const BIconPinterest: typeof import('bootstrap-icons-vue')['BIconPinterest']
-  const BIconPip: typeof import('bootstrap-icons-vue')['BIconPip']
-  const BIconPipFill: typeof import('bootstrap-icons-vue')['BIconPipFill']
-  const BIconPlay: typeof import('bootstrap-icons-vue')['BIconPlay']
-  const BIconPlayBtn: typeof import('bootstrap-icons-vue')['BIconPlayBtn']
-  const BIconPlayBtnFill: typeof import('bootstrap-icons-vue')['BIconPlayBtnFill']
-  const BIconPlayCircle: typeof import('bootstrap-icons-vue')['BIconPlayCircle']
-  const BIconPlayCircleFill: typeof import('bootstrap-icons-vue')['BIconPlayCircleFill']
-  const BIconPlayFill: typeof import('bootstrap-icons-vue')['BIconPlayFill']
-  const BIconPlaystation: typeof import('bootstrap-icons-vue')['BIconPlaystation']
-  const BIconPlug: typeof import('bootstrap-icons-vue')['BIconPlug']
-  const BIconPlugFill: typeof import('bootstrap-icons-vue')['BIconPlugFill']
-  const BIconPlugin: typeof import('bootstrap-icons-vue')['BIconPlugin']
-  const BIconPlus: typeof import('bootstrap-icons-vue')['BIconPlus']
-  const BIconPlusCircle: typeof import('bootstrap-icons-vue')['BIconPlusCircle']
-  const BIconPlusCircleDotted: typeof import('bootstrap-icons-vue')['BIconPlusCircleDotted']
-  const BIconPlusCircleFill: typeof import('bootstrap-icons-vue')['BIconPlusCircleFill']
-  const BIconPlusLg: typeof import('bootstrap-icons-vue')['BIconPlusLg']
-  const BIconPlusSlashMinus: typeof import('bootstrap-icons-vue')['BIconPlusSlashMinus']
-  const BIconPlusSquare: typeof import('bootstrap-icons-vue')['BIconPlusSquare']
-  const BIconPlusSquareDotted: typeof import('bootstrap-icons-vue')['BIconPlusSquareDotted']
-  const BIconPlusSquareFill: typeof import('bootstrap-icons-vue')['BIconPlusSquareFill']
-  const BIconPostage: typeof import('bootstrap-icons-vue')['BIconPostage']
-  const BIconPostageFill: typeof import('bootstrap-icons-vue')['BIconPostageFill']
-  const BIconPostageHeart: typeof import('bootstrap-icons-vue')['BIconPostageHeart']
-  const BIconPostageHeartFill: typeof import('bootstrap-icons-vue')['BIconPostageHeartFill']
-  const BIconPostcard: typeof import('bootstrap-icons-vue')['BIconPostcard']
-  const BIconPostcardFill: typeof import('bootstrap-icons-vue')['BIconPostcardFill']
-  const BIconPostcardHeart: typeof import('bootstrap-icons-vue')['BIconPostcardHeart']
-  const BIconPostcardHeartFill: typeof import('bootstrap-icons-vue')['BIconPostcardHeartFill']
-  const BIconPower: typeof import('bootstrap-icons-vue')['BIconPower']
-  const BIconPrinter: typeof import('bootstrap-icons-vue')['BIconPrinter']
-  const BIconPrinterFill: typeof import('bootstrap-icons-vue')['BIconPrinterFill']
-  const BIconProjector: typeof import('bootstrap-icons-vue')['BIconProjector']
-  const BIconProjectorFill: typeof import('bootstrap-icons-vue')['BIconProjectorFill']
-  const BIconPuzzle: typeof import('bootstrap-icons-vue')['BIconPuzzle']
-  const BIconPuzzleFill: typeof import('bootstrap-icons-vue')['BIconPuzzleFill']
-  const BIconQrCode: typeof import('bootstrap-icons-vue')['BIconQrCode']
-  const BIconQrCodeScan: typeof import('bootstrap-icons-vue')['BIconQrCodeScan']
-  const BIconQuestion: typeof import('bootstrap-icons-vue')['BIconQuestion']
-  const BIconQuestionCircle: typeof import('bootstrap-icons-vue')['BIconQuestionCircle']
-  const BIconQuestionCircleFill: typeof import('bootstrap-icons-vue')['BIconQuestionCircleFill']
-  const BIconQuestionDiamond: typeof import('bootstrap-icons-vue')['BIconQuestionDiamond']
-  const BIconQuestionDiamondFill: typeof import('bootstrap-icons-vue')['BIconQuestionDiamondFill']
-  const BIconQuestionLg: typeof import('bootstrap-icons-vue')['BIconQuestionLg']
-  const BIconQuestionOctagon: typeof import('bootstrap-icons-vue')['BIconQuestionOctagon']
-  const BIconQuestionOctagonFill: typeof import('bootstrap-icons-vue')['BIconQuestionOctagonFill']
-  const BIconQuestionSquare: typeof import('bootstrap-icons-vue')['BIconQuestionSquare']
-  const BIconQuestionSquareFill: typeof import('bootstrap-icons-vue')['BIconQuestionSquareFill']
-  const BIconQuora: typeof import('bootstrap-icons-vue')['BIconQuora']
-  const BIconQuote: typeof import('bootstrap-icons-vue')['BIconQuote']
-  const BIconRadioactive: typeof import('bootstrap-icons-vue')['BIconRadioactive']
-  const BIconRainbow: typeof import('bootstrap-icons-vue')['BIconRainbow']
-  const BIconReceipt: typeof import('bootstrap-icons-vue')['BIconReceipt']
-  const BIconReceiptCutoff: typeof import('bootstrap-icons-vue')['BIconReceiptCutoff']
-  const BIconReception0: typeof import('bootstrap-icons-vue')['BIconReception0']
-  const BIconReception1: typeof import('bootstrap-icons-vue')['BIconReception1']
-  const BIconReception2: typeof import('bootstrap-icons-vue')['BIconReception2']
-  const BIconReception3: typeof import('bootstrap-icons-vue')['BIconReception3']
-  const BIconReception4: typeof import('bootstrap-icons-vue')['BIconReception4']
-  const BIconRecord2: typeof import('bootstrap-icons-vue')['BIconRecord2']
-  const BIconRecord2Fill: typeof import('bootstrap-icons-vue')['BIconRecord2Fill']
-  const BIconRecord: typeof import('bootstrap-icons-vue')['BIconRecord']
-  const BIconRecordBtn: typeof import('bootstrap-icons-vue')['BIconRecordBtn']
-  const BIconRecordBtnFill: typeof import('bootstrap-icons-vue')['BIconRecordBtnFill']
-  const BIconRecordCircle: typeof import('bootstrap-icons-vue')['BIconRecordCircle']
-  const BIconRecordCircleFill: typeof import('bootstrap-icons-vue')['BIconRecordCircleFill']
-  const BIconRecordFill: typeof import('bootstrap-icons-vue')['BIconRecordFill']
-  const BIconRecycle: typeof import('bootstrap-icons-vue')['BIconRecycle']
-  const BIconReddit: typeof import('bootstrap-icons-vue')['BIconReddit']
-  const BIconReply: typeof import('bootstrap-icons-vue')['BIconReply']
-  const BIconReplyAll: typeof import('bootstrap-icons-vue')['BIconReplyAll']
-  const BIconReplyAllFill: typeof import('bootstrap-icons-vue')['BIconReplyAllFill']
-  const BIconReplyFill: typeof import('bootstrap-icons-vue')['BIconReplyFill']
-  const BIconRobot: typeof import('bootstrap-icons-vue')['BIconRobot']
-  const BIconRouter: typeof import('bootstrap-icons-vue')['BIconRouter']
-  const BIconRouterFill: typeof import('bootstrap-icons-vue')['BIconRouterFill']
-  const BIconRss: typeof import('bootstrap-icons-vue')['BIconRss']
-  const BIconRssFill: typeof import('bootstrap-icons-vue')['BIconRssFill']
-  const BIconRulers: typeof import('bootstrap-icons-vue')['BIconRulers']
-  const BIconSafe2: typeof import('bootstrap-icons-vue')['BIconSafe2']
-  const BIconSafe2Fill: typeof import('bootstrap-icons-vue')['BIconSafe2Fill']
-  const BIconSafe: typeof import('bootstrap-icons-vue')['BIconSafe']
-  const BIconSafeFill: typeof import('bootstrap-icons-vue')['BIconSafeFill']
-  const BIconSave2: typeof import('bootstrap-icons-vue')['BIconSave2']
-  const BIconSave2Fill: typeof import('bootstrap-icons-vue')['BIconSave2Fill']
-  const BIconSave: typeof import('bootstrap-icons-vue')['BIconSave']
-  const BIconSaveFill: typeof import('bootstrap-icons-vue')['BIconSaveFill']
-  const BIconScissors: typeof import('bootstrap-icons-vue')['BIconScissors']
-  const BIconScrewdriver: typeof import('bootstrap-icons-vue')['BIconScrewdriver']
-  const BIconSdCard: typeof import('bootstrap-icons-vue')['BIconSdCard']
-  const BIconSdCardFill: typeof import('bootstrap-icons-vue')['BIconSdCardFill']
-  const BIconSearch: typeof import('bootstrap-icons-vue')['BIconSearch']
-  const BIconSearchHeart: typeof import('bootstrap-icons-vue')['BIconSearchHeart']
-  const BIconSearchHeartFill: typeof import('bootstrap-icons-vue')['BIconSearchHeartFill']
-  const BIconSegmentedNav: typeof import('bootstrap-icons-vue')['BIconSegmentedNav']
-  const BIconSend: typeof import('bootstrap-icons-vue')['BIconSend']
-  const BIconSendCheck: typeof import('bootstrap-icons-vue')['BIconSendCheck']
-  const BIconSendCheckFill: typeof import('bootstrap-icons-vue')['BIconSendCheckFill']
-  const BIconSendDash: typeof import('bootstrap-icons-vue')['BIconSendDash']
-  const BIconSendDashFill: typeof import('bootstrap-icons-vue')['BIconSendDashFill']
-  const BIconSendExclamation: typeof import('bootstrap-icons-vue')['BIconSendExclamation']
-  const BIconSendExclamationFill: typeof import('bootstrap-icons-vue')['BIconSendExclamationFill']
-  const BIconSendFill: typeof import('bootstrap-icons-vue')['BIconSendFill']
-  const BIconSendPlus: typeof import('bootstrap-icons-vue')['BIconSendPlus']
-  const BIconSendPlusFill: typeof import('bootstrap-icons-vue')['BIconSendPlusFill']
-  const BIconSendSlash: typeof import('bootstrap-icons-vue')['BIconSendSlash']
-  const BIconSendSlashFill: typeof import('bootstrap-icons-vue')['BIconSendSlashFill']
-  const BIconSendX: typeof import('bootstrap-icons-vue')['BIconSendX']
-  const BIconSendXFill: typeof import('bootstrap-icons-vue')['BIconSendXFill']
-  const BIconServer: typeof import('bootstrap-icons-vue')['BIconServer']
-  const BIconShare: typeof import('bootstrap-icons-vue')['BIconShare']
-  const BIconShareFill: typeof import('bootstrap-icons-vue')['BIconShareFill']
-  const BIconShield: typeof import('bootstrap-icons-vue')['BIconShield']
-  const BIconShieldCheck: typeof import('bootstrap-icons-vue')['BIconShieldCheck']
-  const BIconShieldExclamation: typeof import('bootstrap-icons-vue')['BIconShieldExclamation']
-  const BIconShieldFill: typeof import('bootstrap-icons-vue')['BIconShieldFill']
-  const BIconShieldFillCheck: typeof import('bootstrap-icons-vue')['BIconShieldFillCheck']
-  const BIconShieldFillExclamation: typeof import('bootstrap-icons-vue')['BIconShieldFillExclamation']
-  const BIconShieldFillMinus: typeof import('bootstrap-icons-vue')['BIconShieldFillMinus']
-  const BIconShieldFillPlus: typeof import('bootstrap-icons-vue')['BIconShieldFillPlus']
-  const BIconShieldFillX: typeof import('bootstrap-icons-vue')['BIconShieldFillX']
-  const BIconShieldLock: typeof import('bootstrap-icons-vue')['BIconShieldLock']
-  const BIconShieldLockFill: typeof import('bootstrap-icons-vue')['BIconShieldLockFill']
-  const BIconShieldMinus: typeof import('bootstrap-icons-vue')['BIconShieldMinus']
-  const BIconShieldPlus: typeof import('bootstrap-icons-vue')['BIconShieldPlus']
-  const BIconShieldShaded: typeof import('bootstrap-icons-vue')['BIconShieldShaded']
-  const BIconShieldSlash: typeof import('bootstrap-icons-vue')['BIconShieldSlash']
-  const BIconShieldSlashFill: typeof import('bootstrap-icons-vue')['BIconShieldSlashFill']
-  const BIconShieldX: typeof import('bootstrap-icons-vue')['BIconShieldX']
-  const BIconShift: typeof import('bootstrap-icons-vue')['BIconShift']
-  const BIconShiftFill: typeof import('bootstrap-icons-vue')['BIconShiftFill']
-  const BIconShop: typeof import('bootstrap-icons-vue')['BIconShop']
-  const BIconShopWindow: typeof import('bootstrap-icons-vue')['BIconShopWindow']
-  const BIconShuffle: typeof import('bootstrap-icons-vue')['BIconShuffle']
-  const BIconSignal: typeof import('bootstrap-icons-vue')['BIconSignal']
-  const BIconSignpost2: typeof import('bootstrap-icons-vue')['BIconSignpost2']
-  const BIconSignpost2Fill: typeof import('bootstrap-icons-vue')['BIconSignpost2Fill']
-  const BIconSignpost: typeof import('bootstrap-icons-vue')['BIconSignpost']
-  const BIconSignpostFill: typeof import('bootstrap-icons-vue')['BIconSignpostFill']
-  const BIconSignpostSplit: typeof import('bootstrap-icons-vue')['BIconSignpostSplit']
-  const BIconSignpostSplitFill: typeof import('bootstrap-icons-vue')['BIconSignpostSplitFill']
-  const BIconSim: typeof import('bootstrap-icons-vue')['BIconSim']
-  const BIconSimFill: typeof import('bootstrap-icons-vue')['BIconSimFill']
-  const BIconSkipBackward: typeof import('bootstrap-icons-vue')['BIconSkipBackward']
-  const BIconSkipBackwardBtn: typeof import('bootstrap-icons-vue')['BIconSkipBackwardBtn']
-  const BIconSkipBackwardBtnFill: typeof import('bootstrap-icons-vue')['BIconSkipBackwardBtnFill']
-  const BIconSkipBackwardCircle: typeof import('bootstrap-icons-vue')['BIconSkipBackwardCircle']
-  const BIconSkipBackwardCircleFill: typeof import('bootstrap-icons-vue')['BIconSkipBackwardCircleFill']
-  const BIconSkipBackwardFill: typeof import('bootstrap-icons-vue')['BIconSkipBackwardFill']
-  const BIconSkipEnd: typeof import('bootstrap-icons-vue')['BIconSkipEnd']
-  const BIconSkipEndBtn: typeof import('bootstrap-icons-vue')['BIconSkipEndBtn']
-  const BIconSkipEndBtnFill: typeof import('bootstrap-icons-vue')['BIconSkipEndBtnFill']
-  const BIconSkipEndCircle: typeof import('bootstrap-icons-vue')['BIconSkipEndCircle']
-  const BIconSkipEndCircleFill: typeof import('bootstrap-icons-vue')['BIconSkipEndCircleFill']
-  const BIconSkipEndFill: typeof import('bootstrap-icons-vue')['BIconSkipEndFill']
-  const BIconSkipForward: typeof import('bootstrap-icons-vue')['BIconSkipForward']
-  const BIconSkipForwardBtn: typeof import('bootstrap-icons-vue')['BIconSkipForwardBtn']
-  const BIconSkipForwardBtnFill: typeof import('bootstrap-icons-vue')['BIconSkipForwardBtnFill']
-  const BIconSkipForwardCircle: typeof import('bootstrap-icons-vue')['BIconSkipForwardCircle']
-  const BIconSkipForwardCircleFill: typeof import('bootstrap-icons-vue')['BIconSkipForwardCircleFill']
-  const BIconSkipForwardFill: typeof import('bootstrap-icons-vue')['BIconSkipForwardFill']
-  const BIconSkipStart: typeof import('bootstrap-icons-vue')['BIconSkipStart']
-  const BIconSkipStartBtn: typeof import('bootstrap-icons-vue')['BIconSkipStartBtn']
-  const BIconSkipStartBtnFill: typeof import('bootstrap-icons-vue')['BIconSkipStartBtnFill']
-  const BIconSkipStartCircle: typeof import('bootstrap-icons-vue')['BIconSkipStartCircle']
-  const BIconSkipStartCircleFill: typeof import('bootstrap-icons-vue')['BIconSkipStartCircleFill']
-  const BIconSkipStartFill: typeof import('bootstrap-icons-vue')['BIconSkipStartFill']
-  const BIconSkype: typeof import('bootstrap-icons-vue')['BIconSkype']
-  const BIconSlack: typeof import('bootstrap-icons-vue')['BIconSlack']
-  const BIconSlash: typeof import('bootstrap-icons-vue')['BIconSlash']
-  const BIconSlashCircle: typeof import('bootstrap-icons-vue')['BIconSlashCircle']
-  const BIconSlashCircleFill: typeof import('bootstrap-icons-vue')['BIconSlashCircleFill']
-  const BIconSlashLg: typeof import('bootstrap-icons-vue')['BIconSlashLg']
-  const BIconSlashSquare: typeof import('bootstrap-icons-vue')['BIconSlashSquare']
-  const BIconSlashSquareFill: typeof import('bootstrap-icons-vue')['BIconSlashSquareFill']
-  const BIconSliders2: typeof import('bootstrap-icons-vue')['BIconSliders2']
-  const BIconSliders2Vertical: typeof import('bootstrap-icons-vue')['BIconSliders2Vertical']
-  const BIconSliders: typeof import('bootstrap-icons-vue')['BIconSliders']
-  const BIconSmartwatch: typeof import('bootstrap-icons-vue')['BIconSmartwatch']
-  const BIconSnapchat: typeof import('bootstrap-icons-vue')['BIconSnapchat']
-  const BIconSnow2: typeof import('bootstrap-icons-vue')['BIconSnow2']
-  const BIconSnow3: typeof import('bootstrap-icons-vue')['BIconSnow3']
-  const BIconSnow: typeof import('bootstrap-icons-vue')['BIconSnow']
-  const BIconSortAlphaDown: typeof import('bootstrap-icons-vue')['BIconSortAlphaDown']
-  const BIconSortAlphaDownAlt: typeof import('bootstrap-icons-vue')['BIconSortAlphaDownAlt']
-  const BIconSortAlphaUp: typeof import('bootstrap-icons-vue')['BIconSortAlphaUp']
-  const BIconSortAlphaUpAlt: typeof import('bootstrap-icons-vue')['BIconSortAlphaUpAlt']
-  const BIconSortDown: typeof import('bootstrap-icons-vue')['BIconSortDown']
-  const BIconSortDownAlt: typeof import('bootstrap-icons-vue')['BIconSortDownAlt']
-  const BIconSortNumericDown: typeof import('bootstrap-icons-vue')['BIconSortNumericDown']
-  const BIconSortNumericDownAlt: typeof import('bootstrap-icons-vue')['BIconSortNumericDownAlt']
-  const BIconSortNumericUp: typeof import('bootstrap-icons-vue')['BIconSortNumericUp']
-  const BIconSortNumericUpAlt: typeof import('bootstrap-icons-vue')['BIconSortNumericUpAlt']
-  const BIconSortUp: typeof import('bootstrap-icons-vue')['BIconSortUp']
-  const BIconSortUpAlt: typeof import('bootstrap-icons-vue')['BIconSortUpAlt']
-  const BIconSoundwave: typeof import('bootstrap-icons-vue')['BIconSoundwave']
-  const BIconSpeaker: typeof import('bootstrap-icons-vue')['BIconSpeaker']
-  const BIconSpeakerFill: typeof import('bootstrap-icons-vue')['BIconSpeakerFill']
-  const BIconSpeedometer2: typeof import('bootstrap-icons-vue')['BIconSpeedometer2']
-  const BIconSpeedometer: typeof import('bootstrap-icons-vue')['BIconSpeedometer']
-  const BIconSpellcheck: typeof import('bootstrap-icons-vue')['BIconSpellcheck']
-  const BIconSpotify: typeof import('bootstrap-icons-vue')['BIconSpotify']
-  const BIconSquare: typeof import('bootstrap-icons-vue')['BIconSquare']
-  const BIconSquareFill: typeof import('bootstrap-icons-vue')['BIconSquareFill']
-  const BIconSquareHalf: typeof import('bootstrap-icons-vue')['BIconSquareHalf']
-  const BIconStack: typeof import('bootstrap-icons-vue')['BIconStack']
-  const BIconStackOverflow: typeof import('bootstrap-icons-vue')['BIconStackOverflow']
-  const BIconStar: typeof import('bootstrap-icons-vue')['BIconStar']
-  const BIconStarFill: typeof import('bootstrap-icons-vue')['BIconStarFill']
-  const BIconStarHalf: typeof import('bootstrap-icons-vue')['BIconStarHalf']
-  const BIconStars: typeof import('bootstrap-icons-vue')['BIconStars']
-  const BIconSteam: typeof import('bootstrap-icons-vue')['BIconSteam']
-  const BIconStickies: typeof import('bootstrap-icons-vue')['BIconStickies']
-  const BIconStickiesFill: typeof import('bootstrap-icons-vue')['BIconStickiesFill']
-  const BIconSticky: typeof import('bootstrap-icons-vue')['BIconSticky']
-  const BIconStickyFill: typeof import('bootstrap-icons-vue')['BIconStickyFill']
-  const BIconStop: typeof import('bootstrap-icons-vue')['BIconStop']
-  const BIconStopBtn: typeof import('bootstrap-icons-vue')['BIconStopBtn']
-  const BIconStopBtnFill: typeof import('bootstrap-icons-vue')['BIconStopBtnFill']
-  const BIconStopCircle: typeof import('bootstrap-icons-vue')['BIconStopCircle']
-  const BIconStopCircleFill: typeof import('bootstrap-icons-vue')['BIconStopCircleFill']
-  const BIconStopFill: typeof import('bootstrap-icons-vue')['BIconStopFill']
-  const BIconStoplights: typeof import('bootstrap-icons-vue')['BIconStoplights']
-  const BIconStoplightsFill: typeof import('bootstrap-icons-vue')['BIconStoplightsFill']
-  const BIconStopwatch: typeof import('bootstrap-icons-vue')['BIconStopwatch']
-  const BIconStopwatchFill: typeof import('bootstrap-icons-vue')['BIconStopwatchFill']
-  const BIconStrava: typeof import('bootstrap-icons-vue')['BIconStrava']
-  const BIconSubtract: typeof import('bootstrap-icons-vue')['BIconSubtract']
-  const BIconSuitClub: typeof import('bootstrap-icons-vue')['BIconSuitClub']
-  const BIconSuitClubFill: typeof import('bootstrap-icons-vue')['BIconSuitClubFill']
-  const BIconSuitDiamond: typeof import('bootstrap-icons-vue')['BIconSuitDiamond']
-  const BIconSuitDiamondFill: typeof import('bootstrap-icons-vue')['BIconSuitDiamondFill']
-  const BIconSuitHeart: typeof import('bootstrap-icons-vue')['BIconSuitHeart']
-  const BIconSuitHeartFill: typeof import('bootstrap-icons-vue')['BIconSuitHeartFill']
-  const BIconSuitSpade: typeof import('bootstrap-icons-vue')['BIconSuitSpade']
-  const BIconSuitSpadeFill: typeof import('bootstrap-icons-vue')['BIconSuitSpadeFill']
-  const BIconSun: typeof import('bootstrap-icons-vue')['BIconSun']
-  const BIconSunFill: typeof import('bootstrap-icons-vue')['BIconSunFill']
-  const BIconSunglasses: typeof import('bootstrap-icons-vue')['BIconSunglasses']
-  const BIconSunrise: typeof import('bootstrap-icons-vue')['BIconSunrise']
-  const BIconSunriseFill: typeof import('bootstrap-icons-vue')['BIconSunriseFill']
-  const BIconSunset: typeof import('bootstrap-icons-vue')['BIconSunset']
-  const BIconSunsetFill: typeof import('bootstrap-icons-vue')['BIconSunsetFill']
-  const BIconSymmetryHorizontal: typeof import('bootstrap-icons-vue')['BIconSymmetryHorizontal']
-  const BIconSymmetryVertical: typeof import('bootstrap-icons-vue')['BIconSymmetryVertical']
-  const BIconTable: typeof import('bootstrap-icons-vue')['BIconTable']
-  const BIconTablet: typeof import('bootstrap-icons-vue')['BIconTablet']
-  const BIconTabletFill: typeof import('bootstrap-icons-vue')['BIconTabletFill']
-  const BIconTabletLandscape: typeof import('bootstrap-icons-vue')['BIconTabletLandscape']
-  const BIconTabletLandscapeFill: typeof import('bootstrap-icons-vue')['BIconTabletLandscapeFill']
-  const BIconTag: typeof import('bootstrap-icons-vue')['BIconTag']
-  const BIconTagFill: typeof import('bootstrap-icons-vue')['BIconTagFill']
-  const BIconTags: typeof import('bootstrap-icons-vue')['BIconTags']
-  const BIconTagsFill: typeof import('bootstrap-icons-vue')['BIconTagsFill']
-  const BIconTelegram: typeof import('bootstrap-icons-vue')['BIconTelegram']
-  const BIconTelephone: typeof import('bootstrap-icons-vue')['BIconTelephone']
-  const BIconTelephoneFill: typeof import('bootstrap-icons-vue')['BIconTelephoneFill']
-  const BIconTelephoneForward: typeof import('bootstrap-icons-vue')['BIconTelephoneForward']
-  const BIconTelephoneForwardFill: typeof import('bootstrap-icons-vue')['BIconTelephoneForwardFill']
-  const BIconTelephoneInbound: typeof import('bootstrap-icons-vue')['BIconTelephoneInbound']
-  const BIconTelephoneInboundFill: typeof import('bootstrap-icons-vue')['BIconTelephoneInboundFill']
-  const BIconTelephoneMinus: typeof import('bootstrap-icons-vue')['BIconTelephoneMinus']
-  const BIconTelephoneMinusFill: typeof import('bootstrap-icons-vue')['BIconTelephoneMinusFill']
-  const BIconTelephoneOutbound: typeof import('bootstrap-icons-vue')['BIconTelephoneOutbound']
-  const BIconTelephoneOutboundFill: typeof import('bootstrap-icons-vue')['BIconTelephoneOutboundFill']
-  const BIconTelephonePlus: typeof import('bootstrap-icons-vue')['BIconTelephonePlus']
-  const BIconTelephonePlusFill: typeof import('bootstrap-icons-vue')['BIconTelephonePlusFill']
-  const BIconTelephoneX: typeof import('bootstrap-icons-vue')['BIconTelephoneX']
-  const BIconTelephoneXFill: typeof import('bootstrap-icons-vue')['BIconTelephoneXFill']
-  const BIconTerminal: typeof import('bootstrap-icons-vue')['BIconTerminal']
-  const BIconTerminalDash: typeof import('bootstrap-icons-vue')['BIconTerminalDash']
-  const BIconTerminalFill: typeof import('bootstrap-icons-vue')['BIconTerminalFill']
-  const BIconTerminalPlus: typeof import('bootstrap-icons-vue')['BIconTerminalPlus']
-  const BIconTerminalSplit: typeof import('bootstrap-icons-vue')['BIconTerminalSplit']
-  const BIconTerminalX: typeof import('bootstrap-icons-vue')['BIconTerminalX']
-  const BIconTextCenter: typeof import('bootstrap-icons-vue')['BIconTextCenter']
-  const BIconTextIndentLeft: typeof import('bootstrap-icons-vue')['BIconTextIndentLeft']
-  const BIconTextIndentRight: typeof import('bootstrap-icons-vue')['BIconTextIndentRight']
-  const BIconTextLeft: typeof import('bootstrap-icons-vue')['BIconTextLeft']
-  const BIconTextParagraph: typeof import('bootstrap-icons-vue')['BIconTextParagraph']
-  const BIconTextRight: typeof import('bootstrap-icons-vue')['BIconTextRight']
-  const BIconTextarea: typeof import('bootstrap-icons-vue')['BIconTextarea']
-  const BIconTextareaResize: typeof import('bootstrap-icons-vue')['BIconTextareaResize']
-  const BIconTextareaT: typeof import('bootstrap-icons-vue')['BIconTextareaT']
-  const BIconThermometer: typeof import('bootstrap-icons-vue')['BIconThermometer']
-  const BIconThermometerHalf: typeof import('bootstrap-icons-vue')['BIconThermometerHalf']
-  const BIconThermometerHigh: typeof import('bootstrap-icons-vue')['BIconThermometerHigh']
-  const BIconThermometerLow: typeof import('bootstrap-icons-vue')['BIconThermometerLow']
-  const BIconThermometerSnow: typeof import('bootstrap-icons-vue')['BIconThermometerSnow']
-  const BIconThermometerSun: typeof import('bootstrap-icons-vue')['BIconThermometerSun']
-  const BIconThreeDots: typeof import('bootstrap-icons-vue')['BIconThreeDots']
-  const BIconThreeDotsVertical: typeof import('bootstrap-icons-vue')['BIconThreeDotsVertical']
-  const BIconThunderbolt: typeof import('bootstrap-icons-vue')['BIconThunderbolt']
-  const BIconThunderboltFill: typeof import('bootstrap-icons-vue')['BIconThunderboltFill']
-  const BIconTicket: typeof import('bootstrap-icons-vue')['BIconTicket']
-  const BIconTicketDetailed: typeof import('bootstrap-icons-vue')['BIconTicketDetailed']
-  const BIconTicketDetailedFill: typeof import('bootstrap-icons-vue')['BIconTicketDetailedFill']
-  const BIconTicketFill: typeof import('bootstrap-icons-vue')['BIconTicketFill']
-  const BIconTicketPerforated: typeof import('bootstrap-icons-vue')['BIconTicketPerforated']
-  const BIconTicketPerforatedFill: typeof import('bootstrap-icons-vue')['BIconTicketPerforatedFill']
-  const BIconTiktok: typeof import('bootstrap-icons-vue')['BIconTiktok']
-  const BIconToggle2Off: typeof import('bootstrap-icons-vue')['BIconToggle2Off']
-  const BIconToggle2On: typeof import('bootstrap-icons-vue')['BIconToggle2On']
-  const BIconToggleOff: typeof import('bootstrap-icons-vue')['BIconToggleOff']
-  const BIconToggleOn: typeof import('bootstrap-icons-vue')['BIconToggleOn']
-  const BIconToggles2: typeof import('bootstrap-icons-vue')['BIconToggles2']
-  const BIconToggles: typeof import('bootstrap-icons-vue')['BIconToggles']
-  const BIconTools: typeof import('bootstrap-icons-vue')['BIconTools']
-  const BIconTornado: typeof import('bootstrap-icons-vue')['BIconTornado']
-  const BIconTranslate: typeof import('bootstrap-icons-vue')['BIconTranslate']
-  const BIconTrash2: typeof import('bootstrap-icons-vue')['BIconTrash2']
-  const BIconTrash2Fill: typeof import('bootstrap-icons-vue')['BIconTrash2Fill']
-  const BIconTrash3: typeof import('bootstrap-icons-vue')['BIconTrash3']
-  const BIconTrash3Fill: typeof import('bootstrap-icons-vue')['BIconTrash3Fill']
-  const BIconTrash: typeof import('bootstrap-icons-vue')['BIconTrash']
-  const BIconTrashFill: typeof import('bootstrap-icons-vue')['BIconTrashFill']
-  const BIconTree: typeof import('bootstrap-icons-vue')['BIconTree']
-  const BIconTreeFill: typeof import('bootstrap-icons-vue')['BIconTreeFill']
-  const BIconTriangle: typeof import('bootstrap-icons-vue')['BIconTriangle']
-  const BIconTriangleFill: typeof import('bootstrap-icons-vue')['BIconTriangleFill']
-  const BIconTriangleHalf: typeof import('bootstrap-icons-vue')['BIconTriangleHalf']
-  const BIconTrophy: typeof import('bootstrap-icons-vue')['BIconTrophy']
-  const BIconTrophyFill: typeof import('bootstrap-icons-vue')['BIconTrophyFill']
-  const BIconTropicalStorm: typeof import('bootstrap-icons-vue')['BIconTropicalStorm']
-  const BIconTruck: typeof import('bootstrap-icons-vue')['BIconTruck']
-  const BIconTruckFlatbed: typeof import('bootstrap-icons-vue')['BIconTruckFlatbed']
-  const BIconTsunami: typeof import('bootstrap-icons-vue')['BIconTsunami']
-  const BIconTv: typeof import('bootstrap-icons-vue')['BIconTv']
-  const BIconTvFill: typeof import('bootstrap-icons-vue')['BIconTvFill']
-  const BIconTwitch: typeof import('bootstrap-icons-vue')['BIconTwitch']
-  const BIconTwitter: typeof import('bootstrap-icons-vue')['BIconTwitter']
-  const BIconType: typeof import('bootstrap-icons-vue')['BIconType']
-  const BIconTypeBold: typeof import('bootstrap-icons-vue')['BIconTypeBold']
-  const BIconTypeH1: typeof import('bootstrap-icons-vue')['BIconTypeH1']
-  const BIconTypeH2: typeof import('bootstrap-icons-vue')['BIconTypeH2']
-  const BIconTypeH3: typeof import('bootstrap-icons-vue')['BIconTypeH3']
-  const BIconTypeItalic: typeof import('bootstrap-icons-vue')['BIconTypeItalic']
-  const BIconTypeStrikethrough: typeof import('bootstrap-icons-vue')['BIconTypeStrikethrough']
-  const BIconTypeUnderline: typeof import('bootstrap-icons-vue')['BIconTypeUnderline']
-  const BIconUiChecks: typeof import('bootstrap-icons-vue')['BIconUiChecks']
-  const BIconUiChecksGrid: typeof import('bootstrap-icons-vue')['BIconUiChecksGrid']
-  const BIconUiRadios: typeof import('bootstrap-icons-vue')['BIconUiRadios']
-  const BIconUiRadiosGrid: typeof import('bootstrap-icons-vue')['BIconUiRadiosGrid']
-  const BIconUmbrella: typeof import('bootstrap-icons-vue')['BIconUmbrella']
-  const BIconUmbrellaFill: typeof import('bootstrap-icons-vue')['BIconUmbrellaFill']
-  const BIconUnion: typeof import('bootstrap-icons-vue')['BIconUnion']
-  const BIconUnlock: typeof import('bootstrap-icons-vue')['BIconUnlock']
-  const BIconUnlockFill: typeof import('bootstrap-icons-vue')['BIconUnlockFill']
-  const BIconUpc: typeof import('bootstrap-icons-vue')['BIconUpc']
-  const BIconUpcScan: typeof import('bootstrap-icons-vue')['BIconUpcScan']
-  const BIconUpload: typeof import('bootstrap-icons-vue')['BIconUpload']
-  const BIconUsb: typeof import('bootstrap-icons-vue')['BIconUsb']
-  const BIconUsbC: typeof import('bootstrap-icons-vue')['BIconUsbC']
-  const BIconUsbCFill: typeof import('bootstrap-icons-vue')['BIconUsbCFill']
-  const BIconUsbDrive: typeof import('bootstrap-icons-vue')['BIconUsbDrive']
-  const BIconUsbDriveFill: typeof import('bootstrap-icons-vue')['BIconUsbDriveFill']
-  const BIconUsbFill: typeof import('bootstrap-icons-vue')['BIconUsbFill']
-  const BIconUsbMicro: typeof import('bootstrap-icons-vue')['BIconUsbMicro']
-  const BIconUsbMicroFill: typeof import('bootstrap-icons-vue')['BIconUsbMicroFill']
-  const BIconUsbMini: typeof import('bootstrap-icons-vue')['BIconUsbMini']
-  const BIconUsbMiniFill: typeof import('bootstrap-icons-vue')['BIconUsbMiniFill']
-  const BIconUsbPlug: typeof import('bootstrap-icons-vue')['BIconUsbPlug']
-  const BIconUsbPlugFill: typeof import('bootstrap-icons-vue')['BIconUsbPlugFill']
-  const BIconUsbSymbol: typeof import('bootstrap-icons-vue')['BIconUsbSymbol']
-  const BIconValentine2: typeof import('bootstrap-icons-vue')['BIconValentine2']
-  const BIconValentine: typeof import('bootstrap-icons-vue')['BIconValentine']
-  const BIconVectorPen: typeof import('bootstrap-icons-vue')['BIconVectorPen']
-  const BIconViewList: typeof import('bootstrap-icons-vue')['BIconViewList']
-  const BIconViewStacked: typeof import('bootstrap-icons-vue')['BIconViewStacked']
-  const BIconVimeo: typeof import('bootstrap-icons-vue')['BIconVimeo']
-  const BIconVinyl: typeof import('bootstrap-icons-vue')['BIconVinyl']
-  const BIconVinylFill: typeof import('bootstrap-icons-vue')['BIconVinylFill']
-  const BIconVoicemail: typeof import('bootstrap-icons-vue')['BIconVoicemail']
-  const BIconVolumeDown: typeof import('bootstrap-icons-vue')['BIconVolumeDown']
-  const BIconVolumeDownFill: typeof import('bootstrap-icons-vue')['BIconVolumeDownFill']
-  const BIconVolumeMute: typeof import('bootstrap-icons-vue')['BIconVolumeMute']
-  const BIconVolumeMuteFill: typeof import('bootstrap-icons-vue')['BIconVolumeMuteFill']
-  const BIconVolumeOff: typeof import('bootstrap-icons-vue')['BIconVolumeOff']
-  const BIconVolumeOffFill: typeof import('bootstrap-icons-vue')['BIconVolumeOffFill']
-  const BIconVolumeUp: typeof import('bootstrap-icons-vue')['BIconVolumeUp']
-  const BIconVolumeUpFill: typeof import('bootstrap-icons-vue')['BIconVolumeUpFill']
-  const BIconVr: typeof import('bootstrap-icons-vue')['BIconVr']
-  const BIconWallet2: typeof import('bootstrap-icons-vue')['BIconWallet2']
-  const BIconWallet: typeof import('bootstrap-icons-vue')['BIconWallet']
-  const BIconWalletFill: typeof import('bootstrap-icons-vue')['BIconWalletFill']
-  const BIconWatch: typeof import('bootstrap-icons-vue')['BIconWatch']
-  const BIconWater: typeof import('bootstrap-icons-vue')['BIconWater']
-  const BIconWebcam: typeof import('bootstrap-icons-vue')['BIconWebcam']
-  const BIconWebcamFill: typeof import('bootstrap-icons-vue')['BIconWebcamFill']
-  const BIconWhatsapp: typeof import('bootstrap-icons-vue')['BIconWhatsapp']
-  const BIconWifi1: typeof import('bootstrap-icons-vue')['BIconWifi1']
-  const BIconWifi2: typeof import('bootstrap-icons-vue')['BIconWifi2']
-  const BIconWifi: typeof import('bootstrap-icons-vue')['BIconWifi']
-  const BIconWifiOff: typeof import('bootstrap-icons-vue')['BIconWifiOff']
-  const BIconWind: typeof import('bootstrap-icons-vue')['BIconWind']
-  const BIconWindow: typeof import('bootstrap-icons-vue')['BIconWindow']
-  const BIconWindowDash: typeof import('bootstrap-icons-vue')['BIconWindowDash']
-  const BIconWindowDesktop: typeof import('bootstrap-icons-vue')['BIconWindowDesktop']
-  const BIconWindowDock: typeof import('bootstrap-icons-vue')['BIconWindowDock']
-  const BIconWindowFullscreen: typeof import('bootstrap-icons-vue')['BIconWindowFullscreen']
-  const BIconWindowPlus: typeof import('bootstrap-icons-vue')['BIconWindowPlus']
-  const BIconWindowSidebar: typeof import('bootstrap-icons-vue')['BIconWindowSidebar']
-  const BIconWindowSplit: typeof import('bootstrap-icons-vue')['BIconWindowSplit']
-  const BIconWindowStack: typeof import('bootstrap-icons-vue')['BIconWindowStack']
-  const BIconWindowX: typeof import('bootstrap-icons-vue')['BIconWindowX']
-  const BIconWindows: typeof import('bootstrap-icons-vue')['BIconWindows']
-  const BIconWordpress: typeof import('bootstrap-icons-vue')['BIconWordpress']
-  const BIconWrench: typeof import('bootstrap-icons-vue')['BIconWrench']
-  const BIconWrenchAdjustable: typeof import('bootstrap-icons-vue')['BIconWrenchAdjustable']
-  const BIconWrenchAdjustableCircle: typeof import('bootstrap-icons-vue')['BIconWrenchAdjustableCircle']
-  const BIconWrenchAdjustableCircleFill: typeof import('bootstrap-icons-vue')['BIconWrenchAdjustableCircleFill']
-  const BIconX: typeof import('bootstrap-icons-vue')['BIconX']
-  const BIconXCircle: typeof import('bootstrap-icons-vue')['BIconXCircle']
-  const BIconXCircleFill: typeof import('bootstrap-icons-vue')['BIconXCircleFill']
-  const BIconXDiamond: typeof import('bootstrap-icons-vue')['BIconXDiamond']
-  const BIconXDiamondFill: typeof import('bootstrap-icons-vue')['BIconXDiamondFill']
-  const BIconXLg: typeof import('bootstrap-icons-vue')['BIconXLg']
-  const BIconXOctagon: typeof import('bootstrap-icons-vue')['BIconXOctagon']
-  const BIconXOctagonFill: typeof import('bootstrap-icons-vue')['BIconXOctagonFill']
-  const BIconXSquare: typeof import('bootstrap-icons-vue')['BIconXSquare']
-  const BIconXSquareFill: typeof import('bootstrap-icons-vue')['BIconXSquareFill']
-  const BIconXbox: typeof import('bootstrap-icons-vue')['BIconXbox']
-  const BIconYinYang: typeof import('bootstrap-icons-vue')['BIconYinYang']
-  const BIconYoutube: typeof import('bootstrap-icons-vue')['BIconYoutube']
-  const BIconZoomIn: typeof import('bootstrap-icons-vue')['BIconZoomIn']
-  const BIconZoomOut: typeof import('bootstrap-icons-vue')['BIconZoomOut']
-  const EffectScope: typeof import('vue')['EffectScope']
-  const GQL_URL: typeof import('./shared/constants/url-config')['GQL_URL']
-  const GQL_WS_URL: typeof import('./shared/constants/url-config')['GQL_WS_URL']
-  const RouteName: typeof import('./shared/constants/route-name')['RouteName']
-  const USER: typeof import('./shared/dto/user')['USER']
-  const USER_URL: typeof import('./shared/dto/user-url')['USER_URL']
-  const afterAll: typeof import('vitest')['afterAll']
-  const afterEach: typeof import('vitest')['afterEach']
-  const apolloClient: typeof import('./core/utils/apollo')['apolloClient']
-  const assert: typeof import('vitest')['assert']
-  const asyncComputed: typeof import('@vueuse/core')['asyncComputed']
-  const autoResetRef: typeof import('@vueuse/core')['autoResetRef']
-  const axiosApi: typeof import('./core/utils/axios')['axiosApi']
-  const axiosAuth: typeof import('./core/utils/axios')['axiosAuth']
-  const beforeAll: typeof import('vitest')['beforeAll']
-  const beforeEach: typeof import('vitest')['beforeEach']
-  const chai: typeof import('vitest')['chai']
-  const computed: typeof import('vue')['computed']
-  const computedAsync: typeof import('@vueuse/core')['computedAsync']
-  const computedEager: typeof import('@vueuse/core')['computedEager']
-  const computedInject: typeof import('@vueuse/core')['computedInject']
-  const computedWithControl: typeof import('@vueuse/core')['computedWithControl']
-  const controlledComputed: typeof import('@vueuse/core')['controlledComputed']
-  const controlledRef: typeof import('@vueuse/core')['controlledRef']
-  const createApp: typeof import('vue')['createApp']
-  const createEventHook: typeof import('@vueuse/core')['createEventHook']
-  const createGlobalState: typeof import('@vueuse/core')['createGlobalState']
-  const createInjectionState: typeof import('@vueuse/core')['createInjectionState']
-  const createReactiveFn: typeof import('@vueuse/core')['createReactiveFn']
-  const createSharedComposable: typeof import('@vueuse/core')['createSharedComposable']
-  const createUnrefFn: typeof import('@vueuse/core')['createUnrefFn']
-  const customRef: typeof import('vue')['customRef']
-  const debouncedRef: typeof import('@vueuse/core')['debouncedRef']
-  const debouncedWatch: typeof import('@vueuse/core')['debouncedWatch']
-  const defineAsyncComponent: typeof import('vue')['defineAsyncComponent']
-  const defineComponent: typeof import('vue')['defineComponent']
-  const describe: typeof import('vitest')['describe']
-  const devicesRoute: typeof import('./devices/router/index')['devicesRoute']
-  const eagerComputed: typeof import('@vueuse/core')['eagerComputed']
-  const effectScope: typeof import('vue')['effectScope']
-  const expect: typeof import('vitest')['expect']
-  const extendRef: typeof import('@vueuse/core')['extendRef']
-  const friendsRoute: typeof import('./friends/router/index')['friendsRoute']
-  const getCurrentInstance: typeof import('vue')['getCurrentInstance']
-  const getCurrentScope: typeof import('vue')['getCurrentScope']
-  const gql: typeof import('graphql-tag')['default']
-  const h: typeof import('vue')['h']
-  const homeRoute: typeof import('./home/router/index')['homeRoute']
-  const ignorableWatch: typeof import('@vueuse/core')['ignorableWatch']
-  const inject: typeof import('vue')['inject']
-  const isDefined: typeof import('@vueuse/core')['isDefined']
-  const isProxy: typeof import('vue')['isProxy']
-  const isReactive: typeof import('vue')['isReactive']
-  const isReadonly: typeof import('vue')['isReadonly']
-  const isRef: typeof import('vue')['isRef']
-  const it: typeof import('vitest')['it']
-  const keys: typeof import('./shared/utils/object')['keys']
-  const libraryRoute: typeof import('./library/router/index')['libraryRoute']
-  const makeDestructurable: typeof import('@vueuse/core')['makeDestructurable']
-  const map: typeof import('./shared/utils/object')['map']
-  const markRaw: typeof import('vue')['markRaw']
-  const nextTick: typeof import('vue')['nextTick']
-  const omit: typeof import('./shared/utils/object')['omit']
-  const onActivated: typeof import('vue')['onActivated']
-  const onBeforeMount: typeof import('vue')['onBeforeMount']
-  const onBeforeRouteLeave: typeof import('vue-router')['onBeforeRouteLeave']
-  const onBeforeRouteUpdate: typeof import('vue-router')['onBeforeRouteUpdate']
-  const onBeforeUnmount: typeof import('vue')['onBeforeUnmount']
-  const onBeforeUpdate: typeof import('vue')['onBeforeUpdate']
-  const onClickOutside: typeof import('@vueuse/core')['onClickOutside']
-  const onDeactivated: typeof import('vue')['onDeactivated']
-  const onErrorCaptured: typeof import('vue')['onErrorCaptured']
-  const onKeyStroke: typeof import('@vueuse/core')['onKeyStroke']
-  const onLongPress: typeof import('@vueuse/core')['onLongPress']
-  const onMounted: typeof import('vue')['onMounted']
-  const onRenderTracked: typeof import('vue')['onRenderTracked']
-  const onRenderTriggered: typeof import('vue')['onRenderTriggered']
-  const onScopeDispose: typeof import('vue')['onScopeDispose']
-  const onServerPrefetch: typeof import('vue')['onServerPrefetch']
-  const onStartTyping: typeof import('@vueuse/core')['onStartTyping']
-  const onUnmounted: typeof import('vue')['onUnmounted']
-  const onUpdated: typeof import('vue')['onUpdated']
-  const pausableWatch: typeof import('@vueuse/core')['pausableWatch']
-  const pick: typeof import('./shared/utils/object')['pick']
-  const playlistsRoute: typeof import('./playlists/router/index')['playlistsRoute']
-  const provide: typeof import('vue')['provide']
-  const reactify: typeof import('@vueuse/core')['reactify']
-  const reactifyObject: typeof import('@vueuse/core')['reactifyObject']
-  const reactive: typeof import('vue')['reactive']
-  const reactiveComputed: typeof import('@vueuse/core')['reactiveComputed']
-  const reactiveOmit: typeof import('@vueuse/core')['reactiveOmit']
-  const reactivePick: typeof import('@vueuse/core')['reactivePick']
-  const readonly: typeof import('vue')['readonly']
-  const ref: typeof import('vue')['ref']
-  const refAutoReset: typeof import('@vueuse/core')['refAutoReset']
-  const refDebounced: typeof import('@vueuse/core')['refDebounced']
-  const refDefault: typeof import('@vueuse/core')['refDefault']
-  const refThrottled: typeof import('@vueuse/core')['refThrottled']
-  const refWithControl: typeof import('@vueuse/core')['refWithControl']
-  const resolveComponent: typeof import('vue')['resolveComponent']
-  const resolveDirective: typeof import('vue')['resolveDirective']
-  const resolveRef: typeof import('@vueuse/core')['resolveRef']
-  const resolveUnref: typeof import('@vueuse/core')['resolveUnref']
-  const serializer: typeof import('./shared/utils/serializer')['serializer']
-  const settingsRoute: typeof import('./settings/router/index')['settingsRoute']
-  const shallowReactive: typeof import('vue')['shallowReactive']
-  const shallowReadonly: typeof import('vue')['shallowReadonly']
-  const shallowRef: typeof import('vue')['shallowRef']
-  const suite: typeof import('vitest')['suite']
-  const switchAssign: typeof import('./shared/utils/switch')['switchAssign']
-  const switchExec: typeof import('./shared/utils/switch')['switchExec']
-  const switchThrow: typeof import('./shared/utils/switch')['switchThrow']
-  const syncRef: typeof import('@vueuse/core')['syncRef']
-  const syncRefs: typeof import('@vueuse/core')['syncRefs']
-  const templateRef: typeof import('@vueuse/core')['templateRef']
-  const test: typeof import('vitest')['test']
-  const themesRoute: typeof import('./themes/router/index')['themesRoute']
-  const throttledRef: typeof import('@vueuse/core')['throttledRef']
-  const throttledWatch: typeof import('@vueuse/core')['throttledWatch']
-  const toRaw: typeof import('vue')['toRaw']
-  const toReactive: typeof import('@vueuse/core')['toReactive']
-  const toRef: typeof import('vue')['toRef']
-  const toRefs: typeof import('vue')['toRefs']
-  const triggerRef: typeof import('vue')['triggerRef']
-  const tryOnBeforeMount: typeof import('@vueuse/core')['tryOnBeforeMount']
-  const tryOnBeforeUnmount: typeof import('@vueuse/core')['tryOnBeforeUnmount']
-  const tryOnMounted: typeof import('@vueuse/core')['tryOnMounted']
-  const tryOnScopeDispose: typeof import('@vueuse/core')['tryOnScopeDispose']
-  const tryOnUnmounted: typeof import('@vueuse/core')['tryOnUnmounted']
-  const unref: typeof import('vue')['unref']
-  const unrefElement: typeof import('@vueuse/core')['unrefElement']
-  const until: typeof import('@vueuse/core')['until']
-  const useActiveElement: typeof import('@vueuse/core')['useActiveElement']
-  const useArrayEvery: typeof import('@vueuse/core')['useArrayEvery']
-  const useArrayFilter: typeof import('@vueuse/core')['useArrayFilter']
-  const useArrayFind: typeof import('@vueuse/core')['useArrayFind']
-  const useArrayFindIndex: typeof import('@vueuse/core')['useArrayFindIndex']
-  const useArrayJoin: typeof import('@vueuse/core')['useArrayJoin']
-  const useArrayMap: typeof import('@vueuse/core')['useArrayMap']
-  const useArrayReduce: typeof import('@vueuse/core')['useArrayReduce']
-  const useArraySome: typeof import('@vueuse/core')['useArraySome']
-  const useArrayUnique: typeof import('@vueuse/core')['useArrayUnique']
-  const useAsyncQueue: typeof import('@vueuse/core')['useAsyncQueue']
-  const useAsyncState: typeof import('@vueuse/core')['useAsyncState']
-  const useAttrs: typeof import('vue')['useAttrs']
-  const useAuthentication: typeof import('./auth/stores/authentication')['useAuthentication']
-  const useAuthenticationService: typeof import('./auth/services/authentication')['useAuthenticationService']
-  const useBase64: typeof import('@vueuse/core')['useBase64']
-  const useBattery: typeof import('@vueuse/core')['useBattery']
-  const useBluetooth: typeof import('@vueuse/core')['useBluetooth']
-  const useBreakpoints: typeof import('@vueuse/core')['useBreakpoints']
-  const useBroadcastChannel: typeof import('@vueuse/core')['useBroadcastChannel']
-  const useBrowserLocation: typeof import('@vueuse/core')['useBrowserLocation']
-  const useCached: typeof import('@vueuse/core')['useCached']
-  const useClipboard: typeof import('@vueuse/core')['useClipboard']
-  const useCloned: typeof import('@vueuse/core')['useCloned']
-  const useColorMode: typeof import('@vueuse/core')['useColorMode']
-  const useConfirmDialog: typeof import('@vueuse/core')['useConfirmDialog']
-  const useCounter: typeof import('@vueuse/core')['useCounter']
-  const useCssModule: typeof import('vue')['useCssModule']
-  const useCssVar: typeof import('@vueuse/core')['useCssVar']
-  const useCssVars: typeof import('vue')['useCssVars']
-  const useCurrentElement: typeof import('@vueuse/core')['useCurrentElement']
-  const useCycleList: typeof import('@vueuse/core')['useCycleList']
-  const useDark: typeof import('@vueuse/core')['useDark']
-  const useDateFormat: typeof import('@vueuse/core')['useDateFormat']
-  const useDebounce: typeof import('@vueuse/core')['useDebounce']
-  const useDebounceFn: typeof import('@vueuse/core')['useDebounceFn']
-  const useDebouncedRefHistory: typeof import('@vueuse/core')['useDebouncedRefHistory']
-  const useDeviceMotion: typeof import('@vueuse/core')['useDeviceMotion']
-  const useDeviceOrientation: typeof import('@vueuse/core')['useDeviceOrientation']
-  const useDevicePixelRatio: typeof import('@vueuse/core')['useDevicePixelRatio']
-  const useDevicesList: typeof import('@vueuse/core')['useDevicesList']
-  const useDisplayMedia: typeof import('@vueuse/core')['useDisplayMedia']
-  const useDocumentVisibility: typeof import('@vueuse/core')['useDocumentVisibility']
-  const useDraggable: typeof import('@vueuse/core')['useDraggable']
-  const useDropZone: typeof import('@vueuse/core')['useDropZone']
-  const useElementBounding: typeof import('@vueuse/core')['useElementBounding']
-  const useElementByPoint: typeof import('@vueuse/core')['useElementByPoint']
-  const useElementHover: typeof import('@vueuse/core')['useElementHover']
-  const useElementSize: typeof import('@vueuse/core')['useElementSize']
-  const useElementVisibility: typeof import('@vueuse/core')['useElementVisibility']
-  const useEventBus: typeof import('@vueuse/core')['useEventBus']
-  const useEventListener: typeof import('@vueuse/core')['useEventListener']
-  const useEventSource: typeof import('@vueuse/core')['useEventSource']
-  const useEyeDropper: typeof import('@vueuse/core')['useEyeDropper']
-  const useFavicon: typeof import('@vueuse/core')['useFavicon']
-  const useFetch: typeof import('@vueuse/core')['useFetch']
-  const useFileDialog: typeof import('@vueuse/core')['useFileDialog']
-  const useFileSystemAccess: typeof import('@vueuse/core')['useFileSystemAccess']
-  const useFocus: typeof import('@vueuse/core')['useFocus']
-  const useFocusWithin: typeof import('@vueuse/core')['useFocusWithin']
-  const useFps: typeof import('@vueuse/core')['useFps']
-  const useFullscreen: typeof import('@vueuse/core')['useFullscreen']
-  const useGamepad: typeof import('@vueuse/core')['useGamepad']
-  const useGeolocation: typeof import('@vueuse/core')['useGeolocation']
-  const useIdle: typeof import('@vueuse/core')['useIdle']
-  const useImage: typeof import('@vueuse/core')['useImage']
-  const useInfiniteScroll: typeof import('@vueuse/core')['useInfiniteScroll']
-  const useIntersectionObserver: typeof import('@vueuse/core')['useIntersectionObserver']
-  const useInterval: typeof import('@vueuse/core')['useInterval']
-  const useIntervalFn: typeof import('@vueuse/core')['useIntervalFn']
-  const useKeyModifier: typeof import('@vueuse/core')['useKeyModifier']
-  const useLastChanged: typeof import('@vueuse/core')['useLastChanged']
-  const useLink: typeof import('vue-router')['useLink']
-  const useLocalStorage: typeof import('@vueuse/core')['useLocalStorage']
-  const useMagicKeys: typeof import('@vueuse/core')['useMagicKeys']
-  const useManualRefHistory: typeof import('@vueuse/core')['useManualRefHistory']
-  const useMediaControls: typeof import('@vueuse/core')['useMediaControls']
-  const useMediaQuery: typeof import('@vueuse/core')['useMediaQuery']
-  const useMemoize: typeof import('@vueuse/core')['useMemoize']
-  const useMemory: typeof import('@vueuse/core')['useMemory']
-  const useMobile: typeof import('./shared/hooks/mobile')['useMobile']
-  const useMounted: typeof import('@vueuse/core')['useMounted']
-  const useMouse: typeof import('@vueuse/core')['useMouse']
-  const useMouseInElement: typeof import('@vueuse/core')['useMouseInElement']
-  const useMousePressed: typeof import('@vueuse/core')['useMousePressed']
-  const useMutation: typeof import('@vue/apollo-composable')['useMutation']
-  const useMutationObserver: typeof import('@vueuse/core')['useMutationObserver']
-  const useNavigatorLanguage: typeof import('@vueuse/core')['useNavigatorLanguage']
-  const useNetwork: typeof import('@vueuse/core')['useNetwork']
-  const useNow: typeof import('@vueuse/core')['useNow']
-  const useObjectUrl: typeof import('@vueuse/core')['useObjectUrl']
-  const useOffsetPagination: typeof import('@vueuse/core')['useOffsetPagination']
-  const useOnline: typeof import('@vueuse/core')['useOnline']
-  const usePageLeave: typeof import('@vueuse/core')['usePageLeave']
-  const useParallax: typeof import('@vueuse/core')['useParallax']
-  const usePermission: typeof import('@vueuse/core')['usePermission']
-  const usePointer: typeof import('@vueuse/core')['usePointer']
-  const usePointerSwipe: typeof import('@vueuse/core')['usePointerSwipe']
-  const usePreferredColorScheme: typeof import('@vueuse/core')['usePreferredColorScheme']
-  const usePreferredContrast: typeof import('@vueuse/core')['usePreferredContrast']
-  const usePreferredDark: typeof import('@vueuse/core')['usePreferredDark']
-  const usePreferredLanguages: typeof import('@vueuse/core')['usePreferredLanguages']
-  const usePreferredReducedMotion: typeof import('@vueuse/core')['usePreferredReducedMotion']
-  const useQuery: typeof import('@vue/apollo-composable')['useQuery']
-  const useRafFn: typeof import('@vueuse/core')['useRafFn']
-  const useRefHistory: typeof import('@vueuse/core')['useRefHistory']
-  const useResizeObserver: typeof import('@vueuse/core')['useResizeObserver']
-  const useRoute: typeof import('vue-router')['useRoute']
-  const useRouter: typeof import('vue-router')['useRouter']
-  const useScreenOrientation: typeof import('@vueuse/core')['useScreenOrientation']
-  const useScreenSafeArea: typeof import('@vueuse/core')['useScreenSafeArea']
-  const useScriptTag: typeof import('@vueuse/core')['useScriptTag']
-  const useScroll: typeof import('@vueuse/core')['useScroll']
-  const useScrollLock: typeof import('@vueuse/core')['useScrollLock']
-  const useSessionStorage: typeof import('@vueuse/core')['useSessionStorage']
-  const useShare: typeof import('@vueuse/core')['useShare']
-  const useSlots: typeof import('vue')['useSlots']
-  const useSorted: typeof import('@vueuse/core')['useSorted']
-  const useSpeechRecognition: typeof import('@vueuse/core')['useSpeechRecognition']
-  const useSpeechSynthesis: typeof import('@vueuse/core')['useSpeechSynthesis']
-  const useStepper: typeof import('@vueuse/core')['useStepper']
-  const useStorage: typeof import('@vueuse/core')['useStorage']
-  const useStorageAsync: typeof import('@vueuse/core')['useStorageAsync']
-  const useStyleTag: typeof import('@vueuse/core')['useStyleTag']
-  const useSubscription: typeof import('@vue/apollo-composable')['useSubscription']
-  const useSupported: typeof import('@vueuse/core')['useSupported']
-  const useSwipe: typeof import('@vueuse/core')['useSwipe']
-  const useTemplateRefsList: typeof import('@vueuse/core')['useTemplateRefsList']
-  const useTextDirection: typeof import('@vueuse/core')['useTextDirection']
-  const useTextSelection: typeof import('@vueuse/core')['useTextSelection']
-  const useTextareaAutosize: typeof import('@vueuse/core')['useTextareaAutosize']
-  const useThrottle: typeof import('@vueuse/core')['useThrottle']
-  const useThrottleFn: typeof import('@vueuse/core')['useThrottleFn']
-  const useThrottledRefHistory: typeof import('@vueuse/core')['useThrottledRefHistory']
-  const useTimeAgo: typeof import('@vueuse/core')['useTimeAgo']
-  const useTimeout: typeof import('@vueuse/core')['useTimeout']
-  const useTimeoutFn: typeof import('@vueuse/core')['useTimeoutFn']
-  const useTimeoutPoll: typeof import('@vueuse/core')['useTimeoutPoll']
-  const useTimestamp: typeof import('@vueuse/core')['useTimestamp']
-  const useTitle: typeof import('@vueuse/core')['useTitle']
-  const useToNumber: typeof import('@vueuse/core')['useToNumber']
-  const useToString: typeof import('@vueuse/core')['useToString']
-  const useToggle: typeof import('@vueuse/core')['useToggle']
-  const useTransition: typeof import('@vueuse/core')['useTransition']
-  const useUrlSearchParams: typeof import('@vueuse/core')['useUrlSearchParams']
-  const useUser: typeof import('./core/stores/user')['useUser']
-  const useUserMedia: typeof import('@vueuse/core')['useUserMedia']
-  const useUsersService: typeof import('./shared/services/users')['useUsersService']
-  const useVModel: typeof import('@vueuse/core')['useVModel']
-  const useVModels: typeof import('@vueuse/core')['useVModels']
-  const useVibrate: typeof import('@vueuse/core')['useVibrate']
-  const useVirtualList: typeof import('@vueuse/core')['useVirtualList']
-  const useWakeLock: typeof import('@vueuse/core')['useWakeLock']
-  const useWebNotification: typeof import('@vueuse/core')['useWebNotification']
-  const useWebSocket: typeof import('@vueuse/core')['useWebSocket']
-  const useWebWorker: typeof import('@vueuse/core')['useWebWorker']
-  const useWebWorkerFn: typeof import('@vueuse/core')['useWebWorkerFn']
-  const useWindowFocus: typeof import('@vueuse/core')['useWindowFocus']
-  const useWindowScroll: typeof import('@vueuse/core')['useWindowScroll']
-  const useWindowSize: typeof import('@vueuse/core')['useWindowSize']
-  const vi: typeof import('vitest')['vi']
-  const vitest: typeof import('vitest')['vitest']
-  const watch: typeof import('vue')['watch']
-  const watchArray: typeof import('@vueuse/core')['watchArray']
-  const watchAtMost: typeof import('@vueuse/core')['watchAtMost']
-  const watchDebounced: typeof import('@vueuse/core')['watchDebounced']
-  const watchEffect: typeof import('vue')['watchEffect']
-  const watchIgnorable: typeof import('@vueuse/core')['watchIgnorable']
-  const watchOnce: typeof import('@vueuse/core')['watchOnce']
-  const watchPausable: typeof import('@vueuse/core')['watchPausable']
-  const watchPostEffect: typeof import('vue')['watchPostEffect']
-  const watchSyncEffect: typeof import('vue')['watchSyncEffect']
-  const watchThrottled: typeof import('@vueuse/core')['watchThrottled']
-  const watchTriggerable: typeof import('@vueuse/core')['watchTriggerable']
-  const watchWithFilter: typeof import('@vueuse/core')['watchWithFilter']
-  const whenever: typeof import('@vueuse/core')['whenever']
-}
-// for vue template auto import
-import { UnwrapRef } from 'vue'
-declare module 'vue' {
-  interface ComponentCustomProperties {
-    readonly API_URL: UnwrapRef<typeof import('./shared/constants/url-config')['API_URL']>
-    readonly API_WS_URL: UnwrapRef<typeof import('./shared/constants/url-config')['API_WS_URL']>
-    readonly AUTH_URL: UnwrapRef<typeof import('./shared/constants/url-config')['AUTH_URL']>
-    readonly BASE_URL: UnwrapRef<typeof import('./shared/constants/url-config')['BASE_URL']>
-    readonly BIcon123: UnwrapRef<typeof import('bootstrap-icons-vue')['BIcon123']>
-    readonly BIconActivity: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconActivity']>
-    readonly BIconAlarm: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlarm']>
-    readonly BIconAlarmFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlarmFill']>
-    readonly BIconAlignBottom: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlignBottom']>
-    readonly BIconAlignCenter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlignCenter']>
-    readonly BIconAlignEnd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlignEnd']>
-    readonly BIconAlignMiddle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlignMiddle']>
-    readonly BIconAlignStart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlignStart']>
-    readonly BIconAlignTop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlignTop']>
-    readonly BIconAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAlt']>
-    readonly BIconApp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconApp']>
-    readonly BIconAppIndicator: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAppIndicator']>
-    readonly BIconApple: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconApple']>
-    readonly BIconArchive: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArchive']>
-    readonly BIconArchiveFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArchiveFill']>
-    readonly BIconArrow90degDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrow90degDown']>
-    readonly BIconArrow90degLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrow90degLeft']>
-    readonly BIconArrow90degRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrow90degRight']>
-    readonly BIconArrow90degUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrow90degUp']>
-    readonly BIconArrowBarDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowBarDown']>
-    readonly BIconArrowBarLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowBarLeft']>
-    readonly BIconArrowBarRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowBarRight']>
-    readonly BIconArrowBarUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowBarUp']>
-    readonly BIconArrowClockwise: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowClockwise']>
-    readonly BIconArrowCounterclockwise: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowCounterclockwise']>
-    readonly BIconArrowDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDown']>
-    readonly BIconArrowDownCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownCircle']>
-    readonly BIconArrowDownCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownCircleFill']>
-    readonly BIconArrowDownLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownLeft']>
-    readonly BIconArrowDownLeftCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownLeftCircle']>
-    readonly BIconArrowDownLeftCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownLeftCircleFill']>
-    readonly BIconArrowDownLeftSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownLeftSquare']>
-    readonly BIconArrowDownLeftSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownLeftSquareFill']>
-    readonly BIconArrowDownRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownRight']>
-    readonly BIconArrowDownRightCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownRightCircle']>
-    readonly BIconArrowDownRightCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownRightCircleFill']>
-    readonly BIconArrowDownRightSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownRightSquare']>
-    readonly BIconArrowDownRightSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownRightSquareFill']>
-    readonly BIconArrowDownShort: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownShort']>
-    readonly BIconArrowDownSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownSquare']>
-    readonly BIconArrowDownSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownSquareFill']>
-    readonly BIconArrowDownUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowDownUp']>
-    readonly BIconArrowLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeft']>
-    readonly BIconArrowLeftCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeftCircle']>
-    readonly BIconArrowLeftCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeftCircleFill']>
-    readonly BIconArrowLeftRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeftRight']>
-    readonly BIconArrowLeftShort: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeftShort']>
-    readonly BIconArrowLeftSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeftSquare']>
-    readonly BIconArrowLeftSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowLeftSquareFill']>
-    readonly BIconArrowRepeat: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRepeat']>
-    readonly BIconArrowReturnLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowReturnLeft']>
-    readonly BIconArrowReturnRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowReturnRight']>
-    readonly BIconArrowRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRight']>
-    readonly BIconArrowRightCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRightCircle']>
-    readonly BIconArrowRightCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRightCircleFill']>
-    readonly BIconArrowRightShort: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRightShort']>
-    readonly BIconArrowRightSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRightSquare']>
-    readonly BIconArrowRightSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowRightSquareFill']>
-    readonly BIconArrowThroughHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowThroughHeart']>
-    readonly BIconArrowThroughHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowThroughHeartFill']>
-    readonly BIconArrowUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUp']>
-    readonly BIconArrowUpCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpCircle']>
-    readonly BIconArrowUpCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpCircleFill']>
-    readonly BIconArrowUpLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpLeft']>
-    readonly BIconArrowUpLeftCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpLeftCircle']>
-    readonly BIconArrowUpLeftCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpLeftCircleFill']>
-    readonly BIconArrowUpLeftSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpLeftSquare']>
-    readonly BIconArrowUpLeftSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpLeftSquareFill']>
-    readonly BIconArrowUpRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpRight']>
-    readonly BIconArrowUpRightCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpRightCircle']>
-    readonly BIconArrowUpRightCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpRightCircleFill']>
-    readonly BIconArrowUpRightSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpRightSquare']>
-    readonly BIconArrowUpRightSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpRightSquareFill']>
-    readonly BIconArrowUpShort: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpShort']>
-    readonly BIconArrowUpSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpSquare']>
-    readonly BIconArrowUpSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowUpSquareFill']>
-    readonly BIconArrowsAngleContract: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowsAngleContract']>
-    readonly BIconArrowsAngleExpand: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowsAngleExpand']>
-    readonly BIconArrowsCollapse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowsCollapse']>
-    readonly BIconArrowsExpand: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowsExpand']>
-    readonly BIconArrowsFullscreen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowsFullscreen']>
-    readonly BIconArrowsMove: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconArrowsMove']>
-    readonly BIconAspectRatio: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAspectRatio']>
-    readonly BIconAspectRatioFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAspectRatioFill']>
-    readonly BIconAsterisk: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAsterisk']>
-    readonly BIconAt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAt']>
-    readonly BIconAward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAward']>
-    readonly BIconAwardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconAwardFill']>
-    readonly BIconBack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBack']>
-    readonly BIconBackspace: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBackspace']>
-    readonly BIconBackspaceFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBackspaceFill']>
-    readonly BIconBackspaceReverse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBackspaceReverse']>
-    readonly BIconBackspaceReverseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBackspaceReverseFill']>
-    readonly BIconBadge3d: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadge3d']>
-    readonly BIconBadge3dFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadge3dFill']>
-    readonly BIconBadge4k: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadge4k']>
-    readonly BIconBadge4kFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadge4kFill']>
-    readonly BIconBadge8k: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadge8k']>
-    readonly BIconBadge8kFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadge8kFill']>
-    readonly BIconBadgeAd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeAd']>
-    readonly BIconBadgeAdFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeAdFill']>
-    readonly BIconBadgeAr: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeAr']>
-    readonly BIconBadgeArFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeArFill']>
-    readonly BIconBadgeCc: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeCc']>
-    readonly BIconBadgeCcFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeCcFill']>
-    readonly BIconBadgeHd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeHd']>
-    readonly BIconBadgeHdFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeHdFill']>
-    readonly BIconBadgeSd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeSd']>
-    readonly BIconBadgeSdFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeSdFill']>
-    readonly BIconBadgeTm: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeTm']>
-    readonly BIconBadgeTmFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeTmFill']>
-    readonly BIconBadgeVo: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeVo']>
-    readonly BIconBadgeVoFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeVoFill']>
-    readonly BIconBadgeVr: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeVr']>
-    readonly BIconBadgeVrFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeVrFill']>
-    readonly BIconBadgeWc: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeWc']>
-    readonly BIconBadgeWcFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBadgeWcFill']>
-    readonly BIconBag: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBag']>
-    readonly BIconBagCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagCheck']>
-    readonly BIconBagCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagCheckFill']>
-    readonly BIconBagDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagDash']>
-    readonly BIconBagDashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagDashFill']>
-    readonly BIconBagFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagFill']>
-    readonly BIconBagHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagHeart']>
-    readonly BIconBagHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagHeartFill']>
-    readonly BIconBagPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagPlus']>
-    readonly BIconBagPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagPlusFill']>
-    readonly BIconBagX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagX']>
-    readonly BIconBagXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBagXFill']>
-    readonly BIconBalloon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBalloon']>
-    readonly BIconBalloonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBalloonFill']>
-    readonly BIconBalloonHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBalloonHeart']>
-    readonly BIconBalloonHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBalloonHeartFill']>
-    readonly BIconBandaid: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBandaid']>
-    readonly BIconBandaidFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBandaidFill']>
-    readonly BIconBank2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBank2']>
-    readonly BIconBank: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBank']>
-    readonly BIconBarChart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBarChart']>
-    readonly BIconBarChartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBarChartFill']>
-    readonly BIconBarChartLine: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBarChartLine']>
-    readonly BIconBarChartLineFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBarChartLineFill']>
-    readonly BIconBarChartSteps: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBarChartSteps']>
-    readonly BIconBasket2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBasket2']>
-    readonly BIconBasket2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBasket2Fill']>
-    readonly BIconBasket3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBasket3']>
-    readonly BIconBasket3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBasket3Fill']>
-    readonly BIconBasket: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBasket']>
-    readonly BIconBasketFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBasketFill']>
-    readonly BIconBattery: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBattery']>
-    readonly BIconBatteryCharging: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBatteryCharging']>
-    readonly BIconBatteryFull: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBatteryFull']>
-    readonly BIconBatteryHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBatteryHalf']>
-    readonly BIconBehance: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBehance']>
-    readonly BIconBell: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBell']>
-    readonly BIconBellFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBellFill']>
-    readonly BIconBellSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBellSlash']>
-    readonly BIconBellSlashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBellSlashFill']>
-    readonly BIconBezier2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBezier2']>
-    readonly BIconBezier: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBezier']>
-    readonly BIconBicycle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBicycle']>
-    readonly BIconBinoculars: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBinoculars']>
-    readonly BIconBinocularsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBinocularsFill']>
-    readonly BIconBlockquoteLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBlockquoteLeft']>
-    readonly BIconBlockquoteRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBlockquoteRight']>
-    readonly BIconBluetooth: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBluetooth']>
-    readonly BIconBodyText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBodyText']>
-    readonly BIconBook: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBook']>
-    readonly BIconBookFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookFill']>
-    readonly BIconBookHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookHalf']>
-    readonly BIconBookmark: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmark']>
-    readonly BIconBookmarkCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkCheck']>
-    readonly BIconBookmarkCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkCheckFill']>
-    readonly BIconBookmarkDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkDash']>
-    readonly BIconBookmarkDashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkDashFill']>
-    readonly BIconBookmarkFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkFill']>
-    readonly BIconBookmarkHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkHeart']>
-    readonly BIconBookmarkHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkHeartFill']>
-    readonly BIconBookmarkPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkPlus']>
-    readonly BIconBookmarkPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkPlusFill']>
-    readonly BIconBookmarkStar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkStar']>
-    readonly BIconBookmarkStarFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkStarFill']>
-    readonly BIconBookmarkX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkX']>
-    readonly BIconBookmarkXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarkXFill']>
-    readonly BIconBookmarks: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarks']>
-    readonly BIconBookmarksFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookmarksFill']>
-    readonly BIconBookshelf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBookshelf']>
-    readonly BIconBoombox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoombox']>
-    readonly BIconBoomboxFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoomboxFill']>
-    readonly BIconBootstrap: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBootstrap']>
-    readonly BIconBootstrapFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBootstrapFill']>
-    readonly BIconBootstrapReboot: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBootstrapReboot']>
-    readonly BIconBorder: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorder']>
-    readonly BIconBorderAll: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderAll']>
-    readonly BIconBorderBottom: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderBottom']>
-    readonly BIconBorderCenter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderCenter']>
-    readonly BIconBorderInner: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderInner']>
-    readonly BIconBorderLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderLeft']>
-    readonly BIconBorderMiddle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderMiddle']>
-    readonly BIconBorderOuter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderOuter']>
-    readonly BIconBorderRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderRight']>
-    readonly BIconBorderStyle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderStyle']>
-    readonly BIconBorderTop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderTop']>
-    readonly BIconBorderWidth: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBorderWidth']>
-    readonly BIconBoundingBox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoundingBox']>
-    readonly BIconBoundingBoxCircles: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoundingBoxCircles']>
-    readonly BIconBox2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBox2']>
-    readonly BIconBox2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBox2Fill']>
-    readonly BIconBox2Heart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBox2Heart']>
-    readonly BIconBox2HeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBox2HeartFill']>
-    readonly BIconBox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBox']>
-    readonly BIconBoxArrowDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowDown']>
-    readonly BIconBoxArrowDownLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowDownLeft']>
-    readonly BIconBoxArrowDownRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowDownRight']>
-    readonly BIconBoxArrowInDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInDown']>
-    readonly BIconBoxArrowInDownLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInDownLeft']>
-    readonly BIconBoxArrowInDownRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInDownRight']>
-    readonly BIconBoxArrowInLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInLeft']>
-    readonly BIconBoxArrowInRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInRight']>
-    readonly BIconBoxArrowInUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInUp']>
-    readonly BIconBoxArrowInUpLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInUpLeft']>
-    readonly BIconBoxArrowInUpRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowInUpRight']>
-    readonly BIconBoxArrowLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowLeft']>
-    readonly BIconBoxArrowRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowRight']>
-    readonly BIconBoxArrowUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowUp']>
-    readonly BIconBoxArrowUpLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowUpLeft']>
-    readonly BIconBoxArrowUpRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxArrowUpRight']>
-    readonly BIconBoxSeam: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxSeam']>
-    readonly BIconBoxes: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBoxes']>
-    readonly BIconBraces: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBraces']>
-    readonly BIconBracesAsterisk: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBracesAsterisk']>
-    readonly BIconBricks: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBricks']>
-    readonly BIconBriefcase: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBriefcase']>
-    readonly BIconBriefcaseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBriefcaseFill']>
-    readonly BIconBrightnessAltHigh: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessAltHigh']>
-    readonly BIconBrightnessAltHighFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessAltHighFill']>
-    readonly BIconBrightnessAltLow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessAltLow']>
-    readonly BIconBrightnessAltLowFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessAltLowFill']>
-    readonly BIconBrightnessHigh: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessHigh']>
-    readonly BIconBrightnessHighFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessHighFill']>
-    readonly BIconBrightnessLow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessLow']>
-    readonly BIconBrightnessLowFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrightnessLowFill']>
-    readonly BIconBroadcast: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBroadcast']>
-    readonly BIconBroadcastPin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBroadcastPin']>
-    readonly BIconBrush: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrush']>
-    readonly BIconBrushFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBrushFill']>
-    readonly BIconBucket: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBucket']>
-    readonly BIconBucketFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBucketFill']>
-    readonly BIconBug: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBug']>
-    readonly BIconBugFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBugFill']>
-    readonly BIconBuilding: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBuilding']>
-    readonly BIconBullseye: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconBullseye']>
-    readonly BIconCalculator: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalculator']>
-    readonly BIconCalculatorFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalculatorFill']>
-    readonly BIconCalendar2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2']>
-    readonly BIconCalendar2Check: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Check']>
-    readonly BIconCalendar2CheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2CheckFill']>
-    readonly BIconCalendar2Date: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Date']>
-    readonly BIconCalendar2DateFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2DateFill']>
-    readonly BIconCalendar2Day: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Day']>
-    readonly BIconCalendar2DayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2DayFill']>
-    readonly BIconCalendar2Event: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Event']>
-    readonly BIconCalendar2EventFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2EventFill']>
-    readonly BIconCalendar2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Fill']>
-    readonly BIconCalendar2Heart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Heart']>
-    readonly BIconCalendar2HeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2HeartFill']>
-    readonly BIconCalendar2Minus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Minus']>
-    readonly BIconCalendar2MinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2MinusFill']>
-    readonly BIconCalendar2Month: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Month']>
-    readonly BIconCalendar2MonthFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2MonthFill']>
-    readonly BIconCalendar2Plus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Plus']>
-    readonly BIconCalendar2PlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2PlusFill']>
-    readonly BIconCalendar2Range: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Range']>
-    readonly BIconCalendar2RangeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2RangeFill']>
-    readonly BIconCalendar2Week: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2Week']>
-    readonly BIconCalendar2WeekFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2WeekFill']>
-    readonly BIconCalendar2X: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2X']>
-    readonly BIconCalendar2XFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar2XFill']>
-    readonly BIconCalendar3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3']>
-    readonly BIconCalendar3Event: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3Event']>
-    readonly BIconCalendar3EventFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3EventFill']>
-    readonly BIconCalendar3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3Fill']>
-    readonly BIconCalendar3Range: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3Range']>
-    readonly BIconCalendar3RangeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3RangeFill']>
-    readonly BIconCalendar3Week: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3Week']>
-    readonly BIconCalendar3WeekFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar3WeekFill']>
-    readonly BIconCalendar4: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar4']>
-    readonly BIconCalendar4Event: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar4Event']>
-    readonly BIconCalendar4Range: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar4Range']>
-    readonly BIconCalendar4Week: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar4Week']>
-    readonly BIconCalendar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendar']>
-    readonly BIconCalendarCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarCheck']>
-    readonly BIconCalendarCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarCheckFill']>
-    readonly BIconCalendarDate: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarDate']>
-    readonly BIconCalendarDateFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarDateFill']>
-    readonly BIconCalendarDay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarDay']>
-    readonly BIconCalendarDayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarDayFill']>
-    readonly BIconCalendarEvent: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarEvent']>
-    readonly BIconCalendarEventFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarEventFill']>
-    readonly BIconCalendarFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarFill']>
-    readonly BIconCalendarHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarHeart']>
-    readonly BIconCalendarHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarHeartFill']>
-    readonly BIconCalendarMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarMinus']>
-    readonly BIconCalendarMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarMinusFill']>
-    readonly BIconCalendarMonth: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarMonth']>
-    readonly BIconCalendarMonthFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarMonthFill']>
-    readonly BIconCalendarPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarPlus']>
-    readonly BIconCalendarPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarPlusFill']>
-    readonly BIconCalendarRange: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarRange']>
-    readonly BIconCalendarRangeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarRangeFill']>
-    readonly BIconCalendarWeek: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarWeek']>
-    readonly BIconCalendarWeekFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarWeekFill']>
-    readonly BIconCalendarX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarX']>
-    readonly BIconCalendarXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCalendarXFill']>
-    readonly BIconCamera2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCamera2']>
-    readonly BIconCamera: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCamera']>
-    readonly BIconCameraFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraFill']>
-    readonly BIconCameraReels: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraReels']>
-    readonly BIconCameraReelsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraReelsFill']>
-    readonly BIconCameraVideo: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraVideo']>
-    readonly BIconCameraVideoFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraVideoFill']>
-    readonly BIconCameraVideoOff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraVideoOff']>
-    readonly BIconCameraVideoOffFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCameraVideoOffFill']>
-    readonly BIconCapslock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCapslock']>
-    readonly BIconCapslockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCapslockFill']>
-    readonly BIconCardChecklist: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCardChecklist']>
-    readonly BIconCardHeading: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCardHeading']>
-    readonly BIconCardImage: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCardImage']>
-    readonly BIconCardList: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCardList']>
-    readonly BIconCardText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCardText']>
-    readonly BIconCaretDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretDown']>
-    readonly BIconCaretDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretDownFill']>
-    readonly BIconCaretDownSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretDownSquare']>
-    readonly BIconCaretDownSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretDownSquareFill']>
-    readonly BIconCaretLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretLeft']>
-    readonly BIconCaretLeftFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretLeftFill']>
-    readonly BIconCaretLeftSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretLeftSquare']>
-    readonly BIconCaretLeftSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretLeftSquareFill']>
-    readonly BIconCaretRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretRight']>
-    readonly BIconCaretRightFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretRightFill']>
-    readonly BIconCaretRightSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretRightSquare']>
-    readonly BIconCaretRightSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretRightSquareFill']>
-    readonly BIconCaretUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretUp']>
-    readonly BIconCaretUpFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretUpFill']>
-    readonly BIconCaretUpSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretUpSquare']>
-    readonly BIconCaretUpSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCaretUpSquareFill']>
-    readonly BIconCart2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCart2']>
-    readonly BIconCart3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCart3']>
-    readonly BIconCart4: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCart4']>
-    readonly BIconCart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCart']>
-    readonly BIconCartCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartCheck']>
-    readonly BIconCartCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartCheckFill']>
-    readonly BIconCartDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartDash']>
-    readonly BIconCartDashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartDashFill']>
-    readonly BIconCartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartFill']>
-    readonly BIconCartPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartPlus']>
-    readonly BIconCartPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartPlusFill']>
-    readonly BIconCartX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartX']>
-    readonly BIconCartXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCartXFill']>
-    readonly BIconCash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCash']>
-    readonly BIconCashCoin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCashCoin']>
-    readonly BIconCashStack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCashStack']>
-    readonly BIconCast: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCast']>
-    readonly BIconChat: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChat']>
-    readonly BIconChatDots: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatDots']>
-    readonly BIconChatDotsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatDotsFill']>
-    readonly BIconChatFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatFill']>
-    readonly BIconChatHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatHeart']>
-    readonly BIconChatHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatHeartFill']>
-    readonly BIconChatLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeft']>
-    readonly BIconChatLeftDots: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftDots']>
-    readonly BIconChatLeftDotsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftDotsFill']>
-    readonly BIconChatLeftFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftFill']>
-    readonly BIconChatLeftHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftHeart']>
-    readonly BIconChatLeftHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftHeartFill']>
-    readonly BIconChatLeftQuote: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftQuote']>
-    readonly BIconChatLeftQuoteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftQuoteFill']>
-    readonly BIconChatLeftText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftText']>
-    readonly BIconChatLeftTextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatLeftTextFill']>
-    readonly BIconChatQuote: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatQuote']>
-    readonly BIconChatQuoteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatQuoteFill']>
-    readonly BIconChatRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRight']>
-    readonly BIconChatRightDots: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightDots']>
-    readonly BIconChatRightDotsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightDotsFill']>
-    readonly BIconChatRightFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightFill']>
-    readonly BIconChatRightHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightHeart']>
-    readonly BIconChatRightHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightHeartFill']>
-    readonly BIconChatRightQuote: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightQuote']>
-    readonly BIconChatRightQuoteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightQuoteFill']>
-    readonly BIconChatRightText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightText']>
-    readonly BIconChatRightTextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatRightTextFill']>
-    readonly BIconChatSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquare']>
-    readonly BIconChatSquareDots: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareDots']>
-    readonly BIconChatSquareDotsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareDotsFill']>
-    readonly BIconChatSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareFill']>
-    readonly BIconChatSquareHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareHeart']>
-    readonly BIconChatSquareHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareHeartFill']>
-    readonly BIconChatSquareQuote: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareQuote']>
-    readonly BIconChatSquareQuoteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareQuoteFill']>
-    readonly BIconChatSquareText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareText']>
-    readonly BIconChatSquareTextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatSquareTextFill']>
-    readonly BIconChatText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatText']>
-    readonly BIconChatTextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChatTextFill']>
-    readonly BIconCheck2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheck2']>
-    readonly BIconCheck2All: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheck2All']>
-    readonly BIconCheck2Circle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheck2Circle']>
-    readonly BIconCheck2Square: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheck2Square']>
-    readonly BIconCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheck']>
-    readonly BIconCheckAll: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheckAll']>
-    readonly BIconCheckCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheckCircle']>
-    readonly BIconCheckCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheckCircleFill']>
-    readonly BIconCheckLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheckLg']>
-    readonly BIconCheckSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheckSquare']>
-    readonly BIconCheckSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCheckSquareFill']>
-    readonly BIconChevronBarContract: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronBarContract']>
-    readonly BIconChevronBarDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronBarDown']>
-    readonly BIconChevronBarExpand: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronBarExpand']>
-    readonly BIconChevronBarLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronBarLeft']>
-    readonly BIconChevronBarRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronBarRight']>
-    readonly BIconChevronBarUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronBarUp']>
-    readonly BIconChevronCompactDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronCompactDown']>
-    readonly BIconChevronCompactLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronCompactLeft']>
-    readonly BIconChevronCompactRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronCompactRight']>
-    readonly BIconChevronCompactUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronCompactUp']>
-    readonly BIconChevronContract: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronContract']>
-    readonly BIconChevronDoubleDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronDoubleDown']>
-    readonly BIconChevronDoubleLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronDoubleLeft']>
-    readonly BIconChevronDoubleRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronDoubleRight']>
-    readonly BIconChevronDoubleUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronDoubleUp']>
-    readonly BIconChevronDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronDown']>
-    readonly BIconChevronExpand: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronExpand']>
-    readonly BIconChevronLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronLeft']>
-    readonly BIconChevronRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronRight']>
-    readonly BIconChevronUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconChevronUp']>
-    readonly BIconCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCircle']>
-    readonly BIconCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCircleFill']>
-    readonly BIconCircleHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCircleHalf']>
-    readonly BIconCircleSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCircleSquare']>
-    readonly BIconClipboard2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2']>
-    readonly BIconClipboard2Check: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Check']>
-    readonly BIconClipboard2CheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2CheckFill']>
-    readonly BIconClipboard2Data: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Data']>
-    readonly BIconClipboard2DataFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2DataFill']>
-    readonly BIconClipboard2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Fill']>
-    readonly BIconClipboard2Heart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Heart']>
-    readonly BIconClipboard2HeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2HeartFill']>
-    readonly BIconClipboard2Minus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Minus']>
-    readonly BIconClipboard2MinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2MinusFill']>
-    readonly BIconClipboard2Plus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Plus']>
-    readonly BIconClipboard2PlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2PlusFill']>
-    readonly BIconClipboard2Pulse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2Pulse']>
-    readonly BIconClipboard2PulseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2PulseFill']>
-    readonly BIconClipboard2X: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2X']>
-    readonly BIconClipboard2XFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard2XFill']>
-    readonly BIconClipboard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboard']>
-    readonly BIconClipboardCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardCheck']>
-    readonly BIconClipboardCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardCheckFill']>
-    readonly BIconClipboardData: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardData']>
-    readonly BIconClipboardDataFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardDataFill']>
-    readonly BIconClipboardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardFill']>
-    readonly BIconClipboardHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardHeart']>
-    readonly BIconClipboardHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardHeartFill']>
-    readonly BIconClipboardMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardMinus']>
-    readonly BIconClipboardMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardMinusFill']>
-    readonly BIconClipboardPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardPlus']>
-    readonly BIconClipboardPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardPlusFill']>
-    readonly BIconClipboardPulse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardPulse']>
-    readonly BIconClipboardX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardX']>
-    readonly BIconClipboardXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClipboardXFill']>
-    readonly BIconClock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClock']>
-    readonly BIconClockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClockFill']>
-    readonly BIconClockHistory: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClockHistory']>
-    readonly BIconCloud: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloud']>
-    readonly BIconCloudArrowDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudArrowDown']>
-    readonly BIconCloudArrowDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudArrowDownFill']>
-    readonly BIconCloudArrowUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudArrowUp']>
-    readonly BIconCloudArrowUpFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudArrowUpFill']>
-    readonly BIconCloudCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudCheck']>
-    readonly BIconCloudCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudCheckFill']>
-    readonly BIconCloudDownload: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudDownload']>
-    readonly BIconCloudDownloadFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudDownloadFill']>
-    readonly BIconCloudDrizzle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudDrizzle']>
-    readonly BIconCloudDrizzleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudDrizzleFill']>
-    readonly BIconCloudFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudFill']>
-    readonly BIconCloudFog2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudFog2']>
-    readonly BIconCloudFog2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudFog2Fill']>
-    readonly BIconCloudFog: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudFog']>
-    readonly BIconCloudFogFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudFogFill']>
-    readonly BIconCloudHail: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudHail']>
-    readonly BIconCloudHailFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudHailFill']>
-    readonly BIconCloudHaze2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudHaze2']>
-    readonly BIconCloudHaze2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudHaze2Fill']>
-    readonly BIconCloudHaze: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudHaze']>
-    readonly BIconCloudHazeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudHazeFill']>
-    readonly BIconCloudLightning: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudLightning']>
-    readonly BIconCloudLightningFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudLightningFill']>
-    readonly BIconCloudLightningRain: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudLightningRain']>
-    readonly BIconCloudLightningRainFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudLightningRainFill']>
-    readonly BIconCloudMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudMinus']>
-    readonly BIconCloudMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudMinusFill']>
-    readonly BIconCloudMoon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudMoon']>
-    readonly BIconCloudMoonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudMoonFill']>
-    readonly BIconCloudPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudPlus']>
-    readonly BIconCloudPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudPlusFill']>
-    readonly BIconCloudRain: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudRain']>
-    readonly BIconCloudRainFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudRainFill']>
-    readonly BIconCloudRainHeavy: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudRainHeavy']>
-    readonly BIconCloudRainHeavyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudRainHeavyFill']>
-    readonly BIconCloudSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSlash']>
-    readonly BIconCloudSlashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSlashFill']>
-    readonly BIconCloudSleet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSleet']>
-    readonly BIconCloudSleetFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSleetFill']>
-    readonly BIconCloudSnow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSnow']>
-    readonly BIconCloudSnowFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSnowFill']>
-    readonly BIconCloudSun: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSun']>
-    readonly BIconCloudSunFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudSunFill']>
-    readonly BIconCloudUpload: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudUpload']>
-    readonly BIconCloudUploadFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudUploadFill']>
-    readonly BIconClouds: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconClouds']>
-    readonly BIconCloudsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudsFill']>
-    readonly BIconCloudy: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudy']>
-    readonly BIconCloudyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCloudyFill']>
-    readonly BIconCode: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCode']>
-    readonly BIconCodeSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCodeSlash']>
-    readonly BIconCodeSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCodeSquare']>
-    readonly BIconCoin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCoin']>
-    readonly BIconCollection: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCollection']>
-    readonly BIconCollectionFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCollectionFill']>
-    readonly BIconCollectionPlay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCollectionPlay']>
-    readonly BIconCollectionPlayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCollectionPlayFill']>
-    readonly BIconColumns: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconColumns']>
-    readonly BIconColumnsGap: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconColumnsGap']>
-    readonly BIconCommand: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCommand']>
-    readonly BIconCompass: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCompass']>
-    readonly BIconCompassFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCompassFill']>
-    readonly BIconCone: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCone']>
-    readonly BIconConeStriped: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconConeStriped']>
-    readonly BIconController: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconController']>
-    readonly BIconCpu: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCpu']>
-    readonly BIconCpuFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCpuFill']>
-    readonly BIconCreditCard2Back: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCreditCard2Back']>
-    readonly BIconCreditCard2BackFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCreditCard2BackFill']>
-    readonly BIconCreditCard2Front: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCreditCard2Front']>
-    readonly BIconCreditCard2FrontFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCreditCard2FrontFill']>
-    readonly BIconCreditCard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCreditCard']>
-    readonly BIconCreditCardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCreditCardFill']>
-    readonly BIconCrop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCrop']>
-    readonly BIconCup: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCup']>
-    readonly BIconCupFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCupFill']>
-    readonly BIconCupStraw: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCupStraw']>
-    readonly BIconCurrencyBitcoin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCurrencyBitcoin']>
-    readonly BIconCurrencyDollar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCurrencyDollar']>
-    readonly BIconCurrencyEuro: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCurrencyEuro']>
-    readonly BIconCurrencyExchange: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCurrencyExchange']>
-    readonly BIconCurrencyPound: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCurrencyPound']>
-    readonly BIconCurrencyYen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCurrencyYen']>
-    readonly BIconCursor: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCursor']>
-    readonly BIconCursorFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCursorFill']>
-    readonly BIconCursorText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconCursorText']>
-    readonly BIconDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDash']>
-    readonly BIconDashCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashCircle']>
-    readonly BIconDashCircleDotted: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashCircleDotted']>
-    readonly BIconDashCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashCircleFill']>
-    readonly BIconDashLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashLg']>
-    readonly BIconDashSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashSquare']>
-    readonly BIconDashSquareDotted: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashSquareDotted']>
-    readonly BIconDashSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDashSquareFill']>
-    readonly BIconDeviceHdd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDeviceHdd']>
-    readonly BIconDeviceHddFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDeviceHddFill']>
-    readonly BIconDeviceSsd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDeviceSsd']>
-    readonly BIconDeviceSsdFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDeviceSsdFill']>
-    readonly BIconDiagram2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiagram2']>
-    readonly BIconDiagram2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiagram2Fill']>
-    readonly BIconDiagram3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiagram3']>
-    readonly BIconDiagram3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiagram3Fill']>
-    readonly BIconDiamond: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiamond']>
-    readonly BIconDiamondFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiamondFill']>
-    readonly BIconDiamondHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiamondHalf']>
-    readonly BIconDice1: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice1']>
-    readonly BIconDice1Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice1Fill']>
-    readonly BIconDice2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice2']>
-    readonly BIconDice2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice2Fill']>
-    readonly BIconDice3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice3']>
-    readonly BIconDice3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice3Fill']>
-    readonly BIconDice4: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice4']>
-    readonly BIconDice4Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice4Fill']>
-    readonly BIconDice5: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice5']>
-    readonly BIconDice5Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice5Fill']>
-    readonly BIconDice6: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice6']>
-    readonly BIconDice6Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDice6Fill']>
-    readonly BIconDisc: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDisc']>
-    readonly BIconDiscFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiscFill']>
-    readonly BIconDiscord: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDiscord']>
-    readonly BIconDisplay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDisplay']>
-    readonly BIconDisplayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDisplayFill']>
-    readonly BIconDisplayport: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDisplayport']>
-    readonly BIconDisplayportFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDisplayportFill']>
-    readonly BIconDistributeHorizontal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDistributeHorizontal']>
-    readonly BIconDistributeVertical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDistributeVertical']>
-    readonly BIconDoorClosed: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDoorClosed']>
-    readonly BIconDoorClosedFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDoorClosedFill']>
-    readonly BIconDoorOpen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDoorOpen']>
-    readonly BIconDoorOpenFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDoorOpenFill']>
-    readonly BIconDot: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDot']>
-    readonly BIconDownload: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDownload']>
-    readonly BIconDpad: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDpad']>
-    readonly BIconDpadFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDpadFill']>
-    readonly BIconDribbble: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDribbble']>
-    readonly BIconDroplet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDroplet']>
-    readonly BIconDropletFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDropletFill']>
-    readonly BIconDropletHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconDropletHalf']>
-    readonly BIconEar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEar']>
-    readonly BIconEarFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEarFill']>
-    readonly BIconEarbuds: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEarbuds']>
-    readonly BIconEasel2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEasel2']>
-    readonly BIconEasel2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEasel2Fill']>
-    readonly BIconEasel3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEasel3']>
-    readonly BIconEasel3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEasel3Fill']>
-    readonly BIconEasel: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEasel']>
-    readonly BIconEaselFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEaselFill']>
-    readonly BIconEgg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEgg']>
-    readonly BIconEggFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEggFill']>
-    readonly BIconEggFried: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEggFried']>
-    readonly BIconEject: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEject']>
-    readonly BIconEjectFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEjectFill']>
-    readonly BIconEmojiAngry: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiAngry']>
-    readonly BIconEmojiAngryFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiAngryFill']>
-    readonly BIconEmojiDizzy: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiDizzy']>
-    readonly BIconEmojiDizzyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiDizzyFill']>
-    readonly BIconEmojiExpressionless: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiExpressionless']>
-    readonly BIconEmojiExpressionlessFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiExpressionlessFill']>
-    readonly BIconEmojiFrown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiFrown']>
-    readonly BIconEmojiFrownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiFrownFill']>
-    readonly BIconEmojiHeartEyes: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiHeartEyes']>
-    readonly BIconEmojiHeartEyesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiHeartEyesFill']>
-    readonly BIconEmojiKiss: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiKiss']>
-    readonly BIconEmojiKissFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiKissFill']>
-    readonly BIconEmojiLaughing: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiLaughing']>
-    readonly BIconEmojiLaughingFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiLaughingFill']>
-    readonly BIconEmojiNeutral: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiNeutral']>
-    readonly BIconEmojiNeutralFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiNeutralFill']>
-    readonly BIconEmojiSmile: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiSmile']>
-    readonly BIconEmojiSmileFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiSmileFill']>
-    readonly BIconEmojiSmileUpsideDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiSmileUpsideDown']>
-    readonly BIconEmojiSmileUpsideDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiSmileUpsideDownFill']>
-    readonly BIconEmojiSunglasses: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiSunglasses']>
-    readonly BIconEmojiSunglassesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiSunglassesFill']>
-    readonly BIconEmojiWink: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiWink']>
-    readonly BIconEmojiWinkFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEmojiWinkFill']>
-    readonly BIconEnvelope: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelope']>
-    readonly BIconEnvelopeCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeCheck']>
-    readonly BIconEnvelopeCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeCheckFill']>
-    readonly BIconEnvelopeDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeDash']>
-    readonly BIconEnvelopeDashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeDashFill']>
-    readonly BIconEnvelopeExclamation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeExclamation']>
-    readonly BIconEnvelopeExclamationFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeExclamationFill']>
-    readonly BIconEnvelopeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeFill']>
-    readonly BIconEnvelopeHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeHeart']>
-    readonly BIconEnvelopeHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeHeartFill']>
-    readonly BIconEnvelopeOpen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeOpen']>
-    readonly BIconEnvelopeOpenFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeOpenFill']>
-    readonly BIconEnvelopeOpenHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeOpenHeart']>
-    readonly BIconEnvelopeOpenHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeOpenHeartFill']>
-    readonly BIconEnvelopePaper: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopePaper']>
-    readonly BIconEnvelopePaperFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopePaperFill']>
-    readonly BIconEnvelopePaperHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopePaperHeart']>
-    readonly BIconEnvelopePaperHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopePaperHeartFill']>
-    readonly BIconEnvelopePlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopePlus']>
-    readonly BIconEnvelopePlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopePlusFill']>
-    readonly BIconEnvelopeSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeSlash']>
-    readonly BIconEnvelopeSlashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeSlashFill']>
-    readonly BIconEnvelopeX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeX']>
-    readonly BIconEnvelopeXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEnvelopeXFill']>
-    readonly BIconEraser: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEraser']>
-    readonly BIconEraserFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEraserFill']>
-    readonly BIconEthernet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEthernet']>
-    readonly BIconExclamation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamation']>
-    readonly BIconExclamationCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationCircle']>
-    readonly BIconExclamationCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationCircleFill']>
-    readonly BIconExclamationDiamond: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationDiamond']>
-    readonly BIconExclamationDiamondFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationDiamondFill']>
-    readonly BIconExclamationLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationLg']>
-    readonly BIconExclamationOctagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationOctagon']>
-    readonly BIconExclamationOctagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationOctagonFill']>
-    readonly BIconExclamationSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationSquare']>
-    readonly BIconExclamationSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationSquareFill']>
-    readonly BIconExclamationTriangle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationTriangle']>
-    readonly BIconExclamationTriangleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclamationTriangleFill']>
-    readonly BIconExclude: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExclude']>
-    readonly BIconExplicit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExplicit']>
-    readonly BIconExplicitFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconExplicitFill']>
-    readonly BIconEye: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEye']>
-    readonly BIconEyeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEyeFill']>
-    readonly BIconEyeSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEyeSlash']>
-    readonly BIconEyeSlashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEyeSlashFill']>
-    readonly BIconEyedropper: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEyedropper']>
-    readonly BIconEyeglasses: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconEyeglasses']>
-    readonly BIconFacebook: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFacebook']>
-    readonly BIconFan: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFan']>
-    readonly BIconFile: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFile']>
-    readonly BIconFileArrowDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileArrowDown']>
-    readonly BIconFileArrowDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileArrowDownFill']>
-    readonly BIconFileArrowUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileArrowUp']>
-    readonly BIconFileArrowUpFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileArrowUpFill']>
-    readonly BIconFileBarGraph: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileBarGraph']>
-    readonly BIconFileBarGraphFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileBarGraphFill']>
-    readonly BIconFileBinary: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileBinary']>
-    readonly BIconFileBinaryFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileBinaryFill']>
-    readonly BIconFileBreak: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileBreak']>
-    readonly BIconFileBreakFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileBreakFill']>
-    readonly BIconFileCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileCheck']>
-    readonly BIconFileCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileCheckFill']>
-    readonly BIconFileCode: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileCode']>
-    readonly BIconFileCodeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileCodeFill']>
-    readonly BIconFileDiff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileDiff']>
-    readonly BIconFileDiffFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileDiffFill']>
-    readonly BIconFileEarmark: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmark']>
-    readonly BIconFileEarmarkArrowDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowDown']>
-    readonly BIconFileEarmarkArrowDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowDownFill']>
-    readonly BIconFileEarmarkArrowUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowUp']>
-    readonly BIconFileEarmarkArrowUpFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkArrowUpFill']>
-    readonly BIconFileEarmarkBarGraph: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkBarGraph']>
-    readonly BIconFileEarmarkBarGraphFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkBarGraphFill']>
-    readonly BIconFileEarmarkBinary: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkBinary']>
-    readonly BIconFileEarmarkBinaryFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkBinaryFill']>
-    readonly BIconFileEarmarkBreak: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkBreak']>
-    readonly BIconFileEarmarkBreakFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkBreakFill']>
-    readonly BIconFileEarmarkCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkCheck']>
-    readonly BIconFileEarmarkCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkCheckFill']>
-    readonly BIconFileEarmarkCode: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkCode']>
-    readonly BIconFileEarmarkCodeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkCodeFill']>
-    readonly BIconFileEarmarkDiff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkDiff']>
-    readonly BIconFileEarmarkDiffFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkDiffFill']>
-    readonly BIconFileEarmarkEasel: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkEasel']>
-    readonly BIconFileEarmarkEaselFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkEaselFill']>
-    readonly BIconFileEarmarkExcel: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkExcel']>
-    readonly BIconFileEarmarkExcelFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkExcelFill']>
-    readonly BIconFileEarmarkFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkFill']>
-    readonly BIconFileEarmarkFont: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkFont']>
-    readonly BIconFileEarmarkFontFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkFontFill']>
-    readonly BIconFileEarmarkImage: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkImage']>
-    readonly BIconFileEarmarkImageFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkImageFill']>
-    readonly BIconFileEarmarkLock2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkLock2']>
-    readonly BIconFileEarmarkLock2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkLock2Fill']>
-    readonly BIconFileEarmarkLock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkLock']>
-    readonly BIconFileEarmarkLockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkLockFill']>
-    readonly BIconFileEarmarkMedical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkMedical']>
-    readonly BIconFileEarmarkMedicalFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkMedicalFill']>
-    readonly BIconFileEarmarkMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkMinus']>
-    readonly BIconFileEarmarkMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkMinusFill']>
-    readonly BIconFileEarmarkMusic: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkMusic']>
-    readonly BIconFileEarmarkMusicFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkMusicFill']>
-    readonly BIconFileEarmarkPdf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPdf']>
-    readonly BIconFileEarmarkPdfFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPdfFill']>
-    readonly BIconFileEarmarkPerson: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPerson']>
-    readonly BIconFileEarmarkPersonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPersonFill']>
-    readonly BIconFileEarmarkPlay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlay']>
-    readonly BIconFileEarmarkPlayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlayFill']>
-    readonly BIconFileEarmarkPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlus']>
-    readonly BIconFileEarmarkPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPlusFill']>
-    readonly BIconFileEarmarkPost: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPost']>
-    readonly BIconFileEarmarkPostFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPostFill']>
-    readonly BIconFileEarmarkPpt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPpt']>
-    readonly BIconFileEarmarkPptFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkPptFill']>
-    readonly BIconFileEarmarkRichtext: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkRichtext']>
-    readonly BIconFileEarmarkRichtextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkRichtextFill']>
-    readonly BIconFileEarmarkRuled: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkRuled']>
-    readonly BIconFileEarmarkRuledFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkRuledFill']>
-    readonly BIconFileEarmarkSlides: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkSlides']>
-    readonly BIconFileEarmarkSlidesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkSlidesFill']>
-    readonly BIconFileEarmarkSpreadsheet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkSpreadsheet']>
-    readonly BIconFileEarmarkSpreadsheetFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkSpreadsheetFill']>
-    readonly BIconFileEarmarkText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkText']>
-    readonly BIconFileEarmarkTextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkTextFill']>
-    readonly BIconFileEarmarkWord: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkWord']>
-    readonly BIconFileEarmarkWordFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkWordFill']>
-    readonly BIconFileEarmarkX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkX']>
-    readonly BIconFileEarmarkXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkXFill']>
-    readonly BIconFileEarmarkZip: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkZip']>
-    readonly BIconFileEarmarkZipFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEarmarkZipFill']>
-    readonly BIconFileEasel: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEasel']>
-    readonly BIconFileEaselFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileEaselFill']>
-    readonly BIconFileExcel: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileExcel']>
-    readonly BIconFileExcelFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileExcelFill']>
-    readonly BIconFileFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileFill']>
-    readonly BIconFileFont: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileFont']>
-    readonly BIconFileFontFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileFontFill']>
-    readonly BIconFileImage: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileImage']>
-    readonly BIconFileImageFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileImageFill']>
-    readonly BIconFileLock2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileLock2']>
-    readonly BIconFileLock2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileLock2Fill']>
-    readonly BIconFileLock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileLock']>
-    readonly BIconFileLockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileLockFill']>
-    readonly BIconFileMedical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileMedical']>
-    readonly BIconFileMedicalFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileMedicalFill']>
-    readonly BIconFileMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileMinus']>
-    readonly BIconFileMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileMinusFill']>
-    readonly BIconFileMusic: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileMusic']>
-    readonly BIconFileMusicFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileMusicFill']>
-    readonly BIconFilePdf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePdf']>
-    readonly BIconFilePdfFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePdfFill']>
-    readonly BIconFilePerson: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePerson']>
-    readonly BIconFilePersonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePersonFill']>
-    readonly BIconFilePlay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePlay']>
-    readonly BIconFilePlayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePlayFill']>
-    readonly BIconFilePlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePlus']>
-    readonly BIconFilePlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePlusFill']>
-    readonly BIconFilePost: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePost']>
-    readonly BIconFilePostFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePostFill']>
-    readonly BIconFilePpt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePpt']>
-    readonly BIconFilePptFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilePptFill']>
-    readonly BIconFileRichtext: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileRichtext']>
-    readonly BIconFileRichtextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileRichtextFill']>
-    readonly BIconFileRuled: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileRuled']>
-    readonly BIconFileRuledFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileRuledFill']>
-    readonly BIconFileSlides: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileSlides']>
-    readonly BIconFileSlidesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileSlidesFill']>
-    readonly BIconFileSpreadsheet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileSpreadsheet']>
-    readonly BIconFileSpreadsheetFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileSpreadsheetFill']>
-    readonly BIconFileText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileText']>
-    readonly BIconFileTextFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileTextFill']>
-    readonly BIconFileWord: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileWord']>
-    readonly BIconFileWordFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileWordFill']>
-    readonly BIconFileX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileX']>
-    readonly BIconFileXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileXFill']>
-    readonly BIconFileZip: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileZip']>
-    readonly BIconFileZipFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFileZipFill']>
-    readonly BIconFiles: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiles']>
-    readonly BIconFilesAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilesAlt']>
-    readonly BIconFiletypeAac: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeAac']>
-    readonly BIconFiletypeAi: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeAi']>
-    readonly BIconFiletypeBmp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeBmp']>
-    readonly BIconFiletypeCs: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeCs']>
-    readonly BIconFiletypeCss: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeCss']>
-    readonly BIconFiletypeCsv: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeCsv']>
-    readonly BIconFiletypeDoc: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeDoc']>
-    readonly BIconFiletypeDocx: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeDocx']>
-    readonly BIconFiletypeExe: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeExe']>
-    readonly BIconFiletypeGif: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeGif']>
-    readonly BIconFiletypeHeic: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeHeic']>
-    readonly BIconFiletypeHtml: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeHtml']>
-    readonly BIconFiletypeJava: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeJava']>
-    readonly BIconFiletypeJpg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeJpg']>
-    readonly BIconFiletypeJs: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeJs']>
-    readonly BIconFiletypeJson: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeJson']>
-    readonly BIconFiletypeJsx: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeJsx']>
-    readonly BIconFiletypeKey: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeKey']>
-    readonly BIconFiletypeM4p: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeM4p']>
-    readonly BIconFiletypeMd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeMd']>
-    readonly BIconFiletypeMdx: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeMdx']>
-    readonly BIconFiletypeMov: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeMov']>
-    readonly BIconFiletypeMp3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeMp3']>
-    readonly BIconFiletypeMp4: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeMp4']>
-    readonly BIconFiletypeOtf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeOtf']>
-    readonly BIconFiletypePdf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePdf']>
-    readonly BIconFiletypePhp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePhp']>
-    readonly BIconFiletypePng: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePng']>
-    readonly BIconFiletypePpt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePpt']>
-    readonly BIconFiletypePptx: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePptx']>
-    readonly BIconFiletypePsd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePsd']>
-    readonly BIconFiletypePy: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypePy']>
-    readonly BIconFiletypeRaw: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeRaw']>
-    readonly BIconFiletypeRb: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeRb']>
-    readonly BIconFiletypeSass: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeSass']>
-    readonly BIconFiletypeScss: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeScss']>
-    readonly BIconFiletypeSh: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeSh']>
-    readonly BIconFiletypeSvg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeSvg']>
-    readonly BIconFiletypeTiff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeTiff']>
-    readonly BIconFiletypeTsx: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeTsx']>
-    readonly BIconFiletypeTtf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeTtf']>
-    readonly BIconFiletypeTxt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeTxt']>
-    readonly BIconFiletypeWav: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeWav']>
-    readonly BIconFiletypeWoff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeWoff']>
-    readonly BIconFiletypeXls: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeXls']>
-    readonly BIconFiletypeXlsx: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeXlsx']>
-    readonly BIconFiletypeXml: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeXml']>
-    readonly BIconFiletypeYml: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFiletypeYml']>
-    readonly BIconFilm: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilm']>
-    readonly BIconFilter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilter']>
-    readonly BIconFilterCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilterCircle']>
-    readonly BIconFilterCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilterCircleFill']>
-    readonly BIconFilterLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilterLeft']>
-    readonly BIconFilterRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilterRight']>
-    readonly BIconFilterSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilterSquare']>
-    readonly BIconFilterSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFilterSquareFill']>
-    readonly BIconFingerprint: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFingerprint']>
-    readonly BIconFlag: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFlag']>
-    readonly BIconFlagFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFlagFill']>
-    readonly BIconFlower1: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFlower1']>
-    readonly BIconFlower2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFlower2']>
-    readonly BIconFlower3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFlower3']>
-    readonly BIconFolder2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolder2']>
-    readonly BIconFolder2Open: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolder2Open']>
-    readonly BIconFolder: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolder']>
-    readonly BIconFolderCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderCheck']>
-    readonly BIconFolderFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderFill']>
-    readonly BIconFolderMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderMinus']>
-    readonly BIconFolderPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderPlus']>
-    readonly BIconFolderSymlink: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderSymlink']>
-    readonly BIconFolderSymlinkFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderSymlinkFill']>
-    readonly BIconFolderX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFolderX']>
-    readonly BIconFonts: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFonts']>
-    readonly BIconForward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconForward']>
-    readonly BIconForwardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconForwardFill']>
-    readonly BIconFront: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFront']>
-    readonly BIconFullscreen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFullscreen']>
-    readonly BIconFullscreenExit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFullscreenExit']>
-    readonly BIconFunnel: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFunnel']>
-    readonly BIconFunnelFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconFunnelFill']>
-    readonly BIconGear: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGear']>
-    readonly BIconGearFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGearFill']>
-    readonly BIconGearWide: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGearWide']>
-    readonly BIconGearWideConnected: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGearWideConnected']>
-    readonly BIconGem: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGem']>
-    readonly BIconGenderAmbiguous: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGenderAmbiguous']>
-    readonly BIconGenderFemale: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGenderFemale']>
-    readonly BIconGenderMale: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGenderMale']>
-    readonly BIconGenderTrans: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGenderTrans']>
-    readonly BIconGeo: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGeo']>
-    readonly BIconGeoAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGeoAlt']>
-    readonly BIconGeoAltFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGeoAltFill']>
-    readonly BIconGeoFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGeoFill']>
-    readonly BIconGift: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGift']>
-    readonly BIconGiftFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGiftFill']>
-    readonly BIconGit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGit']>
-    readonly BIconGithub: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGithub']>
-    readonly BIconGlobe2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGlobe2']>
-    readonly BIconGlobe: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGlobe']>
-    readonly BIconGoogle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGoogle']>
-    readonly BIconGpuCard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGpuCard']>
-    readonly BIconGraphDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGraphDown']>
-    readonly BIconGraphDownArrow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGraphDownArrow']>
-    readonly BIconGraphUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGraphUp']>
-    readonly BIconGraphUpArrow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGraphUpArrow']>
-    readonly BIconGrid1x2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid1x2']>
-    readonly BIconGrid1x2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid1x2Fill']>
-    readonly BIconGrid3x2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid3x2']>
-    readonly BIconGrid3x2Gap: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid3x2Gap']>
-    readonly BIconGrid3x2GapFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid3x2GapFill']>
-    readonly BIconGrid3x3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid3x3']>
-    readonly BIconGrid3x3Gap: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid3x3Gap']>
-    readonly BIconGrid3x3GapFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid3x3GapFill']>
-    readonly BIconGrid: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGrid']>
-    readonly BIconGridFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGridFill']>
-    readonly BIconGripHorizontal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGripHorizontal']>
-    readonly BIconGripVertical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconGripVertical']>
-    readonly BIconHammer: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHammer']>
-    readonly BIconHandIndex: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandIndex']>
-    readonly BIconHandIndexFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandIndexFill']>
-    readonly BIconHandIndexThumb: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandIndexThumb']>
-    readonly BIconHandIndexThumbFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandIndexThumbFill']>
-    readonly BIconHandThumbsDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandThumbsDown']>
-    readonly BIconHandThumbsDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandThumbsDownFill']>
-    readonly BIconHandThumbsUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandThumbsUp']>
-    readonly BIconHandThumbsUpFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandThumbsUpFill']>
-    readonly BIconHandbag: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandbag']>
-    readonly BIconHandbagFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHandbagFill']>
-    readonly BIconHash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHash']>
-    readonly BIconHdd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHdd']>
-    readonly BIconHddFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddFill']>
-    readonly BIconHddNetwork: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddNetwork']>
-    readonly BIconHddNetworkFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddNetworkFill']>
-    readonly BIconHddRack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddRack']>
-    readonly BIconHddRackFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddRackFill']>
-    readonly BIconHddStack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddStack']>
-    readonly BIconHddStackFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHddStackFill']>
-    readonly BIconHdmi: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHdmi']>
-    readonly BIconHdmiFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHdmiFill']>
-    readonly BIconHeadphones: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeadphones']>
-    readonly BIconHeadset: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeadset']>
-    readonly BIconHeadsetVr: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeadsetVr']>
-    readonly BIconHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeart']>
-    readonly BIconHeartArrow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartArrow']>
-    readonly BIconHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartFill']>
-    readonly BIconHeartHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartHalf']>
-    readonly BIconHeartPulse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartPulse']>
-    readonly BIconHeartPulseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartPulseFill']>
-    readonly BIconHeartbreak: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartbreak']>
-    readonly BIconHeartbreakFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeartbreakFill']>
-    readonly BIconHearts: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHearts']>
-    readonly BIconHeptagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeptagon']>
-    readonly BIconHeptagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeptagonFill']>
-    readonly BIconHeptagonHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHeptagonHalf']>
-    readonly BIconHexagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHexagon']>
-    readonly BIconHexagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHexagonFill']>
-    readonly BIconHexagonHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHexagonHalf']>
-    readonly BIconHospital: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHospital']>
-    readonly BIconHospitalFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHospitalFill']>
-    readonly BIconHourglass: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHourglass']>
-    readonly BIconHourglassBottom: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHourglassBottom']>
-    readonly BIconHourglassSplit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHourglassSplit']>
-    readonly BIconHourglassTop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHourglassTop']>
-    readonly BIconHouse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHouse']>
-    readonly BIconHouseDoor: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHouseDoor']>
-    readonly BIconHouseDoorFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHouseDoorFill']>
-    readonly BIconHouseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHouseFill']>
-    readonly BIconHouseHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHouseHeart']>
-    readonly BIconHouseHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHouseHeartFill']>
-    readonly BIconHr: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHr']>
-    readonly BIconHurricane: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHurricane']>
-    readonly BIconHypnotize: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconHypnotize']>
-    readonly BIconImage: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconImage']>
-    readonly BIconImageAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconImageAlt']>
-    readonly BIconImageFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconImageFill']>
-    readonly BIconImages: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconImages']>
-    readonly BIconInbox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInbox']>
-    readonly BIconInboxFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInboxFill']>
-    readonly BIconInboxes: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInboxes']>
-    readonly BIconInboxesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInboxesFill']>
-    readonly BIconIncognito: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconIncognito']>
-    readonly BIconInfinity: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfinity']>
-    readonly BIconInfo: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfo']>
-    readonly BIconInfoCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfoCircle']>
-    readonly BIconInfoCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfoCircleFill']>
-    readonly BIconInfoLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfoLg']>
-    readonly BIconInfoSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfoSquare']>
-    readonly BIconInfoSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInfoSquareFill']>
-    readonly BIconInputCursor: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInputCursor']>
-    readonly BIconInputCursorText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInputCursorText']>
-    readonly BIconInstagram: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconInstagram']>
-    readonly BIconIntersect: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconIntersect']>
-    readonly BIconJournal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournal']>
-    readonly BIconJournalAlbum: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalAlbum']>
-    readonly BIconJournalArrowDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalArrowDown']>
-    readonly BIconJournalArrowUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalArrowUp']>
-    readonly BIconJournalBookmark: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalBookmark']>
-    readonly BIconJournalBookmarkFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalBookmarkFill']>
-    readonly BIconJournalCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalCheck']>
-    readonly BIconJournalCode: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalCode']>
-    readonly BIconJournalMedical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalMedical']>
-    readonly BIconJournalMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalMinus']>
-    readonly BIconJournalPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalPlus']>
-    readonly BIconJournalRichtext: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalRichtext']>
-    readonly BIconJournalText: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalText']>
-    readonly BIconJournalX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournalX']>
-    readonly BIconJournals: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJournals']>
-    readonly BIconJoystick: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJoystick']>
-    readonly BIconJustify: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJustify']>
-    readonly BIconJustifyLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJustifyLeft']>
-    readonly BIconJustifyRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconJustifyRight']>
-    readonly BIconKanban: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconKanban']>
-    readonly BIconKanbanFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconKanbanFill']>
-    readonly BIconKey: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconKey']>
-    readonly BIconKeyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconKeyFill']>
-    readonly BIconKeyboard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconKeyboard']>
-    readonly BIconKeyboardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconKeyboardFill']>
-    readonly BIconLadder: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLadder']>
-    readonly BIconLamp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLamp']>
-    readonly BIconLampFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLampFill']>
-    readonly BIconLaptop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLaptop']>
-    readonly BIconLaptopFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLaptopFill']>
-    readonly BIconLayerBackward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayerBackward']>
-    readonly BIconLayerForward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayerForward']>
-    readonly BIconLayers: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayers']>
-    readonly BIconLayersFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayersFill']>
-    readonly BIconLayersHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayersHalf']>
-    readonly BIconLayoutSidebar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutSidebar']>
-    readonly BIconLayoutSidebarInset: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutSidebarInset']>
-    readonly BIconLayoutSidebarInsetReverse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutSidebarInsetReverse']>
-    readonly BIconLayoutSidebarReverse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutSidebarReverse']>
-    readonly BIconLayoutSplit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutSplit']>
-    readonly BIconLayoutTextSidebar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutTextSidebar']>
-    readonly BIconLayoutTextSidebarReverse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutTextSidebarReverse']>
-    readonly BIconLayoutTextWindow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutTextWindow']>
-    readonly BIconLayoutTextWindowReverse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutTextWindowReverse']>
-    readonly BIconLayoutThreeColumns: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutThreeColumns']>
-    readonly BIconLayoutWtf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLayoutWtf']>
-    readonly BIconLifePreserver: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLifePreserver']>
-    readonly BIconLightbulb: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightbulb']>
-    readonly BIconLightbulbFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightbulbFill']>
-    readonly BIconLightbulbOff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightbulbOff']>
-    readonly BIconLightbulbOffFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightbulbOffFill']>
-    readonly BIconLightning: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightning']>
-    readonly BIconLightningCharge: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightningCharge']>
-    readonly BIconLightningChargeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightningChargeFill']>
-    readonly BIconLightningFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLightningFill']>
-    readonly BIconLine: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLine']>
-    readonly BIconLink45deg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLink45deg']>
-    readonly BIconLink: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLink']>
-    readonly BIconLinkedin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLinkedin']>
-    readonly BIconList: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconList']>
-    readonly BIconListCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListCheck']>
-    readonly BIconListColumns: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListColumns']>
-    readonly BIconListColumnsReverse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListColumnsReverse']>
-    readonly BIconListNested: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListNested']>
-    readonly BIconListOl: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListOl']>
-    readonly BIconListStars: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListStars']>
-    readonly BIconListTask: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListTask']>
-    readonly BIconListUl: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconListUl']>
-    readonly BIconLock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLock']>
-    readonly BIconLockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconLockFill']>
-    readonly BIconMagic: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMagic']>
-    readonly BIconMagnet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMagnet']>
-    readonly BIconMagnetFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMagnetFill']>
-    readonly BIconMailbox2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMailbox2']>
-    readonly BIconMailbox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMailbox']>
-    readonly BIconMap: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMap']>
-    readonly BIconMapFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMapFill']>
-    readonly BIconMarkdown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMarkdown']>
-    readonly BIconMarkdownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMarkdownFill']>
-    readonly BIconMask: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMask']>
-    readonly BIconMastodon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMastodon']>
-    readonly BIconMedium: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMedium']>
-    readonly BIconMegaphone: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMegaphone']>
-    readonly BIconMegaphoneFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMegaphoneFill']>
-    readonly BIconMemory: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMemory']>
-    readonly BIconMenuApp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuApp']>
-    readonly BIconMenuAppFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuAppFill']>
-    readonly BIconMenuButton: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuButton']>
-    readonly BIconMenuButtonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuButtonFill']>
-    readonly BIconMenuButtonWide: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuButtonWide']>
-    readonly BIconMenuButtonWideFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuButtonWideFill']>
-    readonly BIconMenuDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuDown']>
-    readonly BIconMenuUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMenuUp']>
-    readonly BIconMessenger: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMessenger']>
-    readonly BIconMeta: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMeta']>
-    readonly BIconMic: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMic']>
-    readonly BIconMicFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMicFill']>
-    readonly BIconMicMute: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMicMute']>
-    readonly BIconMicMuteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMicMuteFill']>
-    readonly BIconMicrosoft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMicrosoft']>
-    readonly BIconMinecart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMinecart']>
-    readonly BIconMinecartLoaded: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMinecartLoaded']>
-    readonly BIconModem: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconModem']>
-    readonly BIconModemFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconModemFill']>
-    readonly BIconMoisture: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMoisture']>
-    readonly BIconMoon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMoon']>
-    readonly BIconMoonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMoonFill']>
-    readonly BIconMoonStars: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMoonStars']>
-    readonly BIconMoonStarsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMoonStarsFill']>
-    readonly BIconMortarboard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMortarboard']>
-    readonly BIconMortarboardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMortarboardFill']>
-    readonly BIconMotherboard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMotherboard']>
-    readonly BIconMotherboardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMotherboardFill']>
-    readonly BIconMouse2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMouse2']>
-    readonly BIconMouse2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMouse2Fill']>
-    readonly BIconMouse3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMouse3']>
-    readonly BIconMouse3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMouse3Fill']>
-    readonly BIconMouse: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMouse']>
-    readonly BIconMouseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMouseFill']>
-    readonly BIconMusicNote: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMusicNote']>
-    readonly BIconMusicNoteBeamed: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMusicNoteBeamed']>
-    readonly BIconMusicNoteList: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMusicNoteList']>
-    readonly BIconMusicPlayer: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMusicPlayer']>
-    readonly BIconMusicPlayerFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconMusicPlayerFill']>
-    readonly BIconNewspaper: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNewspaper']>
-    readonly BIconNintendoSwitch: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNintendoSwitch']>
-    readonly BIconNodeMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNodeMinus']>
-    readonly BIconNodeMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNodeMinusFill']>
-    readonly BIconNodePlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNodePlus']>
-    readonly BIconNodePlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNodePlusFill']>
-    readonly BIconNut: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNut']>
-    readonly BIconNutFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconNutFill']>
-    readonly BIconOctagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOctagon']>
-    readonly BIconOctagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOctagonFill']>
-    readonly BIconOctagonHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOctagonHalf']>
-    readonly BIconOpticalAudio: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOpticalAudio']>
-    readonly BIconOpticalAudioFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOpticalAudioFill']>
-    readonly BIconOption: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOption']>
-    readonly BIconOutlet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconOutlet']>
-    readonly BIconPaintBucket: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPaintBucket']>
-    readonly BIconPalette2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPalette2']>
-    readonly BIconPalette: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPalette']>
-    readonly BIconPaletteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPaletteFill']>
-    readonly BIconPaperclip: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPaperclip']>
-    readonly BIconParagraph: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconParagraph']>
-    readonly BIconPatchCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchCheck']>
-    readonly BIconPatchCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchCheckFill']>
-    readonly BIconPatchExclamation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchExclamation']>
-    readonly BIconPatchExclamationFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchExclamationFill']>
-    readonly BIconPatchMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchMinus']>
-    readonly BIconPatchMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchMinusFill']>
-    readonly BIconPatchPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchPlus']>
-    readonly BIconPatchPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchPlusFill']>
-    readonly BIconPatchQuestion: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchQuestion']>
-    readonly BIconPatchQuestionFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPatchQuestionFill']>
-    readonly BIconPause: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPause']>
-    readonly BIconPauseBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPauseBtn']>
-    readonly BIconPauseBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPauseBtnFill']>
-    readonly BIconPauseCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPauseCircle']>
-    readonly BIconPauseCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPauseCircleFill']>
-    readonly BIconPauseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPauseFill']>
-    readonly BIconPaypal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPaypal']>
-    readonly BIconPc: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPc']>
-    readonly BIconPcDisplay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPcDisplay']>
-    readonly BIconPcDisplayHorizontal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPcDisplayHorizontal']>
-    readonly BIconPcHorizontal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPcHorizontal']>
-    readonly BIconPciCard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPciCard']>
-    readonly BIconPeace: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPeace']>
-    readonly BIconPeaceFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPeaceFill']>
-    readonly BIconPen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPen']>
-    readonly BIconPenFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPenFill']>
-    readonly BIconPencil: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPencil']>
-    readonly BIconPencilFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPencilFill']>
-    readonly BIconPencilSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPencilSquare']>
-    readonly BIconPentagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPentagon']>
-    readonly BIconPentagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPentagonFill']>
-    readonly BIconPentagonHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPentagonHalf']>
-    readonly BIconPeople: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPeople']>
-    readonly BIconPeopleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPeopleFill']>
-    readonly BIconPercent: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPercent']>
-    readonly BIconPerson: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPerson']>
-    readonly BIconPersonBadge: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonBadge']>
-    readonly BIconPersonBadgeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonBadgeFill']>
-    readonly BIconPersonBoundingBox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonBoundingBox']>
-    readonly BIconPersonCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonCheck']>
-    readonly BIconPersonCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonCheckFill']>
-    readonly BIconPersonCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonCircle']>
-    readonly BIconPersonDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonDash']>
-    readonly BIconPersonDashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonDashFill']>
-    readonly BIconPersonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonFill']>
-    readonly BIconPersonHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonHeart']>
-    readonly BIconPersonHearts: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonHearts']>
-    readonly BIconPersonLinesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonLinesFill']>
-    readonly BIconPersonPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonPlus']>
-    readonly BIconPersonPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonPlusFill']>
-    readonly BIconPersonRolodex: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonRolodex']>
-    readonly BIconPersonSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonSquare']>
-    readonly BIconPersonVideo2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonVideo2']>
-    readonly BIconPersonVideo3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonVideo3']>
-    readonly BIconPersonVideo: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonVideo']>
-    readonly BIconPersonWorkspace: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonWorkspace']>
-    readonly BIconPersonX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonX']>
-    readonly BIconPersonXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPersonXFill']>
-    readonly BIconPhone: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhone']>
-    readonly BIconPhoneFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhoneFill']>
-    readonly BIconPhoneFlip: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhoneFlip']>
-    readonly BIconPhoneLandscape: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhoneLandscape']>
-    readonly BIconPhoneLandscapeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhoneLandscapeFill']>
-    readonly BIconPhoneVibrate: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhoneVibrate']>
-    readonly BIconPhoneVibrateFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPhoneVibrateFill']>
-    readonly BIconPieChart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPieChart']>
-    readonly BIconPieChartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPieChartFill']>
-    readonly BIconPiggyBank: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPiggyBank']>
-    readonly BIconPiggyBankFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPiggyBankFill']>
-    readonly BIconPin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPin']>
-    readonly BIconPinAngle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPinAngle']>
-    readonly BIconPinAngleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPinAngleFill']>
-    readonly BIconPinFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPinFill']>
-    readonly BIconPinMap: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPinMap']>
-    readonly BIconPinMapFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPinMapFill']>
-    readonly BIconPinterest: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPinterest']>
-    readonly BIconPip: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPip']>
-    readonly BIconPipFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPipFill']>
-    readonly BIconPlay: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlay']>
-    readonly BIconPlayBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlayBtn']>
-    readonly BIconPlayBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlayBtnFill']>
-    readonly BIconPlayCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlayCircle']>
-    readonly BIconPlayCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlayCircleFill']>
-    readonly BIconPlayFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlayFill']>
-    readonly BIconPlaystation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlaystation']>
-    readonly BIconPlug: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlug']>
-    readonly BIconPlugFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlugFill']>
-    readonly BIconPlugin: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlugin']>
-    readonly BIconPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlus']>
-    readonly BIconPlusCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusCircle']>
-    readonly BIconPlusCircleDotted: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusCircleDotted']>
-    readonly BIconPlusCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusCircleFill']>
-    readonly BIconPlusLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusLg']>
-    readonly BIconPlusSlashMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusSlashMinus']>
-    readonly BIconPlusSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusSquare']>
-    readonly BIconPlusSquareDotted: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusSquareDotted']>
-    readonly BIconPlusSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPlusSquareFill']>
-    readonly BIconPostage: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostage']>
-    readonly BIconPostageFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostageFill']>
-    readonly BIconPostageHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostageHeart']>
-    readonly BIconPostageHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostageHeartFill']>
-    readonly BIconPostcard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostcard']>
-    readonly BIconPostcardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostcardFill']>
-    readonly BIconPostcardHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostcardHeart']>
-    readonly BIconPostcardHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPostcardHeartFill']>
-    readonly BIconPower: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPower']>
-    readonly BIconPrinter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPrinter']>
-    readonly BIconPrinterFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPrinterFill']>
-    readonly BIconProjector: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconProjector']>
-    readonly BIconProjectorFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconProjectorFill']>
-    readonly BIconPuzzle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPuzzle']>
-    readonly BIconPuzzleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconPuzzleFill']>
-    readonly BIconQrCode: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQrCode']>
-    readonly BIconQrCodeScan: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQrCodeScan']>
-    readonly BIconQuestion: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestion']>
-    readonly BIconQuestionCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionCircle']>
-    readonly BIconQuestionCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionCircleFill']>
-    readonly BIconQuestionDiamond: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionDiamond']>
-    readonly BIconQuestionDiamondFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionDiamondFill']>
-    readonly BIconQuestionLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionLg']>
-    readonly BIconQuestionOctagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionOctagon']>
-    readonly BIconQuestionOctagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionOctagonFill']>
-    readonly BIconQuestionSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionSquare']>
-    readonly BIconQuestionSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuestionSquareFill']>
-    readonly BIconQuora: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuora']>
-    readonly BIconQuote: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconQuote']>
-    readonly BIconRadioactive: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRadioactive']>
-    readonly BIconRainbow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRainbow']>
-    readonly BIconReceipt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReceipt']>
-    readonly BIconReceiptCutoff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReceiptCutoff']>
-    readonly BIconReception0: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReception0']>
-    readonly BIconReception1: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReception1']>
-    readonly BIconReception2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReception2']>
-    readonly BIconReception3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReception3']>
-    readonly BIconReception4: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReception4']>
-    readonly BIconRecord2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecord2']>
-    readonly BIconRecord2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecord2Fill']>
-    readonly BIconRecord: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecord']>
-    readonly BIconRecordBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecordBtn']>
-    readonly BIconRecordBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecordBtnFill']>
-    readonly BIconRecordCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecordCircle']>
-    readonly BIconRecordCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecordCircleFill']>
-    readonly BIconRecordFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecordFill']>
-    readonly BIconRecycle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRecycle']>
-    readonly BIconReddit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReddit']>
-    readonly BIconReply: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReply']>
-    readonly BIconReplyAll: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReplyAll']>
-    readonly BIconReplyAllFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReplyAllFill']>
-    readonly BIconReplyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconReplyFill']>
-    readonly BIconRobot: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRobot']>
-    readonly BIconRouter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRouter']>
-    readonly BIconRouterFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRouterFill']>
-    readonly BIconRss: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRss']>
-    readonly BIconRssFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRssFill']>
-    readonly BIconRulers: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconRulers']>
-    readonly BIconSafe2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSafe2']>
-    readonly BIconSafe2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSafe2Fill']>
-    readonly BIconSafe: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSafe']>
-    readonly BIconSafeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSafeFill']>
-    readonly BIconSave2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSave2']>
-    readonly BIconSave2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSave2Fill']>
-    readonly BIconSave: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSave']>
-    readonly BIconSaveFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSaveFill']>
-    readonly BIconScissors: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconScissors']>
-    readonly BIconScrewdriver: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconScrewdriver']>
-    readonly BIconSdCard: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSdCard']>
-    readonly BIconSdCardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSdCardFill']>
-    readonly BIconSearch: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSearch']>
-    readonly BIconSearchHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSearchHeart']>
-    readonly BIconSearchHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSearchHeartFill']>
-    readonly BIconSegmentedNav: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSegmentedNav']>
-    readonly BIconSend: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSend']>
-    readonly BIconSendCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendCheck']>
-    readonly BIconSendCheckFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendCheckFill']>
-    readonly BIconSendDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendDash']>
-    readonly BIconSendDashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendDashFill']>
-    readonly BIconSendExclamation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendExclamation']>
-    readonly BIconSendExclamationFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendExclamationFill']>
-    readonly BIconSendFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendFill']>
-    readonly BIconSendPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendPlus']>
-    readonly BIconSendPlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendPlusFill']>
-    readonly BIconSendSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendSlash']>
-    readonly BIconSendSlashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendSlashFill']>
-    readonly BIconSendX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendX']>
-    readonly BIconSendXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSendXFill']>
-    readonly BIconServer: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconServer']>
-    readonly BIconShare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShare']>
-    readonly BIconShareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShareFill']>
-    readonly BIconShield: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShield']>
-    readonly BIconShieldCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldCheck']>
-    readonly BIconShieldExclamation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldExclamation']>
-    readonly BIconShieldFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldFill']>
-    readonly BIconShieldFillCheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldFillCheck']>
-    readonly BIconShieldFillExclamation: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldFillExclamation']>
-    readonly BIconShieldFillMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldFillMinus']>
-    readonly BIconShieldFillPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldFillPlus']>
-    readonly BIconShieldFillX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldFillX']>
-    readonly BIconShieldLock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldLock']>
-    readonly BIconShieldLockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldLockFill']>
-    readonly BIconShieldMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldMinus']>
-    readonly BIconShieldPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldPlus']>
-    readonly BIconShieldShaded: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldShaded']>
-    readonly BIconShieldSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldSlash']>
-    readonly BIconShieldSlashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldSlashFill']>
-    readonly BIconShieldX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShieldX']>
-    readonly BIconShift: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShift']>
-    readonly BIconShiftFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShiftFill']>
-    readonly BIconShop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShop']>
-    readonly BIconShopWindow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShopWindow']>
-    readonly BIconShuffle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconShuffle']>
-    readonly BIconSignal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignal']>
-    readonly BIconSignpost2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignpost2']>
-    readonly BIconSignpost2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignpost2Fill']>
-    readonly BIconSignpost: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignpost']>
-    readonly BIconSignpostFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignpostFill']>
-    readonly BIconSignpostSplit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignpostSplit']>
-    readonly BIconSignpostSplitFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSignpostSplitFill']>
-    readonly BIconSim: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSim']>
-    readonly BIconSimFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSimFill']>
-    readonly BIconSkipBackward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipBackward']>
-    readonly BIconSkipBackwardBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipBackwardBtn']>
-    readonly BIconSkipBackwardBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipBackwardBtnFill']>
-    readonly BIconSkipBackwardCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipBackwardCircle']>
-    readonly BIconSkipBackwardCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipBackwardCircleFill']>
-    readonly BIconSkipBackwardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipBackwardFill']>
-    readonly BIconSkipEnd: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipEnd']>
-    readonly BIconSkipEndBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipEndBtn']>
-    readonly BIconSkipEndBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipEndBtnFill']>
-    readonly BIconSkipEndCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipEndCircle']>
-    readonly BIconSkipEndCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipEndCircleFill']>
-    readonly BIconSkipEndFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipEndFill']>
-    readonly BIconSkipForward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipForward']>
-    readonly BIconSkipForwardBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipForwardBtn']>
-    readonly BIconSkipForwardBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipForwardBtnFill']>
-    readonly BIconSkipForwardCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipForwardCircle']>
-    readonly BIconSkipForwardCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipForwardCircleFill']>
-    readonly BIconSkipForwardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipForwardFill']>
-    readonly BIconSkipStart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipStart']>
-    readonly BIconSkipStartBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipStartBtn']>
-    readonly BIconSkipStartBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipStartBtnFill']>
-    readonly BIconSkipStartCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipStartCircle']>
-    readonly BIconSkipStartCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipStartCircleFill']>
-    readonly BIconSkipStartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkipStartFill']>
-    readonly BIconSkype: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSkype']>
-    readonly BIconSlack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlack']>
-    readonly BIconSlash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlash']>
-    readonly BIconSlashCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlashCircle']>
-    readonly BIconSlashCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlashCircleFill']>
-    readonly BIconSlashLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlashLg']>
-    readonly BIconSlashSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlashSquare']>
-    readonly BIconSlashSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSlashSquareFill']>
-    readonly BIconSliders2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSliders2']>
-    readonly BIconSliders2Vertical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSliders2Vertical']>
-    readonly BIconSliders: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSliders']>
-    readonly BIconSmartwatch: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSmartwatch']>
-    readonly BIconSnapchat: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSnapchat']>
-    readonly BIconSnow2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSnow2']>
-    readonly BIconSnow3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSnow3']>
-    readonly BIconSnow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSnow']>
-    readonly BIconSortAlphaDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortAlphaDown']>
-    readonly BIconSortAlphaDownAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortAlphaDownAlt']>
-    readonly BIconSortAlphaUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortAlphaUp']>
-    readonly BIconSortAlphaUpAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortAlphaUpAlt']>
-    readonly BIconSortDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortDown']>
-    readonly BIconSortDownAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortDownAlt']>
-    readonly BIconSortNumericDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortNumericDown']>
-    readonly BIconSortNumericDownAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortNumericDownAlt']>
-    readonly BIconSortNumericUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortNumericUp']>
-    readonly BIconSortNumericUpAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortNumericUpAlt']>
-    readonly BIconSortUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortUp']>
-    readonly BIconSortUpAlt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSortUpAlt']>
-    readonly BIconSoundwave: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSoundwave']>
-    readonly BIconSpeaker: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSpeaker']>
-    readonly BIconSpeakerFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSpeakerFill']>
-    readonly BIconSpeedometer2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSpeedometer2']>
-    readonly BIconSpeedometer: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSpeedometer']>
-    readonly BIconSpellcheck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSpellcheck']>
-    readonly BIconSpotify: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSpotify']>
-    readonly BIconSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSquare']>
-    readonly BIconSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSquareFill']>
-    readonly BIconSquareHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSquareHalf']>
-    readonly BIconStack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStack']>
-    readonly BIconStackOverflow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStackOverflow']>
-    readonly BIconStar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStar']>
-    readonly BIconStarFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStarFill']>
-    readonly BIconStarHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStarHalf']>
-    readonly BIconStars: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStars']>
-    readonly BIconSteam: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSteam']>
-    readonly BIconStickies: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStickies']>
-    readonly BIconStickiesFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStickiesFill']>
-    readonly BIconSticky: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSticky']>
-    readonly BIconStickyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStickyFill']>
-    readonly BIconStop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStop']>
-    readonly BIconStopBtn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopBtn']>
-    readonly BIconStopBtnFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopBtnFill']>
-    readonly BIconStopCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopCircle']>
-    readonly BIconStopCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopCircleFill']>
-    readonly BIconStopFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopFill']>
-    readonly BIconStoplights: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStoplights']>
-    readonly BIconStoplightsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStoplightsFill']>
-    readonly BIconStopwatch: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopwatch']>
-    readonly BIconStopwatchFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStopwatchFill']>
-    readonly BIconStrava: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconStrava']>
-    readonly BIconSubtract: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSubtract']>
-    readonly BIconSuitClub: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitClub']>
-    readonly BIconSuitClubFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitClubFill']>
-    readonly BIconSuitDiamond: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitDiamond']>
-    readonly BIconSuitDiamondFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitDiamondFill']>
-    readonly BIconSuitHeart: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitHeart']>
-    readonly BIconSuitHeartFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitHeartFill']>
-    readonly BIconSuitSpade: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitSpade']>
-    readonly BIconSuitSpadeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSuitSpadeFill']>
-    readonly BIconSun: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSun']>
-    readonly BIconSunFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSunFill']>
-    readonly BIconSunglasses: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSunglasses']>
-    readonly BIconSunrise: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSunrise']>
-    readonly BIconSunriseFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSunriseFill']>
-    readonly BIconSunset: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSunset']>
-    readonly BIconSunsetFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSunsetFill']>
-    readonly BIconSymmetryHorizontal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSymmetryHorizontal']>
-    readonly BIconSymmetryVertical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconSymmetryVertical']>
-    readonly BIconTable: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTable']>
-    readonly BIconTablet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTablet']>
-    readonly BIconTabletFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTabletFill']>
-    readonly BIconTabletLandscape: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTabletLandscape']>
-    readonly BIconTabletLandscapeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTabletLandscapeFill']>
-    readonly BIconTag: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTag']>
-    readonly BIconTagFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTagFill']>
-    readonly BIconTags: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTags']>
-    readonly BIconTagsFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTagsFill']>
-    readonly BIconTelegram: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelegram']>
-    readonly BIconTelephone: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephone']>
-    readonly BIconTelephoneFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneFill']>
-    readonly BIconTelephoneForward: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneForward']>
-    readonly BIconTelephoneForwardFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneForwardFill']>
-    readonly BIconTelephoneInbound: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneInbound']>
-    readonly BIconTelephoneInboundFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneInboundFill']>
-    readonly BIconTelephoneMinus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneMinus']>
-    readonly BIconTelephoneMinusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneMinusFill']>
-    readonly BIconTelephoneOutbound: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneOutbound']>
-    readonly BIconTelephoneOutboundFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneOutboundFill']>
-    readonly BIconTelephonePlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephonePlus']>
-    readonly BIconTelephonePlusFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephonePlusFill']>
-    readonly BIconTelephoneX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneX']>
-    readonly BIconTelephoneXFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTelephoneXFill']>
-    readonly BIconTerminal: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTerminal']>
-    readonly BIconTerminalDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTerminalDash']>
-    readonly BIconTerminalFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTerminalFill']>
-    readonly BIconTerminalPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTerminalPlus']>
-    readonly BIconTerminalSplit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTerminalSplit']>
-    readonly BIconTerminalX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTerminalX']>
-    readonly BIconTextCenter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextCenter']>
-    readonly BIconTextIndentLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextIndentLeft']>
-    readonly BIconTextIndentRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextIndentRight']>
-    readonly BIconTextLeft: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextLeft']>
-    readonly BIconTextParagraph: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextParagraph']>
-    readonly BIconTextRight: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextRight']>
-    readonly BIconTextarea: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextarea']>
-    readonly BIconTextareaResize: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextareaResize']>
-    readonly BIconTextareaT: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTextareaT']>
-    readonly BIconThermometer: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThermometer']>
-    readonly BIconThermometerHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThermometerHalf']>
-    readonly BIconThermometerHigh: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThermometerHigh']>
-    readonly BIconThermometerLow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThermometerLow']>
-    readonly BIconThermometerSnow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThermometerSnow']>
-    readonly BIconThermometerSun: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThermometerSun']>
-    readonly BIconThreeDots: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThreeDots']>
-    readonly BIconThreeDotsVertical: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThreeDotsVertical']>
-    readonly BIconThunderbolt: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThunderbolt']>
-    readonly BIconThunderboltFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconThunderboltFill']>
-    readonly BIconTicket: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTicket']>
-    readonly BIconTicketDetailed: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTicketDetailed']>
-    readonly BIconTicketDetailedFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTicketDetailedFill']>
-    readonly BIconTicketFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTicketFill']>
-    readonly BIconTicketPerforated: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTicketPerforated']>
-    readonly BIconTicketPerforatedFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTicketPerforatedFill']>
-    readonly BIconTiktok: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTiktok']>
-    readonly BIconToggle2Off: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconToggle2Off']>
-    readonly BIconToggle2On: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconToggle2On']>
-    readonly BIconToggleOff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconToggleOff']>
-    readonly BIconToggleOn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconToggleOn']>
-    readonly BIconToggles2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconToggles2']>
-    readonly BIconToggles: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconToggles']>
-    readonly BIconTools: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTools']>
-    readonly BIconTornado: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTornado']>
-    readonly BIconTranslate: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTranslate']>
-    readonly BIconTrash2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrash2']>
-    readonly BIconTrash2Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrash2Fill']>
-    readonly BIconTrash3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrash3']>
-    readonly BIconTrash3Fill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrash3Fill']>
-    readonly BIconTrash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrash']>
-    readonly BIconTrashFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrashFill']>
-    readonly BIconTree: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTree']>
-    readonly BIconTreeFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTreeFill']>
-    readonly BIconTriangle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTriangle']>
-    readonly BIconTriangleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTriangleFill']>
-    readonly BIconTriangleHalf: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTriangleHalf']>
-    readonly BIconTrophy: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrophy']>
-    readonly BIconTrophyFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTrophyFill']>
-    readonly BIconTropicalStorm: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTropicalStorm']>
-    readonly BIconTruck: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTruck']>
-    readonly BIconTruckFlatbed: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTruckFlatbed']>
-    readonly BIconTsunami: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTsunami']>
-    readonly BIconTv: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTv']>
-    readonly BIconTvFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTvFill']>
-    readonly BIconTwitch: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTwitch']>
-    readonly BIconTwitter: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTwitter']>
-    readonly BIconType: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconType']>
-    readonly BIconTypeBold: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeBold']>
-    readonly BIconTypeH1: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeH1']>
-    readonly BIconTypeH2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeH2']>
-    readonly BIconTypeH3: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeH3']>
-    readonly BIconTypeItalic: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeItalic']>
-    readonly BIconTypeStrikethrough: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeStrikethrough']>
-    readonly BIconTypeUnderline: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconTypeUnderline']>
-    readonly BIconUiChecks: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUiChecks']>
-    readonly BIconUiChecksGrid: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUiChecksGrid']>
-    readonly BIconUiRadios: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUiRadios']>
-    readonly BIconUiRadiosGrid: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUiRadiosGrid']>
-    readonly BIconUmbrella: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUmbrella']>
-    readonly BIconUmbrellaFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUmbrellaFill']>
-    readonly BIconUnion: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUnion']>
-    readonly BIconUnlock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUnlock']>
-    readonly BIconUnlockFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUnlockFill']>
-    readonly BIconUpc: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUpc']>
-    readonly BIconUpcScan: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUpcScan']>
-    readonly BIconUpload: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUpload']>
-    readonly BIconUsb: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsb']>
-    readonly BIconUsbC: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbC']>
-    readonly BIconUsbCFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbCFill']>
-    readonly BIconUsbDrive: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbDrive']>
-    readonly BIconUsbDriveFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbDriveFill']>
-    readonly BIconUsbFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbFill']>
-    readonly BIconUsbMicro: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbMicro']>
-    readonly BIconUsbMicroFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbMicroFill']>
-    readonly BIconUsbMini: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbMini']>
-    readonly BIconUsbMiniFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbMiniFill']>
-    readonly BIconUsbPlug: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbPlug']>
-    readonly BIconUsbPlugFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbPlugFill']>
-    readonly BIconUsbSymbol: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconUsbSymbol']>
-    readonly BIconValentine2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconValentine2']>
-    readonly BIconValentine: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconValentine']>
-    readonly BIconVectorPen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVectorPen']>
-    readonly BIconViewList: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconViewList']>
-    readonly BIconViewStacked: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconViewStacked']>
-    readonly BIconVimeo: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVimeo']>
-    readonly BIconVinyl: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVinyl']>
-    readonly BIconVinylFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVinylFill']>
-    readonly BIconVoicemail: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVoicemail']>
-    readonly BIconVolumeDown: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeDown']>
-    readonly BIconVolumeDownFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeDownFill']>
-    readonly BIconVolumeMute: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeMute']>
-    readonly BIconVolumeMuteFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeMuteFill']>
-    readonly BIconVolumeOff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeOff']>
-    readonly BIconVolumeOffFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeOffFill']>
-    readonly BIconVolumeUp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeUp']>
-    readonly BIconVolumeUpFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVolumeUpFill']>
-    readonly BIconVr: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconVr']>
-    readonly BIconWallet2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWallet2']>
-    readonly BIconWallet: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWallet']>
-    readonly BIconWalletFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWalletFill']>
-    readonly BIconWatch: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWatch']>
-    readonly BIconWater: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWater']>
-    readonly BIconWebcam: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWebcam']>
-    readonly BIconWebcamFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWebcamFill']>
-    readonly BIconWhatsapp: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWhatsapp']>
-    readonly BIconWifi1: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWifi1']>
-    readonly BIconWifi2: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWifi2']>
-    readonly BIconWifi: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWifi']>
-    readonly BIconWifiOff: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWifiOff']>
-    readonly BIconWind: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWind']>
-    readonly BIconWindow: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindow']>
-    readonly BIconWindowDash: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowDash']>
-    readonly BIconWindowDesktop: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowDesktop']>
-    readonly BIconWindowDock: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowDock']>
-    readonly BIconWindowFullscreen: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowFullscreen']>
-    readonly BIconWindowPlus: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowPlus']>
-    readonly BIconWindowSidebar: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowSidebar']>
-    readonly BIconWindowSplit: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowSplit']>
-    readonly BIconWindowStack: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowStack']>
-    readonly BIconWindowX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindowX']>
-    readonly BIconWindows: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWindows']>
-    readonly BIconWordpress: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWordpress']>
-    readonly BIconWrench: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWrench']>
-    readonly BIconWrenchAdjustable: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWrenchAdjustable']>
-    readonly BIconWrenchAdjustableCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWrenchAdjustableCircle']>
-    readonly BIconWrenchAdjustableCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconWrenchAdjustableCircleFill']>
-    readonly BIconX: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconX']>
-    readonly BIconXCircle: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXCircle']>
-    readonly BIconXCircleFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXCircleFill']>
-    readonly BIconXDiamond: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXDiamond']>
-    readonly BIconXDiamondFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXDiamondFill']>
-    readonly BIconXLg: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXLg']>
-    readonly BIconXOctagon: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXOctagon']>
-    readonly BIconXOctagonFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXOctagonFill']>
-    readonly BIconXSquare: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXSquare']>
-    readonly BIconXSquareFill: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXSquareFill']>
-    readonly BIconXbox: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconXbox']>
-    readonly BIconYinYang: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconYinYang']>
-    readonly BIconYoutube: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconYoutube']>
-    readonly BIconZoomIn: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconZoomIn']>
-    readonly BIconZoomOut: UnwrapRef<typeof import('bootstrap-icons-vue')['BIconZoomOut']>
-    readonly EffectScope: UnwrapRef<typeof import('vue')['EffectScope']>
-    readonly GQL_URL: UnwrapRef<typeof import('./shared/constants/url-config')['GQL_URL']>
-    readonly GQL_WS_URL: UnwrapRef<typeof import('./shared/constants/url-config')['GQL_WS_URL']>
-    readonly RouteName: UnwrapRef<typeof import('./shared/constants/route-name')['RouteName']>
-    readonly USER: UnwrapRef<typeof import('./shared/dto/user')['USER']>
-    readonly USER_URL: UnwrapRef<typeof import('./shared/dto/user-url')['USER_URL']>
-    readonly afterAll: UnwrapRef<typeof import('vitest')['afterAll']>
-    readonly afterEach: UnwrapRef<typeof import('vitest')['afterEach']>
-    readonly apolloClient: UnwrapRef<typeof import('./core/utils/apollo')['apolloClient']>
-    readonly assert: UnwrapRef<typeof import('vitest')['assert']>
-    readonly asyncComputed: UnwrapRef<typeof import('@vueuse/core')['asyncComputed']>
-    readonly autoResetRef: UnwrapRef<typeof import('@vueuse/core')['autoResetRef']>
-    readonly axiosApi: UnwrapRef<typeof import('./core/utils/axios')['axiosApi']>
-    readonly axiosAuth: UnwrapRef<typeof import('./core/utils/axios')['axiosAuth']>
-    readonly beforeAll: UnwrapRef<typeof import('vitest')['beforeAll']>
-    readonly beforeEach: UnwrapRef<typeof import('vitest')['beforeEach']>
-    readonly chai: UnwrapRef<typeof import('vitest')['chai']>
-    readonly computed: UnwrapRef<typeof import('vue')['computed']>
-    readonly computedAsync: UnwrapRef<typeof import('@vueuse/core')['computedAsync']>
-    readonly computedEager: UnwrapRef<typeof import('@vueuse/core')['computedEager']>
-    readonly computedInject: UnwrapRef<typeof import('@vueuse/core')['computedInject']>
-    readonly computedWithControl: UnwrapRef<typeof import('@vueuse/core')['computedWithControl']>
-    readonly controlledComputed: UnwrapRef<typeof import('@vueuse/core')['controlledComputed']>
-    readonly controlledRef: UnwrapRef<typeof import('@vueuse/core')['controlledRef']>
-    readonly createApp: UnwrapRef<typeof import('vue')['createApp']>
-    readonly createEventHook: UnwrapRef<typeof import('@vueuse/core')['createEventHook']>
-    readonly createGlobalState: UnwrapRef<typeof import('@vueuse/core')['createGlobalState']>
-    readonly createInjectionState: UnwrapRef<typeof import('@vueuse/core')['createInjectionState']>
-    readonly createReactiveFn: UnwrapRef<typeof import('@vueuse/core')['createReactiveFn']>
-    readonly createSharedComposable: UnwrapRef<typeof import('@vueuse/core')['createSharedComposable']>
-    readonly createUnrefFn: UnwrapRef<typeof import('@vueuse/core')['createUnrefFn']>
-    readonly customRef: UnwrapRef<typeof import('vue')['customRef']>
-    readonly debouncedRef: UnwrapRef<typeof import('@vueuse/core')['debouncedRef']>
-    readonly debouncedWatch: UnwrapRef<typeof import('@vueuse/core')['debouncedWatch']>
-    readonly defineAsyncComponent: UnwrapRef<typeof import('vue')['defineAsyncComponent']>
-    readonly defineComponent: UnwrapRef<typeof import('vue')['defineComponent']>
-    readonly describe: UnwrapRef<typeof import('vitest')['describe']>
-    readonly devicesRoute: UnwrapRef<typeof import('./devices/router/index')['devicesRoute']>
-    readonly eagerComputed: UnwrapRef<typeof import('@vueuse/core')['eagerComputed']>
-    readonly effectScope: UnwrapRef<typeof import('vue')['effectScope']>
-    readonly expect: UnwrapRef<typeof import('vitest')['expect']>
-    readonly extendRef: UnwrapRef<typeof import('@vueuse/core')['extendRef']>
-    readonly friendsRoute: UnwrapRef<typeof import('./friends/router/index')['friendsRoute']>
-    readonly getCurrentInstance: UnwrapRef<typeof import('vue')['getCurrentInstance']>
-    readonly getCurrentScope: UnwrapRef<typeof import('vue')['getCurrentScope']>
-    readonly gql: UnwrapRef<typeof import('graphql-tag')['default']>
-    readonly h: UnwrapRef<typeof import('vue')['h']>
-    readonly homeRoute: UnwrapRef<typeof import('./home/router/index')['homeRoute']>
-    readonly ignorableWatch: UnwrapRef<typeof import('@vueuse/core')['ignorableWatch']>
-    readonly inject: UnwrapRef<typeof import('vue')['inject']>
-    readonly isDefined: UnwrapRef<typeof import('@vueuse/core')['isDefined']>
-    readonly isProxy: UnwrapRef<typeof import('vue')['isProxy']>
-    readonly isReactive: UnwrapRef<typeof import('vue')['isReactive']>
-    readonly isReadonly: UnwrapRef<typeof import('vue')['isReadonly']>
-    readonly isRef: UnwrapRef<typeof import('vue')['isRef']>
-    readonly it: UnwrapRef<typeof import('vitest')['it']>
-    readonly keys: UnwrapRef<typeof import('./shared/utils/object')['keys']>
-    readonly libraryRoute: UnwrapRef<typeof import('./library/router/index')['libraryRoute']>
-    readonly makeDestructurable: UnwrapRef<typeof import('@vueuse/core')['makeDestructurable']>
-    readonly map: UnwrapRef<typeof import('./shared/utils/object')['map']>
-    readonly markRaw: UnwrapRef<typeof import('vue')['markRaw']>
-    readonly nextTick: UnwrapRef<typeof import('vue')['nextTick']>
-    readonly omit: UnwrapRef<typeof import('./shared/utils/object')['omit']>
-    readonly onActivated: UnwrapRef<typeof import('vue')['onActivated']>
-    readonly onBeforeMount: UnwrapRef<typeof import('vue')['onBeforeMount']>
-    readonly onBeforeRouteLeave: UnwrapRef<typeof import('vue-router')['onBeforeRouteLeave']>
-    readonly onBeforeRouteUpdate: UnwrapRef<typeof import('vue-router')['onBeforeRouteUpdate']>
-    readonly onBeforeUnmount: UnwrapRef<typeof import('vue')['onBeforeUnmount']>
-    readonly onBeforeUpdate: UnwrapRef<typeof import('vue')['onBeforeUpdate']>
-    readonly onClickOutside: UnwrapRef<typeof import('@vueuse/core')['onClickOutside']>
-    readonly onDeactivated: UnwrapRef<typeof import('vue')['onDeactivated']>
-    readonly onErrorCaptured: UnwrapRef<typeof import('vue')['onErrorCaptured']>
-    readonly onKeyStroke: UnwrapRef<typeof import('@vueuse/core')['onKeyStroke']>
-    readonly onLongPress: UnwrapRef<typeof import('@vueuse/core')['onLongPress']>
-    readonly onMounted: UnwrapRef<typeof import('vue')['onMounted']>
-    readonly onRenderTracked: UnwrapRef<typeof import('vue')['onRenderTracked']>
-    readonly onRenderTriggered: UnwrapRef<typeof import('vue')['onRenderTriggered']>
-    readonly onScopeDispose: UnwrapRef<typeof import('vue')['onScopeDispose']>
-    readonly onServerPrefetch: UnwrapRef<typeof import('vue')['onServerPrefetch']>
-    readonly onStartTyping: UnwrapRef<typeof import('@vueuse/core')['onStartTyping']>
-    readonly onUnmounted: UnwrapRef<typeof import('vue')['onUnmounted']>
-    readonly onUpdated: UnwrapRef<typeof import('vue')['onUpdated']>
-    readonly pausableWatch: UnwrapRef<typeof import('@vueuse/core')['pausableWatch']>
-    readonly pick: UnwrapRef<typeof import('./shared/utils/object')['pick']>
-    readonly playlistsRoute: UnwrapRef<typeof import('./playlists/router/index')['playlistsRoute']>
-    readonly provide: UnwrapRef<typeof import('vue')['provide']>
-    readonly reactify: UnwrapRef<typeof import('@vueuse/core')['reactify']>
-    readonly reactifyObject: UnwrapRef<typeof import('@vueuse/core')['reactifyObject']>
-    readonly reactive: UnwrapRef<typeof import('vue')['reactive']>
-    readonly reactiveComputed: UnwrapRef<typeof import('@vueuse/core')['reactiveComputed']>
-    readonly reactiveOmit: UnwrapRef<typeof import('@vueuse/core')['reactiveOmit']>
-    readonly reactivePick: UnwrapRef<typeof import('@vueuse/core')['reactivePick']>
-    readonly readonly: UnwrapRef<typeof import('vue')['readonly']>
-    readonly ref: UnwrapRef<typeof import('vue')['ref']>
-    readonly refAutoReset: UnwrapRef<typeof import('@vueuse/core')['refAutoReset']>
-    readonly refDebounced: UnwrapRef<typeof import('@vueuse/core')['refDebounced']>
-    readonly refDefault: UnwrapRef<typeof import('@vueuse/core')['refDefault']>
-    readonly refThrottled: UnwrapRef<typeof import('@vueuse/core')['refThrottled']>
-    readonly refWithControl: UnwrapRef<typeof import('@vueuse/core')['refWithControl']>
-    readonly resolveComponent: UnwrapRef<typeof import('vue')['resolveComponent']>
-    readonly resolveDirective: UnwrapRef<typeof import('vue')['resolveDirective']>
-    readonly resolveRef: UnwrapRef<typeof import('@vueuse/core')['resolveRef']>
-    readonly resolveUnref: UnwrapRef<typeof import('@vueuse/core')['resolveUnref']>
-    readonly serializer: UnwrapRef<typeof import('./shared/utils/serializer')['serializer']>
-    readonly settingsRoute: UnwrapRef<typeof import('./settings/router/index')['settingsRoute']>
-    readonly shallowReactive: UnwrapRef<typeof import('vue')['shallowReactive']>
-    readonly shallowReadonly: UnwrapRef<typeof import('vue')['shallowReadonly']>
-    readonly shallowRef: UnwrapRef<typeof import('vue')['shallowRef']>
-    readonly suite: UnwrapRef<typeof import('vitest')['suite']>
-    readonly switchAssign: UnwrapRef<typeof import('./shared/utils/switch')['switchAssign']>
-    readonly switchExec: UnwrapRef<typeof import('./shared/utils/switch')['switchExec']>
-    readonly switchThrow: UnwrapRef<typeof import('./shared/utils/switch')['switchThrow']>
-    readonly syncRef: UnwrapRef<typeof import('@vueuse/core')['syncRef']>
-    readonly syncRefs: UnwrapRef<typeof import('@vueuse/core')['syncRefs']>
-    readonly templateRef: UnwrapRef<typeof import('@vueuse/core')['templateRef']>
-    readonly test: UnwrapRef<typeof import('vitest')['test']>
-    readonly themesRoute: UnwrapRef<typeof import('./themes/router/index')['themesRoute']>
-    readonly throttledRef: UnwrapRef<typeof import('@vueuse/core')['throttledRef']>
-    readonly throttledWatch: UnwrapRef<typeof import('@vueuse/core')['throttledWatch']>
-    readonly toRaw: UnwrapRef<typeof import('vue')['toRaw']>
-    readonly toReactive: UnwrapRef<typeof import('@vueuse/core')['toReactive']>
-    readonly toRef: UnwrapRef<typeof import('vue')['toRef']>
-    readonly toRefs: UnwrapRef<typeof import('vue')['toRefs']>
-    readonly triggerRef: UnwrapRef<typeof import('vue')['triggerRef']>
-    readonly tryOnBeforeMount: UnwrapRef<typeof import('@vueuse/core')['tryOnBeforeMount']>
-    readonly tryOnBeforeUnmount: UnwrapRef<typeof import('@vueuse/core')['tryOnBeforeUnmount']>
-    readonly tryOnMounted: UnwrapRef<typeof import('@vueuse/core')['tryOnMounted']>
-    readonly tryOnScopeDispose: UnwrapRef<typeof import('@vueuse/core')['tryOnScopeDispose']>
-    readonly tryOnUnmounted: UnwrapRef<typeof import('@vueuse/core')['tryOnUnmounted']>
-    readonly unref: UnwrapRef<typeof import('vue')['unref']>
-    readonly unrefElement: UnwrapRef<typeof import('@vueuse/core')['unrefElement']>
-    readonly until: UnwrapRef<typeof import('@vueuse/core')['until']>
-    readonly useActiveElement: UnwrapRef<typeof import('@vueuse/core')['useActiveElement']>
-    readonly useArrayEvery: UnwrapRef<typeof import('@vueuse/core')['useArrayEvery']>
-    readonly useArrayFilter: UnwrapRef<typeof import('@vueuse/core')['useArrayFilter']>
-    readonly useArrayFind: UnwrapRef<typeof import('@vueuse/core')['useArrayFind']>
-    readonly useArrayFindIndex: UnwrapRef<typeof import('@vueuse/core')['useArrayFindIndex']>
-    readonly useArrayJoin: UnwrapRef<typeof import('@vueuse/core')['useArrayJoin']>
-    readonly useArrayMap: UnwrapRef<typeof import('@vueuse/core')['useArrayMap']>
-    readonly useArrayReduce: UnwrapRef<typeof import('@vueuse/core')['useArrayReduce']>
-    readonly useArraySome: UnwrapRef<typeof import('@vueuse/core')['useArraySome']>
-    readonly useArrayUnique: UnwrapRef<typeof import('@vueuse/core')['useArrayUnique']>
-    readonly useAsyncQueue: UnwrapRef<typeof import('@vueuse/core')['useAsyncQueue']>
-    readonly useAsyncState: UnwrapRef<typeof import('@vueuse/core')['useAsyncState']>
-    readonly useAttrs: UnwrapRef<typeof import('vue')['useAttrs']>
-    readonly useAuthentication: UnwrapRef<typeof import('./auth/stores/authentication')['useAuthentication']>
-    readonly useAuthenticationService: UnwrapRef<typeof import('./auth/services/authentication')['useAuthenticationService']>
-    readonly useBase64: UnwrapRef<typeof import('@vueuse/core')['useBase64']>
-    readonly useBattery: UnwrapRef<typeof import('@vueuse/core')['useBattery']>
-    readonly useBluetooth: UnwrapRef<typeof import('@vueuse/core')['useBluetooth']>
-    readonly useBreakpoints: UnwrapRef<typeof import('@vueuse/core')['useBreakpoints']>
-    readonly useBroadcastChannel: UnwrapRef<typeof import('@vueuse/core')['useBroadcastChannel']>
-    readonly useBrowserLocation: UnwrapRef<typeof import('@vueuse/core')['useBrowserLocation']>
-    readonly useCached: UnwrapRef<typeof import('@vueuse/core')['useCached']>
-    readonly useClipboard: UnwrapRef<typeof import('@vueuse/core')['useClipboard']>
-    readonly useCloned: UnwrapRef<typeof import('@vueuse/core')['useCloned']>
-    readonly useColorMode: UnwrapRef<typeof import('@vueuse/core')['useColorMode']>
-    readonly useConfirmDialog: UnwrapRef<typeof import('@vueuse/core')['useConfirmDialog']>
-    readonly useCounter: UnwrapRef<typeof import('@vueuse/core')['useCounter']>
-    readonly useCssModule: UnwrapRef<typeof import('vue')['useCssModule']>
-    readonly useCssVar: UnwrapRef<typeof import('@vueuse/core')['useCssVar']>
-    readonly useCssVars: UnwrapRef<typeof import('vue')['useCssVars']>
-    readonly useCurrentElement: UnwrapRef<typeof import('@vueuse/core')['useCurrentElement']>
-    readonly useCycleList: UnwrapRef<typeof import('@vueuse/core')['useCycleList']>
-    readonly useDark: UnwrapRef<typeof import('@vueuse/core')['useDark']>
-    readonly useDateFormat: UnwrapRef<typeof import('@vueuse/core')['useDateFormat']>
-    readonly useDebounce: UnwrapRef<typeof import('@vueuse/core')['useDebounce']>
-    readonly useDebounceFn: UnwrapRef<typeof import('@vueuse/core')['useDebounceFn']>
-    readonly useDebouncedRefHistory: UnwrapRef<typeof import('@vueuse/core')['useDebouncedRefHistory']>
-    readonly useDeviceMotion: UnwrapRef<typeof import('@vueuse/core')['useDeviceMotion']>
-    readonly useDeviceOrientation: UnwrapRef<typeof import('@vueuse/core')['useDeviceOrientation']>
-    readonly useDevicePixelRatio: UnwrapRef<typeof import('@vueuse/core')['useDevicePixelRatio']>
-    readonly useDevicesList: UnwrapRef<typeof import('@vueuse/core')['useDevicesList']>
-    readonly useDisplayMedia: UnwrapRef<typeof import('@vueuse/core')['useDisplayMedia']>
-    readonly useDocumentVisibility: UnwrapRef<typeof import('@vueuse/core')['useDocumentVisibility']>
-    readonly useDraggable: UnwrapRef<typeof import('@vueuse/core')['useDraggable']>
-    readonly useDropZone: UnwrapRef<typeof import('@vueuse/core')['useDropZone']>
-    readonly useElementBounding: UnwrapRef<typeof import('@vueuse/core')['useElementBounding']>
-    readonly useElementByPoint: UnwrapRef<typeof import('@vueuse/core')['useElementByPoint']>
-    readonly useElementHover: UnwrapRef<typeof import('@vueuse/core')['useElementHover']>
-    readonly useElementSize: UnwrapRef<typeof import('@vueuse/core')['useElementSize']>
-    readonly useElementVisibility: UnwrapRef<typeof import('@vueuse/core')['useElementVisibility']>
-    readonly useEventBus: UnwrapRef<typeof import('@vueuse/core')['useEventBus']>
-    readonly useEventListener: UnwrapRef<typeof import('@vueuse/core')['useEventListener']>
-    readonly useEventSource: UnwrapRef<typeof import('@vueuse/core')['useEventSource']>
-    readonly useEyeDropper: UnwrapRef<typeof import('@vueuse/core')['useEyeDropper']>
-    readonly useFavicon: UnwrapRef<typeof import('@vueuse/core')['useFavicon']>
-    readonly useFetch: UnwrapRef<typeof import('@vueuse/core')['useFetch']>
-    readonly useFileDialog: UnwrapRef<typeof import('@vueuse/core')['useFileDialog']>
-    readonly useFileSystemAccess: UnwrapRef<typeof import('@vueuse/core')['useFileSystemAccess']>
-    readonly useFocus: UnwrapRef<typeof import('@vueuse/core')['useFocus']>
-    readonly useFocusWithin: UnwrapRef<typeof import('@vueuse/core')['useFocusWithin']>
-    readonly useFps: UnwrapRef<typeof import('@vueuse/core')['useFps']>
-    readonly useFullscreen: UnwrapRef<typeof import('@vueuse/core')['useFullscreen']>
-    readonly useGamepad: UnwrapRef<typeof import('@vueuse/core')['useGamepad']>
-    readonly useGeolocation: UnwrapRef<typeof import('@vueuse/core')['useGeolocation']>
-    readonly useIdle: UnwrapRef<typeof import('@vueuse/core')['useIdle']>
-    readonly useImage: UnwrapRef<typeof import('@vueuse/core')['useImage']>
-    readonly useInfiniteScroll: UnwrapRef<typeof import('@vueuse/core')['useInfiniteScroll']>
-    readonly useIntersectionObserver: UnwrapRef<typeof import('@vueuse/core')['useIntersectionObserver']>
-    readonly useInterval: UnwrapRef<typeof import('@vueuse/core')['useInterval']>
-    readonly useIntervalFn: UnwrapRef<typeof import('@vueuse/core')['useIntervalFn']>
-    readonly useKeyModifier: UnwrapRef<typeof import('@vueuse/core')['useKeyModifier']>
-    readonly useLastChanged: UnwrapRef<typeof import('@vueuse/core')['useLastChanged']>
-    readonly useLink: UnwrapRef<typeof import('vue-router')['useLink']>
-    readonly useLocalStorage: UnwrapRef<typeof import('@vueuse/core')['useLocalStorage']>
-    readonly useMagicKeys: UnwrapRef<typeof import('@vueuse/core')['useMagicKeys']>
-    readonly useManualRefHistory: UnwrapRef<typeof import('@vueuse/core')['useManualRefHistory']>
-    readonly useMediaControls: UnwrapRef<typeof import('@vueuse/core')['useMediaControls']>
-    readonly useMediaQuery: UnwrapRef<typeof import('@vueuse/core')['useMediaQuery']>
-    readonly useMemoize: UnwrapRef<typeof import('@vueuse/core')['useMemoize']>
-    readonly useMemory: UnwrapRef<typeof import('@vueuse/core')['useMemory']>
-    readonly useMobile: UnwrapRef<typeof import('./shared/hooks/mobile')['useMobile']>
-    readonly useMounted: UnwrapRef<typeof import('@vueuse/core')['useMounted']>
-    readonly useMouse: UnwrapRef<typeof import('@vueuse/core')['useMouse']>
-    readonly useMouseInElement: UnwrapRef<typeof import('@vueuse/core')['useMouseInElement']>
-    readonly useMousePressed: UnwrapRef<typeof import('@vueuse/core')['useMousePressed']>
-    readonly useMutation: UnwrapRef<typeof import('@vue/apollo-composable')['useMutation']>
-    readonly useMutationObserver: UnwrapRef<typeof import('@vueuse/core')['useMutationObserver']>
-    readonly useNavigatorLanguage: UnwrapRef<typeof import('@vueuse/core')['useNavigatorLanguage']>
-    readonly useNetwork: UnwrapRef<typeof import('@vueuse/core')['useNetwork']>
-    readonly useNow: UnwrapRef<typeof import('@vueuse/core')['useNow']>
-    readonly useObjectUrl: UnwrapRef<typeof import('@vueuse/core')['useObjectUrl']>
-    readonly useOffsetPagination: UnwrapRef<typeof import('@vueuse/core')['useOffsetPagination']>
-    readonly useOnline: UnwrapRef<typeof import('@vueuse/core')['useOnline']>
-    readonly usePageLeave: UnwrapRef<typeof import('@vueuse/core')['usePageLeave']>
-    readonly useParallax: UnwrapRef<typeof import('@vueuse/core')['useParallax']>
-    readonly usePermission: UnwrapRef<typeof import('@vueuse/core')['usePermission']>
-    readonly usePointer: UnwrapRef<typeof import('@vueuse/core')['usePointer']>
-    readonly usePointerSwipe: UnwrapRef<typeof import('@vueuse/core')['usePointerSwipe']>
-    readonly usePreferredColorScheme: UnwrapRef<typeof import('@vueuse/core')['usePreferredColorScheme']>
-    readonly usePreferredContrast: UnwrapRef<typeof import('@vueuse/core')['usePreferredContrast']>
-    readonly usePreferredDark: UnwrapRef<typeof import('@vueuse/core')['usePreferredDark']>
-    readonly usePreferredLanguages: UnwrapRef<typeof import('@vueuse/core')['usePreferredLanguages']>
-    readonly usePreferredReducedMotion: UnwrapRef<typeof import('@vueuse/core')['usePreferredReducedMotion']>
-    readonly useQuery: UnwrapRef<typeof import('@vue/apollo-composable')['useQuery']>
-    readonly useRafFn: UnwrapRef<typeof import('@vueuse/core')['useRafFn']>
-    readonly useRefHistory: UnwrapRef<typeof import('@vueuse/core')['useRefHistory']>
-    readonly useResizeObserver: UnwrapRef<typeof import('@vueuse/core')['useResizeObserver']>
-    readonly useRoute: UnwrapRef<typeof import('vue-router')['useRoute']>
-    readonly useRouter: UnwrapRef<typeof import('vue-router')['useRouter']>
-    readonly useScreenOrientation: UnwrapRef<typeof import('@vueuse/core')['useScreenOrientation']>
-    readonly useScreenSafeArea: UnwrapRef<typeof import('@vueuse/core')['useScreenSafeArea']>
-    readonly useScriptTag: UnwrapRef<typeof import('@vueuse/core')['useScriptTag']>
-    readonly useScroll: UnwrapRef<typeof import('@vueuse/core')['useScroll']>
-    readonly useScrollLock: UnwrapRef<typeof import('@vueuse/core')['useScrollLock']>
-    readonly useSessionStorage: UnwrapRef<typeof import('@vueuse/core')['useSessionStorage']>
-    readonly useShare: UnwrapRef<typeof import('@vueuse/core')['useShare']>
-    readonly useSlots: UnwrapRef<typeof import('vue')['useSlots']>
-    readonly useSorted: UnwrapRef<typeof import('@vueuse/core')['useSorted']>
-    readonly useSpeechRecognition: UnwrapRef<typeof import('@vueuse/core')['useSpeechRecognition']>
-    readonly useSpeechSynthesis: UnwrapRef<typeof import('@vueuse/core')['useSpeechSynthesis']>
-    readonly useStepper: UnwrapRef<typeof import('@vueuse/core')['useStepper']>
-    readonly useStorage: UnwrapRef<typeof import('@vueuse/core')['useStorage']>
-    readonly useStorageAsync: UnwrapRef<typeof import('@vueuse/core')['useStorageAsync']>
-    readonly useStyleTag: UnwrapRef<typeof import('@vueuse/core')['useStyleTag']>
-    readonly useSubscription: UnwrapRef<typeof import('@vue/apollo-composable')['useSubscription']>
-    readonly useSupported: UnwrapRef<typeof import('@vueuse/core')['useSupported']>
-    readonly useSwipe: UnwrapRef<typeof import('@vueuse/core')['useSwipe']>
-    readonly useTemplateRefsList: UnwrapRef<typeof import('@vueuse/core')['useTemplateRefsList']>
-    readonly useTextDirection: UnwrapRef<typeof import('@vueuse/core')['useTextDirection']>
-    readonly useTextSelection: UnwrapRef<typeof import('@vueuse/core')['useTextSelection']>
-    readonly useTextareaAutosize: UnwrapRef<typeof import('@vueuse/core')['useTextareaAutosize']>
-    readonly useThrottle: UnwrapRef<typeof import('@vueuse/core')['useThrottle']>
-    readonly useThrottleFn: UnwrapRef<typeof import('@vueuse/core')['useThrottleFn']>
-    readonly useThrottledRefHistory: UnwrapRef<typeof import('@vueuse/core')['useThrottledRefHistory']>
-    readonly useTimeAgo: UnwrapRef<typeof import('@vueuse/core')['useTimeAgo']>
-    readonly useTimeout: UnwrapRef<typeof import('@vueuse/core')['useTimeout']>
-    readonly useTimeoutFn: UnwrapRef<typeof import('@vueuse/core')['useTimeoutFn']>
-    readonly useTimeoutPoll: UnwrapRef<typeof import('@vueuse/core')['useTimeoutPoll']>
-    readonly useTimestamp: UnwrapRef<typeof import('@vueuse/core')['useTimestamp']>
-    readonly useTitle: UnwrapRef<typeof import('@vueuse/core')['useTitle']>
-    readonly useToNumber: UnwrapRef<typeof import('@vueuse/core')['useToNumber']>
-    readonly useToString: UnwrapRef<typeof import('@vueuse/core')['useToString']>
-    readonly useToggle: UnwrapRef<typeof import('@vueuse/core')['useToggle']>
-    readonly useTransition: UnwrapRef<typeof import('@vueuse/core')['useTransition']>
-    readonly useUrlSearchParams: UnwrapRef<typeof import('@vueuse/core')['useUrlSearchParams']>
-    readonly useUser: UnwrapRef<typeof import('./core/stores/user')['useUser']>
-    readonly useUserMedia: UnwrapRef<typeof import('@vueuse/core')['useUserMedia']>
-    readonly useUsersService: UnwrapRef<typeof import('./shared/services/users')['useUsersService']>
-    readonly useVModel: UnwrapRef<typeof import('@vueuse/core')['useVModel']>
-    readonly useVModels: UnwrapRef<typeof import('@vueuse/core')['useVModels']>
-    readonly useVibrate: UnwrapRef<typeof import('@vueuse/core')['useVibrate']>
-    readonly useVirtualList: UnwrapRef<typeof import('@vueuse/core')['useVirtualList']>
-    readonly useWakeLock: UnwrapRef<typeof import('@vueuse/core')['useWakeLock']>
-    readonly useWebNotification: UnwrapRef<typeof import('@vueuse/core')['useWebNotification']>
-    readonly useWebSocket: UnwrapRef<typeof import('@vueuse/core')['useWebSocket']>
-    readonly useWebWorker: UnwrapRef<typeof import('@vueuse/core')['useWebWorker']>
-    readonly useWebWorkerFn: UnwrapRef<typeof import('@vueuse/core')['useWebWorkerFn']>
-    readonly useWindowFocus: UnwrapRef<typeof import('@vueuse/core')['useWindowFocus']>
-    readonly useWindowScroll: UnwrapRef<typeof import('@vueuse/core')['useWindowScroll']>
-    readonly useWindowSize: UnwrapRef<typeof import('@vueuse/core')['useWindowSize']>
-    readonly vi: UnwrapRef<typeof import('vitest')['vi']>
-    readonly vitest: UnwrapRef<typeof import('vitest')['vitest']>
-    readonly watch: UnwrapRef<typeof import('vue')['watch']>
-    readonly watchArray: UnwrapRef<typeof import('@vueuse/core')['watchArray']>
-    readonly watchAtMost: UnwrapRef<typeof import('@vueuse/core')['watchAtMost']>
-    readonly watchDebounced: UnwrapRef<typeof import('@vueuse/core')['watchDebounced']>
-    readonly watchEffect: UnwrapRef<typeof import('vue')['watchEffect']>
-    readonly watchIgnorable: UnwrapRef<typeof import('@vueuse/core')['watchIgnorable']>
-    readonly watchOnce: UnwrapRef<typeof import('@vueuse/core')['watchOnce']>
-    readonly watchPausable: UnwrapRef<typeof import('@vueuse/core')['watchPausable']>
-    readonly watchPostEffect: UnwrapRef<typeof import('vue')['watchPostEffect']>
-    readonly watchSyncEffect: UnwrapRef<typeof import('vue')['watchSyncEffect']>
-    readonly watchThrottled: UnwrapRef<typeof import('@vueuse/core')['watchThrottled']>
-    readonly watchTriggerable: UnwrapRef<typeof import('@vueuse/core')['watchTriggerable']>
-    readonly watchWithFilter: UnwrapRef<typeof import('@vueuse/core')['watchWithFilter']>
-    readonly whenever: UnwrapRef<typeof import('@vueuse/core')['whenever']>
-  }
-}
