--- src/components.d.ts
@@ -1,33 +1,0 @@
-// generated by unplugin-vue-components
-// We suggest you to commit this file into source control
-// Read more: https://github.com/vuejs/core/pull/3399
-import '@vue/runtime-core'
-
-export {}
-
-declare module '@vue/runtime-core' {
-  export interface GlobalComponents {
-    BIconArrowRepeat: typeof import('bootstrap-icons-vue')['BIconArrowRepeat']
-    BIconCircleFill: typeof import('bootstrap-icons-vue')['BIconCircleFill']
-    BIconMusicNoteBeamed: typeof import('bootstrap-icons-vue')['BIconMusicNoteBeamed']
-    BIconPlayCircleFill: typeof import('bootstrap-icons-vue')['BIconPlayCircleFill']
-    BIconShuffle: typeof import('bootstrap-icons-vue')['BIconShuffle']
-    BIconSkipEndFill: typeof import('bootstrap-icons-vue')['BIconSkipEndFill']
-    BIconSkipStartFill: typeof import('bootstrap-icons-vue')['BIconSkipStartFill']
-    BIconThreeDotsVertical: typeof import('bootstrap-icons-vue')['BIconThreeDotsVertical']
-    BIconVolumeUp: typeof import('bootstrap-icons-vue')['BIconVolumeUp']
-    DevicesView: typeof import('./devices/views/DevicesView.vue')['default']
-    FriendsView: typeof import('./friends/views/FriendsView.vue')['default']
-    HomeView: typeof import('./home/views/HomeView.vue')['default']
-    LibraryView: typeof import('./library/views/LibraryView.vue')['default']
-    PageLoader: typeof import('./shared/components/PageLoader.vue')['default']
-    PlaylistView: typeof import('./playlists/views/PlaylistView.vue')['default']
-    RouterLink: typeof import('vue-router')['RouterLink']
-    RouterView: typeof import('vue-router')['RouterView']
-    SettingsView: typeof import('./settings/views/SettingsView.vue')['default']
-    ThemesView: typeof import('./themes/views/ThemesView.vue')['default']
-    ThePlayer: typeof import('./core/components/ThePlayer.vue')['default']
-    TheQueue: typeof import('./core/components/TheQueue.vue')['default']
-    TheSideMenu: typeof import('./core/components/TheSideMenu.vue')['default']
-  }
-}
