+++ src/player/stores/current-track.ts
@@ -7,7 +7,7 @@ import {
 import type { ComputedRef } from 'vue'
 import { computed, watch } from 'vue'
 
-import { useQueue } from '@/core/stores/queue'
+import { useQueue } from '@/queue/stores/queue'
 import { randomArrayElement } from '@/shared/utils/random'
 import type { Track } from '@/shared/dto/track'
 import { RepeatMode } from '@/player/constants/repeat-mode'
