+++ src/themes/stores/colors.ts
